### 2025-05-27 #8292 @editorialbot
[PRE REVIEW]: DeepInverse: A Python package for solving imaging inverse problems with deep learning
**Submitting author:** @tachella (<a href="http://orcid.org/0000-0003-3878-9142">Julian Tachella</a>)
**Repository:** https://github.com/deepinv/deepinv
**Branch with paper.md** (empty if default branch): joss
**Version:** v0.3.0
