+++ README.rst
@@ -124,3 +124,21 @@ If you have any questions or suggestions, please join the conversation in our
    :target: https://discord.gg/qBqY5jKw3p
 .. |colab| image:: https://colab.research.google.com/assets/colab-badge.svg
    :target: https://colab.research.google.com/drive/1XhCO5S1dYN3eKm4NEkczzVU7ZLBuE42J
+
+
+Citation
+--------
+If you use DeepInverse in your research, please cite the following paper (available in `arXiv <https://arxiv.org/abs/2505.20160>`_):
+
+
+.. code-block::
+
+    @misc{tachella2025deepinverse,
+          title={DeepInverse: A Python package for solving imaging inverse problems with deep learning},
+          author={Julián Tachella and Matthieu Terris and Samuel Hurault and Andrew Wang and Dongdong Chen and Minh-Hai Nguyen and Maxime Song and Thomas Davies and Leo Davy and Jonathan Dong and Paul Escande and Johannes Hertrich and Zhiyuan Hu and Tobías I. Liaudat and Nils Laurent and Brett Levac and Mathurin Massias and Thomas Moreau and Thibaut Modrzyk and Brayan Monroy and Sebastian Neumayer and Jérémy Scanvic and Florian Sarron and Victor Sechaud and Georg Schramm and Chao Tang and Romain Vo and Pierre Weiss},
+          year={2025},
+          eprint={2505.20160},
+          archivePrefix={arXiv},
+          primaryClass={eess.IV},
+          url={https://arxiv.org/abs/2505.20160},
+    }
