### 2025-05-27 fe7b828
make copy pastable
