+++ CITATION.cff
@@ -1,71 +1,9 @@
-cff-version: 1.2.0
-message: "If you use this software, please cite it as below."
-title: "DeepInverse: A Python package for solving imaging inverse problems with deep learning"
-authors:
-  - family-names: "Tachella"
-    given-names: "Julián"
-  - family-names: "Terris"
-    given-names: "Matthieu"
-  - family-names: "Hurault"
-    given-names: "Samuel"
-  - family-names: "Wang"
-    given-names: "Andrew"
-  - family-names: "Chen"
-    given-names: "Dongdong"
-  - family-names: "Nguyen"
-    given-names: "Minh-Hai"
-  - family-names: "Song"
-    given-names: "Maxime"
-  - family-names: "Davies"
-    given-names: "Thomas"
-  - family-names: "Davy"
-    given-names: "Leo"
-  - family-names: "Dong"
-    given-names: "Jonathan"
-  - family-names: "Escande"
-    given-names: "Paul"
-  - family-names: "Hertrich"
-    given-names: "Johannes"
-  - family-names: "Hu"
-    given-names: "Zhiyuan"
-  - family-names: "Liaudat"
-    given-names: "Tobías I."
-  - family-names: "Laurent"
-    given-names: "Nils"
-  - family-names: "Levac"
-    given-names: "Brett"
-  - family-names: "Massias"
-    given-names: "Mathurin"
-  - family-names: "Moreau"
-    given-names: "Thomas"
-  - family-names: "Modrzyk"
-    given-names: "Thibaut"
-  - family-names: "Monroy"
-    given-names: "Brayan"
-  - family-names: "Neumayer"
-    given-names: "Sebastian"
-  - family-names: "Scanvic"
-    given-names: "Jérémy"
-  - family-names: "Sarron"
-    given-names: "Florian"
-  - family-names: "Sechaud"
-    given-names: "Victor"
-  - family-names: "Schramm"
-    given-names: "Georg"
-  - family-names: "Tang"
-    given-names: "Chao"
-  - family-names: "Vo"
-    given-names: "Romain"
-  - family-names: "Weiss"
-    given-names: "Pierre"
-version: "0.3.0"
-doi: "https://doi.org/10.48550/arXiv.2505.20160"
-date-released: 2023-06-30
-repository-code: "https://github.com/deepinv/deepinv"
-url: "https://arxiv.org/abs/2505.20160"
-license: "BSD 3"
-keywords:
-  - "inverse problems"
-  - "deep learning"
-  - "imaging"
-  - "Python"
+@misc{tachella2025deepinverse,
+      title={DeepInverse: A Python package for solving imaging inverse problems with deep learning},
+      author={Julián Tachella and Matthieu Terris and Samuel Hurault and Andrew Wang and Dongdong Chen and Minh-Hai Nguyen and Maxime Song and Thomas Davies and Leo Davy and Jonathan Dong and Paul Escande and Johannes Hertrich and Zhiyuan Hu and Tobías I. Liaudat and Nils Laurent and Brett Levac and Mathurin Massias and Thomas Moreau and Thibaut Modrzyk and Brayan Monroy and Sebastian Neumayer and Jérémy Scanvic and Florian Sarron and Victor Sechaud and Georg Schramm and Chao Tang and Romain Vo and Pierre Weiss},
+      year={2025},
+      eprint={2505.20160},
+      archivePrefix={arXiv},
+      primaryClass={eess.IV},
+      url={https://arxiv.org/abs/2505.20160},
+}
