+++ docs/source/index.rst
@@ -45,10 +45,11 @@ If you use DeepInverse in your research, please cite the following paper (availa
 
 .. code-block:: bash
 
-    @misc{tachella2025deepinverse,
+    @article{tachella2025deepinverse,
           title={DeepInverse: A Python package for solving imaging inverse problems with deep learning},
           author={Julián Tachella and Matthieu Terris and Samuel Hurault and Andrew Wang and Dongdong Chen and Minh-Hai Nguyen and Maxime Song and Thomas Davies and Leo Davy and Jonathan Dong and Paul Escande and Johannes Hertrich and Zhiyuan Hu and Tobías I. Liaudat and Nils Laurent and Brett Levac and Mathurin Massias and Thomas Moreau and Thibaut Modrzyk and Brayan Monroy and Sebastian Neumayer and Jérémy Scanvic and Florian Sarron and Victor Sechaud and Georg Schramm and Chao Tang and Romain Vo and Pierre Weiss},
           year={2025},
+          journal={arXiv preprint arXiv:2503.08915},
           eprint={2505.20160},
           archivePrefix={arXiv},
           primaryClass={eess.IV},
