### 2025-05-28 3ce1236
Single test with conditional mock for TomographyWithAstra. Fix doc. Change api img_shape -> img_size.
