+++ deepinv/tests/test_physics.py
@@ -1,4 +1,4 @@
-from math import sqrt, pi as PI
+from math import sqrt
 from typing import Optional, List
 import pytest
 import torch
@@ -1017,7 +1017,6 @@ def test_tomography(device):
 
     :param device: (torch.device) cpu or cuda:x
     """
-    PI = 4 * torch.ones(1).atan()
     for normalize in [True, False]:
         for parallel_computation in [True, False]:
             for fan_beam in [True, False]:
@@ -1036,115 +1035,12 @@ def test_tomography(device):
                     x = torch.randn(imsize, device=device).unsqueeze(0)
                     r = (
                         physics.A_adjoint(physics.A(x))
-                        * PI.item()
+                        * torch.pi
                         / (2 * len(physics.radon.theta))
                     )
                     y = physics.A(r)
                     error = (physics.A_dagger(y) - r).flatten().mean().abs()
                     assert error < 0.2
-
-
-@pytest.mark.skipif(
-    not torch.cuda.is_available(), reason="TomographyWithAstra requires GPU to run"
-)
-@pytest.mark.parametrize(
-    "is_2d,geometry_type,normalize",
-    [
-        (True, "parallel", False),
-        (True, "parallel", True),
-        (True, "fanbeam", False),
-        (True, "fanbeam", True),
-        (False, "parallel", False),
-        (False, "parallel", True),
-        (False, "conebeam", False),
-        (False, "conebeam", True),
-    ],
-)
-def test_tomography_with_astra(is_2d, geometry_type, normalize):
-    r"""
-    Tests tomography operator with astra backend which does not have a numerically precise adjoint.
-
-    :param bool is_2d: Runs the test with 2D geometry, else 3D.
-    :param str geometry_type: In 2D, expects ``parallel`` or ``fanbeam``. In 3D expects ``parallel`` or ``conebeam``.
-    :param bool normalize: Initializes the operator with ``normalize=normalize``.
-    """
-
-    device = "cuda" if torch.cuda.is_available() else "cpu"
-
-    pytest.importorskip(
-        "astra",
-        reason="This test requires astra-toolbox. It should be "
-        "installed with `conda install -c astra-toolbox -c nvidia astra-toolbox`",
-    )
-
-    ## Test 2d transforms
-    if is_2d:
-        img_shape = (16, 16)
-        physics = dinv.physics.TomographyWithAstra(
-            img_shape=img_shape,
-            num_detectors=2 * img_shape[0],
-            num_angles=2 * img_shape[0],
-            angular_range=(0, PI) if geometry_type == "parallel" else (0, 2 * PI),
-            geometry_type=geometry_type,
-            normalize=normalize,
-            device=device,
-        )
-
-        x = torch.rand(1, 1, *img_shape, device=device)
-
-        ## --- Test adjointness ---
-        Ax = physics.A(x)
-        y = torch.rand_like(Ax)
-        At_y = physics.A_adjoint(y)
-
-        Ax_y = torch.sum(Ax * y).item()
-        At_y_x = torch.sum(At_y * x).item()
-
-        relative_error = abs(Ax_y - At_y_x) / At_y_x
-        assert relative_error < 0.01  # at least 99% adjoint
-
-        ## --- Test pseudoinverse ---
-        r_tol = 0.05 if geometry_type == "parallel" else 0.1
-        r = physics.A_adjoint(physics.A(x))
-        y = physics.A(r)
-        error = torch.linalg.norm(physics.A_dagger(y) - r) / torch.linalg.norm(r)
-        assert error < r_tol
-
-    else:
-        ## Test 3d transforms
[…]
