+++ deepinv/physics/tomography.py
@@ -22,15 +22,12 @@ except:
     astra = ImportError("The astra-toolbox package is not installed.")
 
 
-PI = 4 * torch.ones(1).atan()
-
-
 class Tomography(LinearPhysics):
     r"""
     (Computed) Tomography operator.
 
     The Radon transform is the integral transform which takes a square image :math:`x` defined on the plane to a function
-    :math:`y=Rx` defined on the (two-dimensional) space of lines in the plane, whose value at a particular line is equal
+    :math:`y=Ax` defined on the (two-dimensional) space of lines in the plane, whose value at a particular line is equal
     to the line integral of the function over that line.
 
     .. note::
@@ -168,7 +165,7 @@ class Tomography(LinearPhysics):
         if self.fan_beam:
             y = self.filter(y)
             output = (
-                self.A_adjoint(y, **kwargs) * PI.item() / (2 * len(self.radon.theta))
+                self.A_adjoint(y, **kwargs) * torch.pi / (2 * len(self.radon.theta))
             )
             if self.normalize:
                 output = output * output.shape[-1] ** 2
@@ -195,7 +192,7 @@ class Tomography(LinearPhysics):
             # IRadon is not exactly the adjoint but a rescaled version of it...
             output = (
                 self.iradon(y, filtering=False)
-                / PI.item()
+                / torch.pi
                 * (2 * len(self.iradon.theta))
             )
             if self.normalize:
@@ -210,11 +207,11 @@ class TomographyWithAstra(LinearPhysics):
     information on the ``astra`` wrapper.
 
     Mathematically, it is described as a ray transform
-    :math:`R` which linearly integrates an object :math:`x` along straight
+    :math:`A` which linearly integrates an object :math:`x` along straight
     lines
 
     .. math::
-        y = Rx
+        y = Ax
 
     where :math:`y` is the set of line integrals, called sinogram in 2D, or
     radiographs in 3D. An object is typically scanned using a surrounding circular
@@ -256,10 +253,10 @@ class TomographyWithAstra(LinearPhysics):
 
         The :class:`deepinv.physics.functional.XrayTransform` used in :class:`TomographyWithAstra` sequentially processes batch elements, which can make the 2d parallel beam operator significantly slower than the its native torch counterpart with :class:`deepinv.physics.Tomography` (though still more memory-efficient).
 
-    :param tuple[int, ...] img_shape: Shape of the object grid, either a 2 or 3-element tuple, for respectively 2d or 3d.
+    :param tuple[int, ...] img_size: Shape of the object grid, either a 2 or 3-element tuple, for respectively 2d or 3d.
     :param int num_angles: Number of angular positions sampled uniformly in ``angular_range``. (default: 180)
     :param int | tuple[int, ...], None num_detectors: In 2d, specify an integer for a single line of detector cells. In 3d, specify a 2-element tuple for (row,col) shape of the detector. (default: None)
-    :param tuple[float, float] angular_range: Angular range, defaults to (0,``math.pi``).
+    :param tuple[float, float] angular_range: Angular range, defaults to (0,``torch.pi``).
     :param float | tuple[float, float] detector_spacing: In 2d the width of a detector cell. In 3d a 2-element tuple specifying the (vertical, horizontal) dimensions of a detector cell. (default: 1.0)
[…]

[…]
