### 2025-05-28 10c4e50
Update examples/external-libraries/_demo_astra_tomography.py
Co-authored-by: Andrewwango <andrew.wang27@gmail.com>
