+++ examples/external-libraries/_demo_astra_tomography.py
@@ -2,7 +2,9 @@
 Low-dose CT with ASTRA backend and Total-Variation (TV) prior
 ==============================================================================
 
-This example shows how to use Total-Variation regularization to solve a low-dose computed tomography problem. We use the ``TomographyWithAstra`` operator which wraps the [astra-toolbox](https://astra-toolbox.com/) backend in a custom ``LinearPhysics`` operator.
+This example shows how to use the Astra tomography toolbox with deepinv, a popular toolbox for tomography with GPU acceleration. 
+
+We show how to use the :class:`deepinv.physics.TomographyWithAstra` operator (which wraps the [astra-toolbox](https://astra-toolbox.com/) backend) to solve a low-dose computed tomography problem with Total-Variation regularization.
 
 ``TomographyWithAstra`` requires the astra-toolbox to function correctly, which can be easily installed using the command: `conda install -c astra-toolbox -c nvidia astra-toolbox`.
 
