### 2025-05-28 484fffe
change some patch calls to patch.object calls
