### 2025-05-28 e792f77
black. Fix demo_vanilla_PnP
