### 2025-05-30 bde6e51
Typo in Documention formula of SaltPepperNoise (#504)
