### 2025-05-30 4a76235
wavelet denoiser batch
