### 2025-05-30 be5b5d9
wavelet denoiser batch, level, dimension
