+++ deepinv/models/wavdict.py
@@ -218,7 +218,7 @@ class WaveletDenoiser(Denoiser):
         for level in range(1, self.level + 1):
             ths_cur = self.reshape_ths(ths, level)
             for c, key in enumerate(["aad", "ada", "daa", "add", "dad", "dda", "ddd"]):
-                coeffs[level][key] = self.prox_l1(coeffs[level][key], ths_cur[c])
+                coeffs[level][key] = self.thresold_func(coeffs[level][key], ths_cur[c])
         return coeffs
 
     def threshold_ND(self, coeffs, ths):
@@ -243,7 +243,9 @@ class WaveletDenoiser(Denoiser):
             padding_bottom = h % 2
             padding_right = w % 2
             p = (padding_bottom, padding_right)
-            x = torch.nn.functional.pad(x, (0, p[0], 0, p[1]), mode="replicate")
+            x = torch.nn.functional.pad(
+                x, (0, padding_right, 0, padding_bottom), mode="replicate"
+            )
         elif self.dimension == 3:
             d, h, w = x.size()[-3:]
             padding_depth = d % 2
@@ -251,7 +253,9 @@ class WaveletDenoiser(Denoiser):
             padding_right = w % 2
             p = (padding_depth, padding_bottom, padding_right)
             x = torch.nn.functional.pad(
-                x, (0, p[0], 0, p[1], 0, p[2]), mode="replicate"
+                x,
+                (0, padding_right, 0, padding_bottom, 0, padding_depth),
+                mode="replicate",
             )
         return x, p
 
@@ -275,9 +279,11 @@ class WaveletDenoiser(Denoiser):
         Since the approximation coefficients are not thresholded, we do not need to provide a thresholding parameter,
         ths has shape (n_levels-1, 3).
         """
+
+        # Return tensor of shape (B, n_level - 1, numel)
         numel = 3 if self.dimension == 2 else 7
         if not torch.is_tensor(ths):
-            if isinstance(ths, int) or isinstance(ths, float):
+            if isinstance(ths, (int, float)):
                 ths_cur = [ths] * numel
             elif len(ths) == 1:
                 ths_cur = [ths[0]] * numel
@@ -286,12 +292,47 @@ class WaveletDenoiser(Denoiser):
                 if len(ths_cur) == 1:
                     ths_cur = [ths_cur[0]] * numel
         else:
-            if len(ths.shape) == 1:  # Needs to reshape to shape (n_levels-1, 3)
-                ths_cur = ths.squeeze().repeat(numel)
+            if ths.ndim == 0 or ths.ndim == 1:  # a tensor of shape 0 or (B,)
+                return self._reshape_ths_one_dim(ths, level)
+            elif ths.ndim == 2:  # (B, n_levels-1)
+                return self._reshape_ths_two_dim(ths, level)
+            elif ths.ndim == 3:
+                # (B, n_levels-1, numel) or (B, n_levels-1, 1)
+                ths_cur = self._reshape_ths_three_dim(ths, level)
             else:
-                ths_cur = ths[level - 2]
+                raise ValueError(
+                    f"Expected tensor of 0, 1, 2 or 3 dimensions. Got tensor of {ths.ndim} dimensions"
+                )
 
         return ths_cur
+
+    def _reshape_ths_one_dim(self, ths, level):
+        numel = 3 if self.dimension == 2 else 7
+        return [ths] * numel
+
+    def _reshape_ths_two_dim(self, ths, level):
+        numel = 3 if self.dimension == 2 else 7
+        if ths.size(1) == 1:
+            return [ths[:, 0]] * numel
+        else:
+            assert ths.size(1) == self.level
+            return [ths[:, level - 2]] * numel
+
[…]
