### 2025-05-31 0754de7
fixed epll batch sigma
