+++ deepinv/models/epll.py
@@ -2,6 +2,8 @@ import torch
 from deepinv.optim.epll import EPLL
 from deepinv.physics import Denoising, GaussianNoise
 from .base import Denoiser
+from typing import Union
+from deepinv.optim.utils import GaussianMixtureModel
 
 
 class EPLLDenoiser(Denoiser):
@@ -30,12 +32,12 @@ class EPLLDenoiser(Denoiser):
 
     def __init__(
         self,
-        GMM=None,
-        n_components=200,
-        pretrained="download",
-        patch_size=6,
-        channels=1,
-        device="cpu",
+        GMM: GaussianMixtureModel = None,
+        n_components: int = 200,
+        pretrained: str = "download",
+        patch_size: int = 6,
+        channels: int = 1,
+        device: torch.device = torch.device("cpu"),
     ):
         super(EPLLDenoiser, self).__init__()
         self.PatchGMM = EPLL(
@@ -43,7 +45,13 @@ class EPLLDenoiser(Denoiser):
         )
         self.denoising_operator = Denoising(GaussianNoise(0))
 
-    def forward(self, x, sigma, betas=None, batch_size=-1):
+    def forward(
+        self,
+        x: torch.Tensor,
+        sigma: Union[float, torch.Tensor, list[float]],
+        betas: list[float] = None,
+        batch_size: int = -1,
+    ) -> torch.Tensor:
         r"""
         Denoising method based on the minimization problem.
 
@@ -55,6 +63,7 @@ class EPLLDenoiser(Denoiser):
             but a small value reduces the memory consumption
             and might increase the computation time. ``-1`` for considering all patches at once.
         """
+        sigma = self._handle_sigma(sigma, x.size(0)).to(x.device)
         return self.PatchGMM(
             x,
             x_init=x.clone(),
@@ -63,3 +72,34 @@ class EPLLDenoiser(Denoiser):
             batch_size=batch_size,
             betas=betas,
         )
+
+    @staticmethod
+    def _handle_sigma(sigma, batch_size):
+        r"""
+        Handle the sigma parameter for the denoising method.
+
+        :param float, torch.Tensor sigma: noise level, can be a scalar or a list of scalars.
+        :param batch_size: size of the batch.
+        :return: sigma as a tensor of shape (batch_size, 1).
+        """
+        if isinstance(sigma, (int, float)):
+            sigma = torch.tensor([float(sigma)] * batch_size)
+        elif isinstance(sigma, list):
+            assert (
+                len(sigma) == batch_size
+            ), "Length of sigma list must match batch size."
+            sigma = (
+                torch.tensor(sigma, dtype=torch.float32)
+                .squeeze()
+                .view(-1)
+                .expand(batch_size)
+            )
+        elif isinstance(sigma, torch.Tensor):
+            sigma = sigma.squeeze().view(-1).expand(batch_size)
+        else:
+            raise TypeError(
+                "sigma must be a scalar, list, or torch.Tensor, "
+                f"but got {type(sigma)}."
+            )
+
+        return sigma
