### 2025-05-31 ebb91c2
test wavelet denoiser ths
