+++ deepinv/loss/metric/distortion.py
@@ -297,6 +297,9 @@ class LpNorm(Metric):
         the data must either be of complex dtype or have size 2 in the channel dimension (usually the second dimension after batch).
     :param str reduction: a method to reduce metric score over individual batch scores. ``mean``: takes the mean, ``sum`` takes the sum, ``none`` or None no reduction will be applied (default).
     :param str norm_inputs: normalize images before passing to metric. ``l2``normalizes by L2 spatial norm, ``min_max`` normalizes by min and max of each input.
+    
+    :References:
+    .. footbibliography::
     """
 
     def __init__(self, p=2, onesided=False, **kwargs):
