+++ deepinv/loss/score.py
@@ -74,7 +74,7 @@ class ScoreLoss(Loss):
 
     :References:
     .. footbibliography::
-    
+
     """
 
     def __init__(self, noise_model=None, total_batches=1000, delta=(0.001, 0.1)):
