+++ docs/source/refs.bib
@@ -899,14 +899,3 @@ journal = {ACM Trans. Graph.}
   year={2011},
   publisher={Wiley Online Library}
 }
-
-@article{meng2020large,
-  title={A large-scale benchmark data set for evaluating pansharpening performance: Overview and implementation},
-  author={Meng, Xiangchao and Xiong, Yiming and Shao, Feng and Shen, Huanfeng and Sun, Weiwei and Yang, Gang and Yuan, Qiangqiang and Fu, Randi and Zhang, Hongyan},
-  journal={IEEE Geoscience and Remote Sensing Magazine},
-  volume={9},
-  number={1},
-  pages={18--52},
-  year={2020},
-  publisher={IEEE}
-}
