+++ examples/sampling/demo_ddrm.py
@@ -10,6 +10,8 @@ The DDRM method requires that:
 * The operator has a singular value decomposition (i.e., the operator is a :class:`deepinv.physics.DecomposablePhysics`).
 * The noise is Gaussian with known standard deviation (i.e., the noise model is :class:`deepinv.physics.GaussianNoise`).
 
+:References:
+.. footbibliography::
 """
 
 import deepinv as dinv
