+++ examples/unfolded/demo_learned_primal_dual.py
@@ -7,6 +7,10 @@ where both the data fidelity and the prior are learned modules, distinct for eac
 
 The algorithm is used for CT reconstruction trained on random phantoms.
 The phantoms are generated on the fly during training using the odl library (https://odlgroup.github.io/odl/).
+
+
+:References:
+.. footbibliography::
 """
 
 import deepinv as dinv
