### 2025-06-03 74983d9
Refactor PoissonNoise initialization and update demo to use Poisson noise model with adjusted gain and learning rate
