+++ deepinv/physics/noise.py
@@ -330,7 +330,7 @@ class PoissonNoise(NoiseModel):
         self, gain=1.0, normalize=True, clip_positive=False, rng: torch.Generator = None
     ):
         super().__init__(rng=rng)
-        self.normalize = normalize
+        self.register_buffer("normalize", torch.tensor(normalize, dtype=torch.bool))
         self.update_parameters(gain=gain)
         self.clip_positive = clip_positive
 
