+++ examples/adversarial-learning/demo_gan_imaging.py
@@ -10,9 +10,9 @@ your own GAN by using :class:`deepinv.training.AdversarialTrainer`. These
 examples can also be easily extended to train more complicated GANs such
 as CycleGAN.
 
-This example is based on the papers DeblurGAN :footcite:t:`kupyn2018deblurgan`,
-Compressed Sensing using Generative Models (CSGM) :footcite:t:`bora2017compressed`,
-AmbiantGAN :footcite:t:`bora2018ambientgan`, and Unsupervised Adversarial Image Reconstruction (UAIR) :footcite:t:`pajot2019unsupervised`.
+This example is based on the papers DeblurGAN :footcite:ps:`kupyn2018deblurgan`,
+Compressed Sensing using Generative Models (CSGM) :footcite:ps:`bora2017compressed`,
+AmbiantGAN :footcite:ps:`bora2018ambientgan`, and Unsupervised Adversarial Image Reconstruction (UAIR) :footcite:ps:`pajot2019unsupervised`.
 
 Adversarial networks are characterised by the addition of an adversarial
 loss :math:`\mathcal{L}_\text{adv}` to the standard reconstruction loss:
@@ -95,7 +95,7 @@ test_dataloader = DataLoader(
 # We first define reconstruction network (i.e conditional generator) and
 # discriminator network to use for adversarial training. For demonstration
 # we use a simple U-Net as the reconstruction network and the
-# discriminator from PatchGAN :footcite:t:`isola2017image`, but
+# discriminator from PatchGAN :footcite:ps:`isola2017image`, but
 # these can be replaced with any architecture e.g transformers, unrolled
 # etc. Further discriminator models are in :ref:`adversarial models <adversarial>`.
 #
@@ -130,7 +130,7 @@ def get_models(model=None, D=None, lr_g=1e-4, lr_d=1e-4, device=device):
 # Conditional GAN training
 # ~~~~~~~~~~~~~~~~~~~~~~~~
 #
-# Conditional GANs :footcite:t:`kupyn2018deblurgan` are a type of GAN where the generator is conditioned on a label or input.
+# Conditional GANs :footcite:ps:`kupyn2018deblurgan` are a type of GAN where the generator is conditioned on a label or input.
 # In the context of imaging, this can be used to generate images from a given measurement.
 # In this example, we use a simple U-Net as the generator
 # and a PatchGAN discriminator. The forward pass of the generator is given by:
@@ -211,7 +211,7 @@ trainer.test(test_dataloader)
 # UAIR training
 # ~~~~~~~~~~~~~
 #
-# Unsupervised Adversarial Image Reconstruction (UAIR) :footcite:t:`pajot2019unsupervised`
+# Unsupervised Adversarial Image Reconstruction (UAIR) :footcite:ps:`pajot2019unsupervised`
 # is a method for solving inverse problems using generative models. In this
 # example, we use a simple U-Net as the generator and discriminator, and
 # train using the adversarial loss. The forward pass of the generator is defined as:
@@ -279,7 +279,7 @@ trainer.test(test_dataloader)
 # CSGM / AmbientGAN training
 # ~~~~~~~~~~~~~~~~~~~~~~~~~~
 #
-# Compressed Sensing using Generative Models (CSGM) :footcite:t:`bora2017compressed` and AmbientGAN :footcite:t:`bora2018ambientgan` are two methods for solving inverse problems
+# Compressed Sensing using Generative Models (CSGM) :footcite:ps:`bora2017compressed` and AmbientGAN :footcite:ps:`bora2018ambientgan` are two methods for solving inverse problems
 # using generative models. CSGM uses a generative model to solve the inverse problem by optimising the latent
[…]
