+++ deepinv/optim/optim_iterators/__init__.py
@@ -6,4 +6,3 @@ from .hqs import HQSIteration
 from .drs import DRSIteration
 from .gradient_descent import GDIteration, MDIteration
 from .spectral_methods import SMIteration
-
