### 2025-06-04 96f4646
reformat params_algo
