+++ deepinv/optim/optimizers.py
@@ -140,6 +140,8 @@ class BaseOptim(Reconstructor):
     :param int history_size: size of the history of iterates used for Anderson acceleration. Default: ``5``.
     :param float beta_anderson_acc: momentum of the Anderson acceleration step. Default: ``1.0``.
     :param float eps_anderson_acc: regularization parameter of the Anderson acceleration step. Default: ``1e-4``.
+    :param bool unfold: whether to unfold the algorithm and make the model parameters trainable. Default: ``False``.
+    :param list trainable_params: list of the algorithmic parameters among the keys of the dictionery params_algo to be made trainable. Default: ``None``, which means that all parameters in params_algo are trainable. For no trainable parameters, set to an empty list ``[]``.
     :param bool verbose: whether to print relevant information of the algorithm during its run,
         such as convergence criterion at each iterate. Default: ``False``.
     :return: a torch model that solves the optimization problem.
@@ -167,7 +169,7 @@ class BaseOptim(Reconstructor):
         beta_anderson_acc=1.0,
         eps_anderson_acc=1e-4,
         unfold=False,
-        trainable_params=["lambda", "stepsize"],
+        trainable_params=None,
         verbose=False,
         device=torch.device("cpu"),
         **kwargs,
@@ -257,6 +259,8 @@ class BaseOptim(Reconstructor):
 
         # set trainable parameters
         if self.unfold:
+            if trainable_params is None:
+                trainable_params = params_algo.keys()
             for param_key in trainable_params:
                 if param_key in self.init_params_algo.keys():
                     param_value = self.init_params_algo[param_key]
@@ -671,7 +675,7 @@ class ADMM(BaseOptim):
         \end{equation}
 
     where :math:`\datafid{x}{y}` is the data-fidelity term, :math:`\reg{x}` is the regularization term.
-    If the attribute ``g_first`` is set to False (by default), the ADMM algorithm  writes (`see this paper <https://www.nowpublishers.com/article/Details/MAL-016>`_)
+    If the attribute ``g_first`` is set to False (by default), the ADMM iterations write (`see this paper <https://www.nowpublishers.com/article/Details/MAL-016>`_)
 
     .. math::
         \begin{equation*}
@@ -685,7 +689,9 @@ class ADMM(BaseOptim):
     where :math:`\gamma>0` is a stepsize and :math:`\beta>0` is a relaxation parameter.  If the attribute ``g_first`` is set to ``True``, the functions :math:`f` and :math:`\regname` are
     inverted in the previous iterations. The ADMM iterations are defined in the iterator class :class:`deepinv.optim.ADMMIteration`.
 
-    If the attribute ``unfold`` is set to ``True``, the algorithm is unfolded and the parameters of the algorithm are trainable.
+    If the attribute ``unfold`` is set to ``True``, the algorithm is unfolded and the algorithmic parameters of the algorithm are trainable.
+    By default, all the algorithm parameters are trainiable : the stepsize :math:`\gamma`, the regularization parameter :math:`\lambda`, the prior parameter and the relaxation parameter :math:`\beta`.
+    Use the ``trainable_params`` argument to adjust the list of trainable parameters.
 
     :param list, deepinv.optim.DataFidelity data_fidelity: data-fidelity term :math:`\datafid{x}{y}`.
         Either a single instance (same data-fidelity for each iteration) or a list of instances of
@@ -699,10 +705,10 @@ class ADMM(BaseOptim):
     :param float g_param: parameter of the prior function. For example the noise level for a denoising prior. Default: ``None``.
     :param int max_iter: maximum number of iterations of the optimization algorithm. Default: ``100``.
[…]
