+++ deepinv/tests/test_unfolded.py
@@ -59,21 +59,17 @@ def test_unfolded(unfolded_algo, imsize, dummy_dataset, device):
     sigma_denoiser_init = 0.01
     sigma_denoiser = [sigma_denoiser_init * torch.ones(level, 3)] * max_iter
     # sigma_denoiser = [torch.Tensor([sigma_denoiser_init])]*max_iter
-    params_algo = {  # wrap all the restoration parameters in a 'params_algo' dictionary
-        "stepsize": stepsize,
-        "g_param": sigma_denoiser,
-        "lambda": lamb,
-    }
-
     trainable_params = [
         "g_param",
         "stepsize",
-    ]  # define which parameters from 'params_algo' are trainable
+    ]  # define which parameters  are trainable
 
     # Define the unfolded trainable model.
     model = getattr(optim, unfolded_algo)(
         unfold=True,
-        params_algo=params_algo,
+        stepsize=stepsize,
+        g_param=sigma_denoiser,
+        lambda_reg=lamb,
         trainable_params=trainable_params,
         data_fidelity=data_fidelity,
         max_iter=max_iter,
@@ -113,9 +109,7 @@ def test_unfolded(unfolded_algo, imsize, dummy_dataset, device):
     loss.backward()
 
 
-DEQ_ALGO = ["DEQ_ProximalGradientDescent", "DEQ_HQS"]
-
-
+DEQ_ALGO = ["PGD", "HQS"]
 @pytest.mark.parametrize("unfolded_algo", DEQ_ALGO)
 def test_DEQ(unfolded_algo, imsize, dummy_dataset, device):
     pytest.importorskip("ptwt")
