+++ deepinv/tests/test_optim.py
@@ -334,9 +334,8 @@ def test_optim_algo(name_algo, imsize, dummy_dataset, device):
             stepsize = 0.9 / physics.compute_norm(x, tol=1e-4).item()
             sigma = None
 
-        lamb = 0.9
+        lambda_reg = 0.9
         max_iter = 1000
-        params_algo = {"stepsize": stepsize, "lambda": lamb, "sigma": sigma}
 
         optimalgo = getattr(optim, name_algo)(
             prior=prior,
@@ -345,7 +344,9 @@ def test_optim_algo(name_algo, imsize, dummy_dataset, device):
             crit_conv="residual",
             thres_conv=1e-11,
             verbose=True,
-            params_algo=params_algo,
+            stepsize=stepsize,
+            lambda_reg=lambda_reg,
+            g_param=sigma,
             early_stop=True,
             g_first=g_first,
         )
@@ -370,15 +371,15 @@ def test_optim_algo(name_algo, imsize, dummy_dataset, device):
                     stepsize
                 )  # Gradient of the moreau envelope
                 assert torch.allclose(
-                    moreau_grad, -lamb * subdiff, atol=1e-8
+                    moreau_grad, -lambda_reg * subdiff, atol=1e-8
                 )  # Optimality condition
             else:
                 subdiff = data_fidelity.grad(x, y, physics)
-                moreau_grad = (x - prior.prox(x, gamma=lamb * stepsize)) / (
-                    lamb * stepsize
+                moreau_grad = (x - prior.prox(x, gamma=lambda_reg * stepsize)) / (
+                    lambda_reg * stepsize
                 )  # Gradient of the moreau envelope
                 assert torch.allclose(
-                    lamb * moreau_grad, -subdiff, atol=1e-8
+                    lambda_reg * moreau_grad, -subdiff, atol=1e-8
                 )  # Optimality condition
         else:
             subdiff = prior.grad(x)
@@ -386,7 +387,7 @@ def test_optim_algo(name_algo, imsize, dummy_dataset, device):
             # The optimality condition is then :math:`0 \in  \nabla f(x)+ \lambda \partial g(x)`
             grad_deepinv = data_fidelity.grad(x, y, physics)
             assert torch.allclose(
-                grad_deepinv, -lamb * subdiff, atol=1e-8
+                grad_deepinv, -lambda_reg * subdiff, atol=1e-8
             )  # Optimality condition
 
 
@@ -444,7 +445,7 @@ def test_pnp_algo(pnp_algo, imsize, dummy_dataset, device):
     # Note: results are better for sigma_denoiser=0.001, but it takes longer to run.
     sigma_denoiser = torch.tensor([[0.1]])
     stepsize = 1.0
-    lamb = 1.0
+    lambda_reg = 1.0
 
     data_fidelity = L2()
 
@@ -452,12 +453,6 @@ def test_pnp_algo(pnp_algo, imsize, dummy_dataset, device):
     prior = PnP(denoiser=dinv.models.WaveletDenoiser(wv="db8", level=3, device=device))
 
     stepsize_dual = 1.0 if pnp_algo == "PrimalDualCP" else None
-    params_algo = {
-        "stepsize": stepsize,
-        "g_param": sigma_denoiser,
-        "lambda": lamb,
-        "stepsize_dual": stepsize_dual,
-    }
 
     custom_init = custom_init_CP if pnp_algo == "PrimalDualCP" else None
 
@@ -467,7 +462,10 @@ def test_pnp_algo(pnp_algo, imsize, dummy_dataset, device):
         max_iter=max_iter,
         thres_conv=1e-4,
         verbose=True,
[…]
