+++ deepinv/tests/test_unfolded.py
@@ -110,6 +110,8 @@ def test_unfolded(unfolded_algo, imsize, dummy_dataset, device):
 
 
 DEQ_ALGO = ["PGD", "HQS"]
+
+
 @pytest.mark.parametrize("unfolded_algo", DEQ_ALGO)
 def test_DEQ(unfolded_algo, imsize, dummy_dataset, device):
     pytest.importorskip("ptwt")
