+++ deepinv/models/dsccp.py
@@ -1,41 +1,29 @@
 import os
-import numpy as np
 import torch
 from torch import nn
 from torch.nn import functional
-import torch.utils.data as td
-import torchvision as tv
-from PIL import Image
-import nntools as nt
-from utils import DenoisingStatsManager
-from utils import NNRegressor
 
+class DScCP(nn.Module):
+    r"""
+    DScCP denoiser network.
 
-class HT(nn.Module):
-    def __init__(self, n_ch=64):
-        super(HT, self).__init__()
-    def forward(self, x, thres):
-        out=torch.clone(x)
-        for i in range(x.size(0)):
-            out[i] = torch.clamp(x[i],min=-thres[i],max=thres[i])
-        return out
+    The network architecture is based on the paper
+    `Unfolded proximal neural networks for robust image Gaussian denoising <https://arxiv.org/pdf/2308.03139>`_,
+    and has an unrolled architecture.
+    
+    # TODO
+    The network takes into account the noise level of the input image, which is encoded as an additional input channel.
 
-class SoftShk(nn.Module):
-    def __init__(self, n_ch=64):
-        super(SoftShk, self).__init__()
+    A pretrained network for RGB images with K=20, and F=64
+    can be downloaded via setting ``pretrained='download'``.
 
-        self.n_ch = n_ch
-        self.relu1 = nn.ReLU()
-        self.relu2 = nn.ReLU()
+    :param int K: number of layers.
+    :param int F: number of convolutional filters.
+    :param str device: gpu or cpu.
 
-    def forward(self, x, l):
-        out = self.relu1(x - l) - self.relu2(-x - l)
-        return out
-
-
-class unfolded_ScCP_ver2(NNRegressor):
-    def __init__(self, K, F, device='cpu'):
-        super(unfolded_ScCP_ver2, self).__init__()
+    """
+    def __init__(self, K=20, F=64, device='cpu'):
+        super(DScCP, self).__init__()
         self.K = K
         self.F = F
         self.norm_net = 0
@@ -48,136 +36,49 @@ class unfolded_ScCP_ver2(NNRegressor):
             self.conv[i * 2 + 1].weight = self.conv[i * 2].weight
             nn.init.kaiming_normal_(self.conv[i * 2].weight.data, nonlinearity = 'relu')
 
-        x=np.ones(K)
+        x=torch.ones(K)
         self.mu = nn.Parameter(torch.tensor(x,requires_grad=True,dtype=torch.float).cpu())
-        # self.alpha = nn.Parameter(torch.tensor(x,requires_grad=True,dtype=torch.float).cuda())
-        self.ht = HT()
-        self.lip = torch.tensor(np.ones(K),requires_grad=False,dtype=torch.float).cpu()
+        self.lip = torch.tensor(torch.ones(K),requires_grad=False,dtype=torch.float).cpu()
         # apply He's initialization
         for i in range(K):
             nn.init.kaiming_normal_(
                 self.conv[i].weight.data, nonlinearity='relu')
 
         outputdir= 'checkpoints/unfolded_ScCP_ver2/unfolded_ScCP_ver2_F64_K20_batchsize200_param_34580_data_vary'
-        # outputdir= 'checkpoints/unfolded_ISTA_ver2/unfolded_ISTA_ver2_F64_K20_batchsize200_param_34560_data_vary'
-
         checkpoint_path = os.path.join(outputdir, "checkpoint.pth.tar")
         checkpoint = torch.load(checkpoint_path,map_location=device)
         self.load_state_dict(checkpoint['Net'])
 
-    def lip_cal(self,x):
-        K=self.K
-        for i in range(K):
[…]
