+++ deepinv/models/dsccp.py
@@ -3,6 +3,7 @@ import torch
 from torch import nn
 from torch.nn import functional
 
+
 class DScCP(nn.Module):
     r"""
     DScCP denoiser network.
@@ -10,7 +11,7 @@ class DScCP(nn.Module):
     The network architecture is based on the paper
     `Unfolded proximal neural networks for robust image Gaussian denoising <https://arxiv.org/pdf/2308.03139>`_,
     and has an unrolled architecture.
-    
+
     # TODO
     The network takes into account the noise level of the input image, which is encoded as an additional input channel.
 
@@ -22,7 +23,8 @@ class DScCP(nn.Module):
     :param str device: gpu or cpu.
 
     """
-    def __init__(self, K=20, F=64, device='cpu'):
+
+    def __init__(self, K=20, F=64, device="cpu"):
         super(DScCP, self).__init__()
         self.K = K
         self.F = F
@@ -30,25 +32,44 @@ class DScCP(nn.Module):
         # convolution layers
         self.conv = nn.ModuleList()
         for i in range(self.K):
-            self.conv.append(nn.Conv2d(in_channels = 3, out_channels = F, kernel_size = 3, padding = 1, bias = False, dtype=torch.float))
             self.conv.append(
-                nn.ConvTranspose2d(in_channels = F, out_channels = 3, kernel_size = 3, padding = 1, bias = False, dtype=torch.float))
+                nn.Conv2d(
+                    in_channels=3,
+                    out_channels=F,
+                    kernel_size=3,
+                    padding=1,
+                    bias=False,
+                    dtype=torch.float,
+                )
+            )
+            self.conv.append(
+                nn.ConvTranspose2d(
+                    in_channels=F,
+                    out_channels=3,
+                    kernel_size=3,
+                    padding=1,
+                    bias=False,
+                    dtype=torch.float,
+                )
+            )
             self.conv[i * 2 + 1].weight = self.conv[i * 2].weight
-            nn.init.kaiming_normal_(self.conv[i * 2].weight.data, nonlinearity = 'relu')
+            nn.init.kaiming_normal_(self.conv[i * 2].weight.data, nonlinearity="relu")
 
-        x=torch.ones(K)
-        self.mu = nn.Parameter(torch.tensor(x,requires_grad=True,dtype=torch.float).cpu())
-        self.lip = torch.tensor(torch.ones(K),requires_grad=False,dtype=torch.float).cpu()
+        x = torch.ones(K)
+        self.mu = nn.Parameter(
+            torch.tensor(x, requires_grad=True, dtype=torch.float).cpu()
+        )
+        self.lip = torch.tensor(
+            torch.ones(K), requires_grad=False, dtype=torch.float
+        ).cpu()
         # apply He's initialization
         for i in range(K):
-            nn.init.kaiming_normal_(
-                self.conv[i].weight.data, nonlinearity='relu')
+            nn.init.kaiming_normal_(self.conv[i].weight.data, nonlinearity="relu")
 
-        outputdir= 'checkpoints/unfolded_ScCP_ver2/unfolded_ScCP_ver2_F64_K20_batchsize200_param_34580_data_vary'
+        outputdir = "checkpoints/unfolded_ScCP_ver2/unfolded_ScCP_ver2_F64_K20_batchsize200_param_34580_data_vary"
         checkpoint_path = os.path.join(outputdir, "checkpoint.pth.tar")
-        checkpoint = torch.load(checkpoint_path,map_location=device)
-        self.load_state_dict(checkpoint['Net'])
-
[…]
