+++ deepinv/unfolded/deep_equilibrium.py
@@ -10,6 +10,11 @@ class BaseDEQ(BaseUnfold):
     r"""
     Base class for deep equilibrium (DEQ) algorithms. Child of :class:`deepinv.unfolded.BaseUnfold`.
 
+    .. note::
+
+        Since 0.3.1, instead of using this class, it is possible to set directly an optimization algorithm as a DEQ algorithm using the algorithm name e.g.
+        ``model = ProximalGradientDescent(data_fidelity, prior, ..., DEQ = True, ...)``.
+
     Enables to turn any fixed-point algorithm into a DEQ algorithm, i.e. an algorithm
     that can be virtually unrolled infinitely leveraging the implicit function theorem.
     The backward pass is performed using fixed point iterations to find solutions of the fixed-point equation
@@ -179,6 +184,7 @@ def DEQ_builder(
     :param deepinv.optim.Bregman bregman_potential: Bregman potential used for Bregman optimization algorithms such as Mirror Descent. Default: None, comes back to standart Euclidean optimization.
     :param kwargs: additional arguments to be passed to the :class:`deepinv.unfolded.BaseUnfold` class.
     """
+    raise DeprecationWarning("The DEQ_builder function will be deprecated in 0.3.1.  Instead of using this function, it is possible to define DEQ algorithms using the algorithm name e.g. model = ProximalGradientDescent(data_fidelity, prior, ..., DEQ=True, ...).")    
     iterator = create_iterator(
         iteration,
         prior=prior,
