+++ deepinv/optim/optim_iterators/primal_dual_CP.py
@@ -79,7 +79,7 @@ class CPIteration(OptimIterator):
         z = x + cur_params["beta"] * (x - x_prev)
         F = (
             self.F_fn(x, cur_data_fidelity, cur_prior, cur_params, y, physics)
-            if self.F_fn is not None and self.has_cost
+            if self.has_cost and self.F_fn is not None and cur_data_fidelity is not None and cur_prior is not None
             else None
         )
         return {"est": (x, z, u), "cost": F}
