+++ deepinv/optim/optimizers.py
@@ -142,6 +142,13 @@ class BaseOptim(Reconstructor):
     :param float eps_anderson_acc: regularization parameter of the Anderson acceleration step. Default: ``1e-4``.
     :param bool unfold: whether to unfold the algorithm and make the model parameters trainable. Default: ``False``.
     :param list trainable_params: list of the algorithmic parameters among the keys of the dictionery params_algo to be made trainable. Default: ``None``, which means that all parameters in params_algo are trainable. For no trainable parameters, set to an empty list ``[]``.
+    :param bool DEQ: whether to use a Deep Equilibrium approach as unfolding strategy i.e. the  algorithm is virtually unrolled infinitely leveraging the implicit function theorem. Default: ``False``.
+    :param bool DEQ_jacobian_free: whether to use a Jacobian-free approach for the backward pass in the Deep Equilibrium model. The expansive Jacobian is removed in the implicit differentiation theorem. See https://ojs.aaai.org/index.php/AAAI/article/view/20619. Default: ``False``.
+    :param bool DEQ_anderson_acceleration_backward: whether to use Anderson acceleration for the backward pass in the Deep Equilibrium model. Default: ``False``.
+    :param int DEQ_history_size_backward: size of the history of iterates used for Anderson acceleration in the backward pass in the Deep Equilibrium model. Default: ``5``.
+    :param float DEQ_beta_anderson_acc_backward: momentum of the Anderson acceleration step in the backward pass in the Deep Equilibrium model. Default: ``1.0``.
+    :param float DEQ_eps_anderson_acc_backward: regularization parameter of the Anderson acceleration step in the backward pass in the Deep Equilibrium model. Default: ``1e-4``.
+    :param int DEQ_max_iter_backward: maximum number of iterations for the backward pass in the Deep Equilibrium model.
     :param bool verbose: whether to print relevant information of the algorithm during its run,
         such as convergence criterion at each iterate. Default: ``False``.
     :return: a torch model that solves the optimization problem.
@@ -169,7 +176,14 @@ class BaseOptim(Reconstructor):
         beta_anderson_acc=1.0,
         eps_anderson_acc=1e-4,
         unfold=False,
+        DEQ=False,
         trainable_params=None,
+        DEQ_jacobian_free=False,
+        DEQ_max_iter_backward=50,
+        DEQ_anderson_acceleration_backward=False,
+        DEQ_history_size_backward=5,
+        DEQ_beta_anderson_acc_backward=1.0,
+        DEQ_eps_anderson_acc_backward=1e-4,
         verbose=False,
         device=torch.device("cpu"),
         **kwargs,
@@ -190,6 +204,14 @@ class BaseOptim(Reconstructor):
         self.get_output = get_output
         self.has_cost = has_cost
         self.unfold = unfold
+        self.DEQ = DEQ
+        self.DEQ_jacobian_free = DEQ_jacobian_free
+        self.DEQ_max_iter_backward = DEQ_max_iter_backward
+        self.DEQ_anderson_acceleration_backward = DEQ_anderson_acceleration_backward
+        self.DEQ_history_size_backward = DEQ_history_size_backward
+        self.DEQ_beta_anderson_acc_backward = DEQ_beta_anderson_acc_backward
+        self.DEQ_eps_anderson_acc_backward = DEQ_eps_anderson_acc_backward
+        self.device = device
 
         # By default, ``self.prior`` should be a list of elements of the class :meth:`deepinv.optim.Prior`. The user could want the prior to change at each iteration. If no prior is given, we set it to a zero prior.
         if prior is None:
@@ -258,7 +280,7 @@ class BaseOptim(Reconstructor):
         self.init_params_algo = params_algo
 
         # set trainable parameters
-        if self.unfold:
+        if self.unfold or self.DEQ:
[…]

[…]
