### 2025-06-05 ffb20e8
julian comments to improve ensure docs
