+++ deepinv/unfolded/unfolded.py
@@ -190,7 +190,9 @@ def unfolded_builder(
 
 
     """
-    raise DeprecationWarning("The unfolded_builder function will be deprecated in 0.3.1.  Instead of using this function, it is possible to define unfolded algorithms using the algorithm name e.g. model = ProximalGradientDescent(data_fidelity, prior, ..., unfold=True, ...).")    
+    raise DeprecationWarning(
+        "The unfolded_builder function will be deprecated in 0.3.1.  Instead of using this function, it is possible to define unfolded algorithms using the algorithm name e.g. model = ProximalGradientDescent(data_fidelity, prior, ..., unfold=True, ...)."
+    )
     iterator = create_iterator(
         iteration,
         prior=prior,
