+++ deepinv/optim/optim_iterators/pgd.py
@@ -58,7 +58,9 @@ class FISTAIteration(OptimIterator):
         self.f_step = fStepPGD(**kwargs)
         self.a = a
         if self.a != 3:
-            raise DeprecationWarning('setting the a parameter in the FISTAIteration is deprecated, it should now be set in the cur_params dictionary instead.')
+            raise DeprecationWarning(
+                "setting the a parameter in the FISTAIteration is deprecated, it should now be set in the cur_params dictionary instead."
+            )
 
     def forward(
         self, X, cur_data_fidelity, cur_prior, cur_params, y, physics, *args, **kwargs
@@ -90,7 +92,10 @@ class FISTAIteration(OptimIterator):
 
         F = (
             self.F_fn(x, cur_data_fidelity, cur_prior, cur_params, y, physics)
-            if self.has_cost and self.F_fn is not None and cur_data_fidelity is not None and cur_prior is not None
+            if self.has_cost
+            and self.F_fn is not None
+            and cur_data_fidelity is not None
+            and cur_prior is not None
             else None
         )
 
