+++ deepinv/optim/optim_iterators/gradient_descent.py
@@ -50,7 +50,10 @@ class GDIteration(OptimIterator):
         x = x_prev - grad
         F = (
             self.F_fn(x, cur_data_fidelity, cur_prior, cur_params, y, physics)
-            if self.has_cost and self.F_fn is not None and cur_data_fidelity is not None and cur_prior is not None
+            if self.has_cost
+            and self.F_fn is not None
+            and cur_data_fidelity is not None
+            and cur_prior is not None
             else None
         )
         return {"est": (x,), "cost": F}
@@ -108,7 +111,10 @@ class MDIteration(OptimIterator):
         x = self.bregman_potential.grad_conj(self.bregman_potential.grad(x_prev) - grad)
         F = (
             self.F_fn(x, cur_data_fidelity, cur_prior, cur_params, y, physics)
-            if self.has_cost and self.F_fn is not None and cur_data_fidelity is not None and cur_prior is not None
+            if self.has_cost
+            and self.F_fn is not None
+            and cur_data_fidelity is not None
+            and cur_prior is not None
             else None
         )
         return {"est": (x,), "cost": F}
