+++ deepinv/tests/test_unfolded.py
@@ -109,10 +109,9 @@ def test_unfolded(unfolded_algo, imsize, dummy_dataset, device):
     loss.backward()
 
 
-DEQ_ALGO = [
-    "ProximalGradientDescent",
-    "HQS"
-]
+DEQ_ALGO = ["ProximalGradientDescent", "HQS"]
+
+
 @pytest.mark.parametrize("unfolded_algo", DEQ_ALGO)
 def test_DEQ(unfolded_algo, imsize, dummy_dataset, device):
     pytest.importorskip("ptwt")
