+++ deepinv/optim/optim_iterators/drs.py
@@ -58,7 +58,10 @@ class DRSIteration(OptimIterator):
         z = z + cur_params["beta"] * (x - u)
         F = (
             self.F_fn(x, cur_data_fidelity, cur_prior, cur_params, y, physics)
-            if self.has_cost and self.F_fn is not None and cur_data_fidelity is not None and cur_prior is not None
+            if self.has_cost
+            and self.F_fn is not None
+            and cur_data_fidelity is not None
+            and cur_prior is not None
             else None
         )
         return {"est": (x, z), "cost": F}

[…]
