+++ docs/source/conf.py
@@ -44,7 +44,6 @@ extensions = [
     "sphinx_copybutton",
     "sphinx_design",
 ]
-
 copybutton_exclude = ".linenos, .gp"
 
 intersphinx_mapping = {
