+++ CHANGELOG.rst
@@ -4,8 +4,20 @@ Change Log
 This change log is for the `main` branch. It contains changes for each release, with the date and author of each change.
 
 
-
 Current
+----------------
+
+New Features
+^^^^^^^^^^^^
+
+Changed
+^^^^^^^
+
+Fixed
+^^^^^
+
+
+v0.3.1
 ----------------
 
 New Features
