### 2025-06-05 4f1a9e2
fix the reported total loss
