+++ deepinv/optim/optim_iterators/mace.py
@@ -1,0 +1,186 @@
+# -*- coding: utf-8 -*-
+import torch
+from .optim_iterator import OptimIterator
+import warnings
+
+
+class MACEIteration(OptimIterator):
+    r"""Iterator for Multi-Agent Consensus Equilibrium (MACE).
+
+    This iterator implements the MACE algorithm. The update is based on Mann iterations:
+
+    .. math::
+        v_i^{(k+1)} = (1-\rho)v_i^{(k)} + \rho \left( (2G-I)(2F-I)v^{(k)} \right)_i
+
+    where :math:`v^{(k)}` is a list of variables, one for each agent,
+    :math:`F` applies each agent's proximal operator to its corresponding :math:`v_i`,
+    and :math:`G` is an averaging operator based on weights :math:`\mu`.
+
+    This iterator expects the following in ``cur_params`` from ``params_algo``:
+    - ``rho``: Mann iteration relaxation parameter (float).
+    - ``mu``: List of weights for averaging agent outputs, summing to 1 (list of floats).
+    - ``stepsize``: Base stepsize (gamma) for data fidelity agents (float or list).
+    - ``lambda``: Regularization parameter for prior agents (float or list).
+    - ``g_param``: Parameter for prior agents, e.g., noise level for PnP (any type, or list).
+
+    The agents themselves (data fidelity and prior objects) are configured at initialization.
+    """
+
+    def __init__(self, data_fidelity_list, prior_list, **kwargs):
+        super().__init__(F_fn=None, has_cost=False)
+        self.v_list = None
+        self.mu = None
+
+        self.agents_ops = []
+        self.agents_gammas_keys = []
+        self.agents_sigmas_keys = []
+        self._y_ref = None
+        self._physics_ref = None
+
+        self._configure_agents(data_fidelity_list, prior_list)
+        self.num_agents = len(self.agents_ops)
+        if self.num_agents == 0:
+            raise ValueError(
+                "MACEIteration requires at least one data_fidelity or prior agent."
+            )
+
+        self.requires_prox_g = any(ptype == "prior" for ptype, _ in self.agents_ops)
+        
+        # Build the cost function and check if it's valid
+        self.F_fn = self._build_cost_function()
+ 
+    def _build_cost_function(self):
+        """
+        Builds a cost function that sums the data fidelity and all explicit prior costs.
+        """
+        def F_fn(x, data_fidelity, prior, cur_params, y, physics):
+            total_cost = 0
+            prior_idx_counter = 0
+            # We need to iterate over the configured agents
+            for agent_type, agent_obj in self.agents_ops:
+                if agent_type == 'data_fidelity':
+                    total_cost += agent_obj(x, y, physics)
+                elif agent_type == 'prior':
+                    if hasattr(agent_obj, 'explicit_prior') and agent_obj.explicit_prior:
+                        lambda_key = f"lambda_{prior_idx_counter}"
+                        lambda_val = cur_params.get(lambda_key, cur_params.get("lambda", 1.0))
+                        g_param = cur_params.get("g_param")
+                        prior_cost = agent_obj(x, g_param)
+                        total_cost += (lambda_val * prior_cost).sum()
+                    prior_idx_counter += 1
+            return total_cost
+        return F_fn
+
+    def _configure_agents(self, data_fidelity_list, prior_list):
+        """Collects and configures all agent operations from initial lists."""
+        df_list = (
+            data_fidelity_list
[…]
