### 2025-06-06 32f8c09
demo PnP custom optim
