+++ deepinv/optim/optimizers.py
@@ -74,6 +74,7 @@ class BaseOptim(Reconstructor):
 
     If ``DEQ`` is set to ``True``, the algorithm is unfolded as a Deep Equilibrium model, i.e. the algorithm is virtually unrolled infinitely leveraging the implicit function theorem.
     The backward pass is then performed using fixed point iterations to find solutions of the fixed-point equation
+    
     .. math::
 
         \begin{equation}
@@ -749,7 +750,7 @@ def optim_builder(
         Since 0.3.1, instead of using this function, it is possible to define optimization algorithms using directly the algorithm name e.g.
         ``model = ProximalGradientDescent(data_fidelity, prior, ...)``.
 
-    :param str, deepinv.optim.OptimIterator iteration: either the name of the algorithm to be used,
+    :param str, deepinv.optim.optim_iterators.OptimIterator iteration: either the name of the algorithm to be used,
         or directly an optim iterator.
         If an algorithm name (string), should be either ``"GD"`` (gradient descent),
         ``"PGD"`` (proximal gradient descent), ``"ADMM"`` (ADMM),
@@ -819,7 +820,7 @@ class ADMM(BaseOptim):
         \end{equation*}
 
     where :math:`\gamma>0` is a stepsize and :math:`\beta>0` is a relaxation parameter.  If the attribute ``g_first`` is set to ``True``, the functions :math:`f` and :math:`\regname` are
-    inverted in the previous iterations. The ADMM iterations are defined in the iterator class :class:`deepinv.optim.ADMMIteration`.
+    inverted in the previous iterations. The ADMM iterations are defined in the iterator class :class:`deepinv.optim.optim_iterators.ADMMIteration`.
 
     If the attribute ``unfold`` is set to ``True``, the algorithm is unfolded and the algorithmic parameters of the algorithm are trainable.
     By default, all the algorithm parameters are trainiable : the stepsize :math:`\gamma`, the regularization parameter :math:`\lambda`, the prior parameter and the relaxation parameter :math:`\beta`.
@@ -903,7 +904,7 @@ class DRS(BaseOptim):
         \end{equation*}
 
     where :math:`\gamma>0` is a stepsize and :math:`\beta>0` is a relaxation parameter. If the attribute ``g_first`` is set to True, the functions :math:`f` and :math:`\regname` are inverted in the previous iteration.
-    The DRS iterations are defined in the iterator class :class:`deepinv.optim.DRSIteration`.
+    The DRS iterations are defined in the iterator class :class:`deepinv.optim.optim_iterators.DRSIteration`.
 
     If the attribute ``unfold`` is set to ``True``, the algorithm is unfolded and the parameters of the algorithm are trainable.
     By default, all the algorithm parameters are trainiable : the stepsize :math:`\gamma`, the regularization parameter :math:`\lambda`, the prior parameter and the relaxation parameter :math:`\beta`.
@@ -965,20 +966,24 @@ class DRS(BaseOptim):
 class GradientDescent(BaseOptim):
     r"""
     Gradient Descent module for solving the problem
+    
     .. math::
         \begin{equation}
         \label{eq:min_prob}
         \tag{1}
         \underset{x}{\arg\min} \quad  \datafid{x}{y} + \lambda \reg{x},
         \end{equation}
+    
     where :math:`\datafid{x}{y}` is the data-fidelity term, :math:`\reg{x}` is the regularization term.
 
     The Gradient Descent iterations are given by
+    
     .. math::
         \begin{equation*}
[…]
