+++ docs/source/user_guide/reconstruction/optimization.rst
@@ -291,11 +291,11 @@ Some predefined optimizers are provided:
        | :math:`x_{k+1} = \nabla h^*(\nabla h(x_k) - \gamma v_{k})`
      - ``"stepsize"``, ``"lambda"``, ``"g_param"``
 
-  * - :class:`Proximal Mirror Descent (MD) <deepinv.optim.optim_iterators.PMDIteration>`
-    - | :math:`v_{k} = \nabla f(x_k) + \lambda \nabla \reg{x_k}`
-      | :math:`u_{k} = \nabla h^*(\nabla h(x_k) - \gamma v_{k})`
-      | :math:`x_{k+1} = \operatorname{prox^h}_{\gamma \lambda \regname}(u_k)`
-    - ``"stepsize"``, ``"lambda"``, ``"g_param"``
+   * - :class:`Proximal Mirror Descent (MD) <deepinv.optim.optim_iterators.PMDIteration>`
+     - | :math:`v_{k} = \nabla f(x_k) + \lambda \nabla \reg{x_k}`
+       | :math:`u_{k} = \nabla h^*(\nabla h(x_k) - \gamma v_{k})`
+       | :math:`x_{k+1} = \operatorname{prox^h}_{\gamma \lambda \regname}(u_k)`
+     - ``"stepsize"``, ``"lambda"``, ``"g_param"``
 
    * - :class:`Spectral Methods (SM) <deepinv.optim.optim_iterators.SMIteration>`
      - :math:`M = \conj{B} \text{diag}(T(y)) B + \lambda I`
