### 2025-06-06 6470337
fix the reported total loss (#515)
