### 2025-06-06 f1283d9
make adjoint_function use A in autograd
