### 2025-06-07 fc76037
skip radio physics which can't be autodiffed
