### 2025-06-07 d6ff61f
fix the gradient norm reported by Trainer
