### 2025-06-07 b88f6aa
Make adjoint_function use A in autograd (#519)
* make adjoint_function use A in autograd

* use mandatory setup_context

* skip radio physics which can't be autodiffed
