+++ _downloads/4c16f979114fe6fc154ea1a7f9b6d843/demo_phase_retrieval.py
@@ -1,0 +1,281 @@
+r"""
+Random phase retrieval and reconstruction methods.
+===================================================
+
+This example shows how to create a random phase retrieval operator and generate phaseless measurements from a given image. The example showcases 4 different reconstruction methods to recover the image from the phaseless measurements:
+
+#. Gradient descent with random initialization;
+#. Spectral methods;
+#. Gradient descent with spectral methods initialization;
+#. Gradient descent with PnP denoisers.
+
+"""
+
+# %%
+# General setup
+# ----------------------------
+import deepinv as dinv
+from pathlib import Path
+import torch
+import matplotlib.pyplot as plt
+from deepinv.models import DRUNet
+from deepinv.optim.data_fidelity import L2
+from deepinv.optim.prior import PnP, Zero
+from deepinv.optim.optimizers import optim_builder
+from deepinv.utils.demo import load_example
+from deepinv.utils.plotting import plot
+from deepinv.optim.phase_retrieval import (
+    correct_global_phase,
+    cosine_similarity,
+    spectral_methods,
+)
+from deepinv.models.complex import to_complex_denoiser
+
+BASE_DIR = Path(".")
+RESULTS_DIR = BASE_DIR / "results"
+# Set global random seed to ensure reproducibility.
+torch.manual_seed(0)
+
+device = dinv.utils.get_freer_gpu() if torch.cuda.is_available() else "cpu"
+
+# %%
+# Load image from the internet
+# ----------------------------
+#
+# We use the standard test image "Shepp–Logan phantom".
+
+# Image size
+img_size = 32
+# The pixel values of the image are in the range [0, 1].
+x = load_example(
+    "SheppLogan.png",
+    img_size=img_size,
+    grayscale=True,
+    resize_mode="resize",
+    device=device,
+)
+print(x.min(), x.max())
+
+# %%
+# Visualization
+# ---------------------------------------
+#
+# We use the customized plot() function in deepinv to visualize the original image.
+plot(x, titles="Original image")
+
+# %%
+# Signal construction
+# ---------------------------------------
+# We use the original image as the phase information for the complex signal. The original value range is [0, 1], and we map it to the phase range [-pi/2, pi/2].
+x_phase = torch.exp(1j * x * torch.pi - 0.5j * torch.pi)
+
+# Every element of the signal should have unit norm.
+assert torch.allclose(x_phase.real**2 + x_phase.imag**2, torch.tensor(1.0))
+
+# %%
+# Measurements generation
+# ---------------------------------------
+# Create a random phase retrieval operator with an
+# oversampling ratio (measurements/pixels) of 5.0,
+# and generate measurements from the signal with additive Gaussian noise.
+
+# Define physics information
+oversampling_ratio = 5.0
+img_shape = x.shape[1:]
+m = int(oversampling_ratio * torch.prod(torch.tensor(img_shape)))
+n_channels = 1  # 3 for color images, 1 for gray-scale images
+
+# Create the physics
+physics = dinv.physics.RandomPhaseRetrieval(
+    m=m,
+    img_shape=img_shape,
+    device=device,
+)
+
+# Generate measurements
+y = physics(x_phase)
+
+# %%
+# Reconstruction with gradient descent and random initialization
+# ---------------------------------------------------------------
+# First, we use the function :class:`deepinv.optim.L2` as the data fidelity function, and the class :class:`deepinv.optim.optim_iterators.GDIteration` as the optimizer to run a gradient descent algorithm. The initial guess is a random complex signal.
+
+data_fidelity = L2()
[…]
