+++ _downloads/faf1222544a1ffa01f52a6b80e689f71/demo_dps.py
@@ -1,0 +1,361 @@
+r"""
+Implementing DPS
+================
+
+In this tutorial, we will go over the steps in the Diffusion Posterior Sampling (DPS) algorithm introduced in
+`Chung et al. <https://arxiv.org/abs/2209.14687>`_ The full algorithm is implemented in
+:class:`deepinv.sampling.DPS`.
+"""
+
+# %%
+# Installing dependencies
+# -----------------------
+# Let us ``import`` the relevant packages, and load a sample
+# image of size 64 x 64. This will be used as our ground truth image.
+#
+# .. note::
+#           We work with an image of size 64 x 64 to reduce the computational time of this example.
+#           The algorithm works best with images of size 256 x 256.
+#
+
+import numpy as np
+import torch
+
+import deepinv as dinv
+from deepinv.utils.plotting import plot
+from deepinv.optim.data_fidelity import L2
+from deepinv.utils.demo import load_example
+from tqdm import tqdm  # to visualize progress
+
+device = dinv.utils.get_freer_gpu() if torch.cuda.is_available() else "cpu"
+
+x_true = load_example("butterfly.png", img_size=64).to(device)
+x = x_true.clone()
+
+# %%
+# In this tutorial we consider random inpainting as the inverse problem, where the forward operator is implemented
+# in :class:`deepinv.physics.Inpainting`. In the example that we use, 90% of the pixels will be masked out randomly,
+# and we will additionally have Additive White Gaussian Noise (AWGN) of standard deviation  12.75/255.
+
+sigma = 12.75 / 255.0  # noise level
+
+physics = dinv.physics.Inpainting(
+    tensor_size=(3, x.shape[-2], x.shape[-1]),
+    mask=0.1,
+    pixelwise=True,
+    device=device,
+)
+
+y = physics(x_true)
+
+imgs = [y, x_true]
+plot(
+    imgs,
+    titles=["measurement", "groundtruth"],
+)
+
+
+# %%
+# Diffusion model loading
+# -----------------------
+#
+# We will take a pre-trained diffusion model that was also used for the DiffPIR algorithm, namely the one trained on
+# the FFHQ 256x256 dataset. Note that this means that the diffusion model was trained with human face images,
+# which is very different from the image that we consider in our example. Nevertheless, we will see later on that
+# ``DPS`` generalizes sufficiently well even in such case.
+
+
+model = dinv.models.DiffUNet(large_model=False).to(device)
+
+# %%
+# Define diffusion schedule
+# -------------------------
+#
+# We will use the standard linear diffusion noise schedule. Once :math:`\beta_t` is defined to follow a linear schedule
+# that interpolates between :math:`\beta_{\rm min}` and :math:`\beta_{\rm max}`,
+# we have the following additional definitions:
+# :math:`\alpha_t := 1 - \beta_t`, :math:`\bar\alpha_t := \prod_{j=1}^t \alpha_j`.
+# The following equations will also be useful
+# later on (we always assume that :math:`\mathbf{\epsilon} \sim \mathcal{N}(\mathbf{0}, \mathbf{I})` hereafter.)
+#
+# .. math::
+#
[…]
