+++ _downloads/873d523bccb69fb019bf7f5379602f26/demo_vanilla_PnP.py
@@ -1,0 +1,155 @@
+"""
+Vanilla PnP for computed tomography (CT).
+====================================================================================================
+
+This example shows how to use a standart PnP algorithm with DnCNN denoiser for computed tomography.
+
+"""
+
+import deepinv as dinv
+from pathlib import Path
+import torch
+from deepinv.models import DnCNN
+from deepinv.optim.data_fidelity import L2
+from deepinv.optim.prior import PnP
+from deepinv.optim.optimizers import optim_builder
+from deepinv.utils.demo import load_example
+from deepinv.utils.plotting import plot, plot_curves
+
+# %%
+# Setup paths for data loading and results.
+# ----------------------------------------------------------------------------------------
+#
+BASE_DIR = Path(".")
+RESULTS_DIR = BASE_DIR / "results"
+
+# %%
+# Load image and parameters
+# ----------------------------------------------------------------------------------------
+
+# Set the global random seed from pytorch to ensure reproducibility of the example.
+torch.manual_seed(0)
+device = dinv.utils.get_freer_gpu() if torch.cuda.is_available() else "cpu"
+# Set up the variable to fetch dataset and operators.
+method = "PnP"
+img_size = 32
+x = load_example(
+    "SheppLogan.png",
+    img_size=img_size,
+    grayscale=True,
+    resize_mode="resize",
+    device=device,
+)
+operation = "tomography"
+
+
+# %%
+# Set the forward operator
+# --------------------------------------------------------------------------------
+# We use the :class:`deepinv.physics.Tomography`
+# class from the physics module to generate a CT measurements.
+
+
+noise_level_img = 0.03  # Gaussian Noise standard deviation for the degradation
+angles = 100
+n_channels = 1  # 3 for color images, 1 for gray-scale images
+physics = dinv.physics.Tomography(
+    img_width=img_size,
+    angles=angles,
+    circle=False,
+    device=device,
+    noise_model=dinv.physics.GaussianNoise(sigma=noise_level_img),
+)
+
+PI = 4 * torch.ones(1).atan()
+SCALING = (PI / (2 * angles)).to(device)  # approximate operator norm of A^T A
+
+# Use parallel dataloader if using a GPU to fasten training,
+# otherwise, as all computes are on CPU, use synchronous data loading.
+num_workers = 4 if torch.cuda.is_available() else 0
+
+
+# %%
+# Set up the PnP algorithm to solve the inverse problem.
+# --------------------------------------------------------------------------------
+# We use the Proximal Gradient Descent optimization algorithm.
+# The algorithm alternates between a denoising step and a gradient descent step.
+# The denoising step is performed by a DNCNN pretrained denoiser :class:`deepinv.models.DnCNN`.
+#
+# Set up the PnP algorithm parameters : the ``stepsize``, ``g_param`` the noise level of the denoiser.
+# Attention: The choice of the stepsize is crucial as it also defines the amount of regularization.  Indeed, the regularization parameter ``lambda`` is implicitly defined by the stepsize.
+# Both the stepsize and the noise level of the denoiser control the regularization power and should be tuned to the specific problem.
+# The following parameters have been chosen manually.
+params_algo = {"stepsize": 0.01 * SCALING, "g_param": noise_level_img}
+max_iter = 100
+early_stop = True
+
+# Select the data fidelity term
+data_fidelity = L2()
+
+# Specify the denoising prior
+denoiser = DnCNN(
+    in_channels=n_channels,
+    out_channels=n_channels,
[…]

[…]
