### 2025-06-07 bb03fd3
add test suite profiling
