+++ .github/workflows/test.yml
@@ -45,15 +45,23 @@ jobs:
       - name: Install deepinv and its dependencies
         run: |
           pip install .[${{ matrix.extra }}]
+          pip install pytest-profiling
 
       - name: Test with pytest and generate coverage report
         run: |
-          python3 -m pytest --cov=deepinv --junitxml=junit.xml -o junit_family=legacy -v
+          python3 -m pytest --cov=deepinv --junitxml=junit.xml -o junit_family=legacy --profile -v
+
+      - name: Upload combined profile
+        if: ${{ always() }}
+        uses: actions/upload-artifact@v4
+        with:
+          name: combined-prof-${{ matrix.os }}
+          path: prof/combined.prof
+
       - name: Upload coverage to Codecov
         uses: codecov/codecov-action@v5
 
       - name: Upload test results to Codecov
         if: ${{ !cancelled() }}
         uses: codecov/test-results-action@v1
-
 
