### 2025-06-07 c9a7f8e
make test_sde run in parallel of other tests
