### 2025-06-07 ec79e86
add custom marker to pytest config
