### 2025-06-07 7721e83
set the groups before they are read
