+++ docs/source/conf.py
@@ -64,7 +64,8 @@ autodoc_preserve_defaults = True
 nitpicky = True
 # Create link to the API in the auto examples
 autodoc_inherit_docstrings = False
-
+# for sitemap
+html_baseurl = 'https://deepinv.github.io/deepinv/'
 
 ####  userguide directive ###
 from docutils import nodes
