### 2025-06-09 ab1de5f
Add sitemap for docs (#523)
* add sitemap

* address

* black
