+++ deepinv/models/ram.py
@@ -1,0 +1,833 @@
+import torch
+import torch.nn as nn
+import torch.nn.functional as F
+
+import deepinv as dinv
+from deepinv.physics import Physics, LinearPhysics, Downsampling
+from deepinv.utils.tensorlist import TensorList
+
+from huggingface_hub import hf_hub_download
+
+cuda = True if torch.cuda.is_available() else False
+Tensor = torch.cuda.FloatTensor if cuda else torch.FloatTensor
+
+
+class RAM(nn.Module):
+    r"""
+    RAM model
+
+    This model is a convolutional neural network (CNN) designed for image reconstruction tasks.
+
+    :param in_channels: Number of input channels. If a list is provided, the model will have separate heads for each channel.
+    :param device: Device to which the model should be moved. If None, the model will be created on the default device.
+    :param pretrained: If True, the model will be initialized with pretrained weights.
+    """
+    def __init__(
+        self,
+        in_channels=[1, 2, 3],
+        device=None,
+        pretrained=True,
+    ):
+        super(RAM, self).__init__()
+
+        nc = [64, 128, 256, 512] # number of channels in the network
+        self.in_channels = in_channels
+        self.fact_realign = torch.nn.Parameter(torch.tensor([1.0], device=device))
+
+        self.separate_head = isinstance(in_channels, list)
+
+        if isinstance(in_channels, list):
+            in_channels_first = []
+            for i in range(len(in_channels)):
+                in_channels_first.append(in_channels[i] + 2)
+
+        # check if in_channels is a list
+        self.m_head = InHead(in_channels_first, nc[0])
+
+        self.m_down1 = BaseEncBlock(nc[0], nc[0], img_channels=in_channels, decode_upscale=1)
+        self.m_down2 = BaseEncBlock(nc[1], nc[1], img_channels=in_channels, decode_upscale=2)
+        self.m_down3 = BaseEncBlock(nc[2], nc[2], img_channels=in_channels, decode_upscale=4)
+        self.m_body = BaseEncBlock(nc[3], nc[3], img_channels=in_channels, decode_upscale=8)
+        self.m_up3 = BaseEncBlock(nc[2], nc[2], img_channels=in_channels, decode_upscale=4)
+        self.m_up2 = BaseEncBlock(nc[1], nc[1], img_channels=in_channels, decode_upscale=2)
+        self.m_up1 = BaseEncBlock(nc[0], nc[0], img_channels=in_channels, decode_upscale=1)
+
+        self.pool1 = downsample_strideconv(nc[0], nc[1], bias=False, mode="2")
+        self.pool2 = downsample_strideconv(nc[1], nc[2], bias=False, mode="2")
+        self.pool3 = downsample_strideconv(nc[2], nc[3], bias=False, mode="2")
+        self.up3 = upsample_convtranspose(nc[3], nc[2], bias=False, mode="2")
+        self.up2 = upsample_convtranspose(nc[2], nc[1], bias=False, mode="2")
+        self.up1 = upsample_convtranspose(nc[1], nc[0], bias=False, mode="2")
+
+        self.m_tail = OutTail(nc[0], in_channels)
+
+        # load pretrained weights from hugging face
+        if pretrained:
+            self.load_state_dict(torch.load(hf_hub_download(repo_id="mterris/ram", filename="ram.pth.tar"), map_location=device))
+
+        if device is not None:
+            self.to(device)
+
+    def constant2map(self, value, x):
+        r"""
+        Converts a constant value to a map of the same size as the input tensor x.
+
+        :params float value: constant value
+        :params torch.Tensor x: input tensor
+        """
+        if isinstance(value, torch.Tensor):
+            if value.ndim > 0:
+                value_map = value.view(x.size(0), 1, 1, 1)
[…]
