### 2025-06-10 9297cac
Fix the gradient norm reported by the trainer (#520)
* fix the gradient norm reported by Trainer

* increase error tolerance
