+++ deepinv/models/ram.py
@@ -14,14 +14,16 @@ Tensor = torch.cuda.FloatTensor if cuda else torch.FloatTensor
 
 class RAM(nn.Module):
     r"""
-    RAM model
+    RAM model.
 
-    This model is a convolutional neural network (CNN) designed for image reconstruction tasks.
+    This model (proposed in `this paper <https://arxiv.org/abs/2503.08915>`_) is a convolutional neural network (CNN)
+    designed for image reconstruction tasks.
 
     :param in_channels: Number of input channels. If a list is provided, the model will have separate heads for each channel.
     :param device: Device to which the model should be moved. If None, the model will be created on the default device.
     :param pretrained: If True, the model will be initialized with pretrained weights.
     """
+
     def __init__(
         self,
         in_channels=[1, 2, 3],
@@ -30,7 +32,7 @@ class RAM(nn.Module):
     ):
         super(RAM, self).__init__()
 
-        nc = [64, 128, 256, 512] # number of channels in the network
+        nc = [64, 128, 256, 512]  # number of channels in the network
         self.in_channels = in_channels
         self.fact_realign = torch.nn.Parameter(torch.tensor([1.0], device=device))
 
@@ -44,13 +46,27 @@ class RAM(nn.Module):
         # check if in_channels is a list
         self.m_head = InHead(in_channels_first, nc[0])
 
-        self.m_down1 = BaseEncBlock(nc[0], nc[0], img_channels=in_channels, decode_upscale=1)
-        self.m_down2 = BaseEncBlock(nc[1], nc[1], img_channels=in_channels, decode_upscale=2)
-        self.m_down3 = BaseEncBlock(nc[2], nc[2], img_channels=in_channels, decode_upscale=4)
-        self.m_body = BaseEncBlock(nc[3], nc[3], img_channels=in_channels, decode_upscale=8)
-        self.m_up3 = BaseEncBlock(nc[2], nc[2], img_channels=in_channels, decode_upscale=4)
-        self.m_up2 = BaseEncBlock(nc[1], nc[1], img_channels=in_channels, decode_upscale=2)
-        self.m_up1 = BaseEncBlock(nc[0], nc[0], img_channels=in_channels, decode_upscale=1)
+        self.m_down1 = BaseEncBlock(
+            nc[0], nc[0], img_channels=in_channels, decode_upscale=1
+        )
+        self.m_down2 = BaseEncBlock(
+            nc[1], nc[1], img_channels=in_channels, decode_upscale=2
+        )
+        self.m_down3 = BaseEncBlock(
+            nc[2], nc[2], img_channels=in_channels, decode_upscale=4
+        )
+        self.m_body = BaseEncBlock(
+            nc[3], nc[3], img_channels=in_channels, decode_upscale=8
+        )
+        self.m_up3 = BaseEncBlock(
+            nc[2], nc[2], img_channels=in_channels, decode_upscale=4
+        )
+        self.m_up2 = BaseEncBlock(
+            nc[1], nc[1], img_channels=in_channels, decode_upscale=2
+        )
+        self.m_up1 = BaseEncBlock(
+            nc[0], nc[0], img_channels=in_channels, decode_upscale=1
+        )
 
         self.pool1 = downsample_strideconv(nc[0], nc[1], bias=False, mode="2")
         self.pool2 = downsample_strideconv(nc[1], nc[2], bias=False, mode="2")
@@ -63,7 +79,12 @@ class RAM(nn.Module):
 
         # load pretrained weights from hugging face
         if pretrained:
-            self.load_state_dict(torch.load(hf_hub_download(repo_id="mterris/ram", filename="ram.pth.tar"), map_location=device))
+            self.load_state_dict(
+                torch.load(
[…]
