+++ deepinv/tests/test_models.py
@@ -27,6 +27,7 @@ MODEL_LIST = MODEL_LIST_1_CHANNEL + [
     "unet",
     "waveletdict_hard",
     "waveletdict_topk",
+    "dsccp",
 ]
 
 
@@ -685,5 +686,22 @@ def test_ncsnpp_net(device, image_size, n_channels, batch_size, precond, use_fp1
     assert y.shape == x.shape
 
     y = model(x, torch.tensor([0.01]))
+    # Check the output tensor shape
+    assert y.shape == x.shape
+
+@pytest.mark.parametrize("device", [torch.device("cpu")])
+@pytest.mark.parametrize("n_channels", [3])
+def test_dsccp_net(device, n_channels):
+    # Load the pretrained model
+
+    image_size = (3, 37, 28)
+    model = dinv.models.DScCP().to(device)
+    x = torch.rand(image_size, device=device).unsqueeze(0)
+
+    y = model(x, 0.01)
+    # Check the output tensor shape
+    assert y.shape == x.shape
+
+    y = model(x, torch.tensor(0.01))
     # Check the output tensor shape
     assert y.shape == x.shape
