+++ deepinv/physics/singlepixel.py
@@ -111,10 +111,11 @@ def hadamard_2d_ishift(x):
     return x
 
 
-def sequency_mask(img_shape, m):
+@_deprecated_alias(img_shape="img_size")
+def sequency_mask(img_size, m):
     """
     Generates a sequency-ordered binary mask for single-pixel imaging.
-    :param tuple img_shape: The shape of the input image as a tuple (C, H, W),
+    :param tuple img_size: The shape of the input image as a tuple (C, H, W),
                             where C is the number of channels, H is the height,
                             and W is the width.
     :param int m: The number of pixels to select based on sequency order.
@@ -123,7 +124,7 @@ def sequency_mask(img_shape, m):
     :rtype: torch.Tensor
     """
 
-    _, H, W = img_shape
+    _, H, W = img_size
     n = H * W
 
     indexes = sequency_order(n)[:m]
@@ -131,23 +132,24 @@ def sequency_mask(img_shape, m):
     i = i.flatten(order="F")
     j = j.flatten(order="F")
 
-    mask = torch.zeros((1, *img_shape))
+    mask = torch.zeros((1, *img_size))
     mask[:, :, i[indexes], j[indexes]] = 1.0
 
     return mask
 
 
-def old_sequency_mask(img_shape, m):
+@_deprecated_alias(img_shape="img_size")
+def old_sequency_mask(img_size, m):
     """
     Generates a binary mask for a single-pixel camera based on a sequency ordering.
-    :param tuple img_shape: The shape of the input image as a tuple (C, H, W), where C is the number of channels,
+    :param tuple img_size: The shape of the input image as a tuple (C, H, W), where C is the number of channels,
                             H is the height, and W is the width.
     :param int m: The number of pixels to include in the mask, selected based on the sequency ordering.
     :return: A binary mask of shape (1, C, H, W) with `m` pixels set to 1 and the rest set to 0.
     :rtype: torch.Tensor
     """
 
-    _, H, W = img_shape
+    _, H, W = img_size
     n = H * W
 
     indexes = get_permutation_list(n)[:m]
@@ -155,7 +157,7 @@ def old_sequency_mask(img_shape, m):
     i = i.flatten(order="F")
     j = j.flatten(order="F")
 
-    mask = torch.zeros((1, *img_shape))
+    mask = torch.zeros((1, *img_size))
     mask[:, :, i[indexes], j[indexes]] = 1.0
 
     return mask
@@ -200,10 +202,10 @@ def cake_cutting_order(n):
     return np.argsort(seq)
 
 
-def cake_cutting_mask(img_shape, m):
+def cake_cutting_mask(img_size, m):
     """
     Generates a "cake cutting" mask for single-pixel imaging.
-    :param tuple img_shape: A tuple representing the shape of the input image
+    :param tuple img_size: A tuple representing the shape of the input image
                             in the format (channels, height, width). The height
                             and width must be equal.
     :param int m: The number of pixels to include in the mask.
@@ -213,7 +215,7 @@ def cake_cutting_mask(img_shape, m):
     :raises Warning: If the height and width of the image are not equal.
     """
 
-    _, H, W = img_shape
+    _, H, W = img_size
 
     if H != W:
         warnings.warn("Image height and width must be equal for cake cutting mask.")
@@ -227,7 +229,7 @@ def cake_cutting_mask(img_shape, m):
     i = i.flatten(order="F")
     j = j.flatten(order="F")
 
-    mask = torch.zeros((1, *img_shape))
+    mask = torch.zeros((1, *img_size))
     mask[:, :, i[indexes], j[indexes]] = 1.0
 
     return mask
[…]
