+++ deepinv/physics/structured_random.py
@@ -7,20 +7,21 @@ from deepinv.physics.forward import LinearPhysics
 from deepinv.utils.decorators import _deprecated_alias
 
 
-def compare(input_shape: tuple, output_shape: tuple) -> str:
+@_deprecated_alias(input_shape="img_size", output_shape="output_size")
+def compare(img_size: tuple, output_size: tuple) -> str:
     r"""
     Compare input and output shape to determine the sampling mode.
 
-    :param tuple input_shape: Input shape (C, H, W).
-    :param tuple output_shape: Output shape (C, H, W).
+    :param tuple img_size: Input shape (C, H, W).
+    :param tuple output_size: Output shape (C, H, W).
 
     :return: The sampling mode in ["oversampling","undersampling","equisampling].
     """
-    if input_shape[1] == output_shape[1] and input_shape[2] == output_shape[2]:
+    if img_size[1] == output_size[1] and img_size[2] == output_size[2]:
         return "equisampling"
-    elif input_shape[1] <= output_shape[1] and input_shape[2] <= output_shape[2]:
+    elif img_size[1] <= output_size[1] and img_size[2] <= output_size[2]:
         return "oversampling"
-    elif input_shape[1] >= output_shape[1] and input_shape[2] >= output_shape[2]:
+    elif img_size[1] >= output_size[1] and img_size[2] >= output_size[2]:
         return "undersampling"
     else:
         raise ValueError(
@@ -28,43 +29,48 @@ def compare(input_shape: tuple, output_shape: tuple) -> str:
         )
 
 
-def padding(tensor: torch.Tensor, input_shape: tuple, output_shape: tuple):
+@_deprecated_alias(input_shape="img_size", output_shape="output_size")
+def padding(tensor: torch.Tensor, img_size: tuple, output_size: tuple):
     r"""
     Zero padding function for oversampling in structured random phase retrieval.
 
     :param torch.Tensor tensor: input tensor.
-    :param tuple input_shape: shape of the input tensor.
-    :param tuple output_shape: shape of the output tensor.
+    :param tuple img_size: shape of the input tensor.
+    :param tuple output_size: shape of the output tensor.
 
     :return: (:class:`torch.Tensor`) the zero-padded tensor.
     """
-    change_top = math.ceil(abs(input_shape[1] - output_shape[1]) / 2)
-    change_bottom = math.floor(abs(input_shape[1] - output_shape[1]) / 2)
-    change_left = math.ceil(abs(input_shape[2] - output_shape[2]) / 2)
-    change_right = math.floor(abs(input_shape[2] - output_shape[2]) / 2)
-    assert change_top + change_bottom == abs(input_shape[1] - output_shape[1])
-    assert change_left + change_right == abs(input_shape[2] - output_shape[2])
-    return torch.nn.ZeroPad2d((change_left, change_right, change_top, change_bottom))(
-        tensor
+    change_top = math.ceil(abs(img_size[1] - output_size[1]) / 2)
+    change_bottom = math.floor(abs(img_size[1] - output_size[1]) / 2)
+    change_left = math.ceil(abs(img_size[2] - output_size[2]) / 2)
+    change_right = math.floor(abs(img_size[2] - output_size[2]) / 2)
+    assert change_top + change_bottom == abs(img_size[1] - output_size[1])
+    assert change_left + change_right == abs(img_size[2] - output_size[2])
+    return torch.nn.functional.pad(
+        tensor,
+        (change_left, change_right, change_top, change_bottom),
+        mode="constant",
+        value=0,
     )
 
 
-def trimming(tensor: torch.Tensor, input_shape: tuple, output_shape: tuple):
[…]

[…]
