### 2025-06-10 e2c3db0
Speed up the test suite by parallelizing the slowest outlier (#522)
* make test_sde run in parallel of other tests

* black

* add custom marker to pytest config

* make ordering deterministic as mandatory

* set the groups before they are read

* update the changelog
