+++ CHANGELOG.rst
@@ -13,6 +13,7 @@ New Features
 Changed
 ^^^^^^^
 - Make autograd use the base linear operator for `deepinv.physics.adjoint_function` (:gh:`519` by `Jérémy Scanvic`_)
+- Parallelize the test suite making it 15% faster (:gh:`522` by `Jérémy Scanvic`_)
 
 Fixed
 ^^^^^
