### 2025-06-11 943a4b9
Fix documentation and example syntax
