+++ deepinv/physics/tomography.py
@@ -202,7 +202,7 @@ class Tomography(LinearPhysics):
 
 class TomographyWithAstra(LinearPhysics):
     r"""Computed Tomography operator with `astra-toolbox <https://astra-toolbox.com/>`_ backend.
-    It is more memory efficient than the :class:`deepinv.physics.Tomography` operator and support 3d geometries.
+    It is more memory efficient than the :class:`deepinv.physics.Tomography` operator and support 3D geometries.
     See documentation of :class:`deepinv.physics.functional.XrayTransform` for more
     information on the ``astra`` wrapper.
 
@@ -218,27 +218,27 @@ class TomographyWithAstra(LinearPhysics):
     trajectory. Given different acquisition systems, the lines along which
     the integrals are computed follow different geometries:
 
-    * parallel. (2d and 3d)
-        Per view, all rays intersecting the object are parallel. In 2d, all rays live on the same plane, perpendicular
+    * parallel. (2D and 3D)
+        Per view, all rays intersecting the object are parallel. In 2D, all rays live on the same plane, perpendicular
         to the axis of rotation.
-    * fanbeam. (2d)
+    * fanbeam. (2D)
         Per view, all rays come from a single source and intersect the object at a certain angle. The detector consists of a 1d line of cells. Similar to
-        the 2d "parallel", all rays live on the same plane, perpendicular to the axis of rotation.
-    * conebeam. (3d)
-        Per view, all rays come from a single source. The detector consists of a 2d grid of cells. Apart from the central plane, the set of rays coming onto
+        the 2D "parallel", all rays live on the same plane, perpendicular to the axis of rotation.
+    * conebeam. (3D)
+        Per view, all rays come from a single source. The detector consists of a 2D grid of cells. Apart from the central plane, the set of rays coming onto
         a line of cells live on a tilted plane.
 
     .. note::
 
         The pseudo-inverse is computed using the filtered back-projection
-        algorithm with a Ramp filter, and its equivalent for conebeam 3d, the
+        algorithm with a Ramp filter, and its equivalent for conebeam 3D, the
         Feldkamp-Davis-Kress algorithm. This is not the exact linear pseudo-inverse
         of the Ray Transform, but it is a good approximation which is robust to noise.
 
     .. note::
 
         In the default configuration, reconstruction cells and detector cells are
-        set to have isotropic unit lengths. The geometry is set to 2d parallel
+        set to have isotropic unit lengths. The geometry is set to 2D parallel
         and matches the default configuration of the :class:`deepinv.physics.Tomography` operator with
         ``circle=False``.
 
@@ -251,17 +251,17 @@ class TomographyWithAstra(LinearPhysics):
 
     .. warning::
 
-        The :class:`deepinv.physics.functional.XrayTransform` used in :class:`TomographyWithAstra` sequentially processes batch elements, which can make the 2d parallel beam operator significantly slower than the its native torch counterpart with :class:`deepinv.physics.Tomography` (though still more memory-efficient).
+        The :class:`deepinv.physics.functional.XrayTransform` used in :class:`deepinv.physics.TomographyWithAstra` sequentially processes batch elements, which can make the 2D parallel beam operator significantly slower than its native torch counterpart with :class:`deepinv.physics.Tomography` (though still more memory-efficient).
 
[…]
