+++ examples/external-libraries/_demo_astra_tomography.py
@@ -110,8 +110,8 @@ plot_convergence_metrics = (
 )
 
 # Algorithm parameters
-SCALING = 1 / physics.compute_norm(torch.rand_like(test_imgs)).item()
-stepsize = 0.99 * SCALING
+scaling = 1 / physics.compute_norm(torch.rand_like(test_imgs)).item()
+stepsize = 0.99 * scaling
 lamb = 3.0  # TV regularisation parameter
 params_algo = {"stepsize": stepsize, "lambda": lamb}
 max_iter = 300
