+++ deepinv/physics/functional/astra.py
@@ -28,7 +28,7 @@ class XrayTransform:
 
     :param dict[str, Any] projection_geometry: Dictionnary containing the parameters of the projection geometry. It is passed to the ``astra.create_projector()`` function to instanciate the projector.
     :param dict[str, Any] object_geometry:  Dictionnary containing the parameters of the object geometry. It is passed to the ``astra.create_projector()`` function to instanciate the projector.
-    :param bool is_2d: Specifies if the geometry is flat (2d) or describe a real 3d reconstruction setup.
+    :param bool is_2d: Specifies if the geometry is flat (2D) or describe a real 3D reconstruction setup.
     """
 
     def __init__(
@@ -150,10 +150,10 @@ class XrayTransform:
     ) -> torch.Tensor:
         r"""Forward projection.
 
-        :param torch.Tensor x: Tensor of shape [1,H,W] in 2d, or [D,H,W] in 3d.
+        :param torch.Tensor x: Tensor of shape [1,H,W] in 2D, or [D,H,W] in 3D.
         :param torch.Tensor, None out: To avoid unecessary copies, provide tensor of shape [...,A,N]
         to store output results
-        :return: Sinogram of shape [1,A,N] in 2d or set of sinograms [V,A,N] in 3d.
+        :return: Sinogram of shape [1,A,N] in 2D or set of sinograms [V,A,N] in 3D.
         """
 
         assert (
@@ -191,10 +191,10 @@ class XrayTransform:
             ) -> torch.Tensor:
                 r"""Backprojection.
 
-                :param torch.Tensor x: Tensor of shape [1,A,N] in 2d, or [V,A,N] in 3d.
+                :param torch.Tensor x: Tensor of shape [1,A,N] in 2D, or [V,A,N] in 3D.
                 :param torch.Tensor, None out: To avoid unecessary copies, provide tensor of shape [...,H,W]
                 to store output results
-                :return: Image of shape [1,H,W] in 2d or volume [D,H,W] in 3d.
+                :return: Image of shape [1,H,W] in 2D or volume [D,H,W] in 3D.
                 """
                 assert (
                     x.shape == self.domain_shape
@@ -355,10 +355,10 @@ def create_projection_geometry(
     """Utility function that produces a "projection geometry", a dict of parameters
     used by ``astra`` to parametrize the geometry of the detector and the x-ray source.
 
-    :param str geometry_type: The type of geometry among ``'parallel'``, ``'fanbeam'`` in 2d and ``'parallel'`` and ``'conebeam'`` in 3d.
-    :param int | tuple[int, int]: In 2d the width of a detector cell. In 3d a 2-element tuple specifying the (vertical, horizontal) dimensions of a detector cell. (default: 1.0)
+    :param str geometry_type: The type of geometry among ``'parallel'``, ``'fanbeam'`` in 2D and ``'parallel'`` and ``'conebeam'`` in 3D.
+    :param int | tuple[int, int]: In 2D the width of a detector cell. In 3D a 2-element tuple specifying the (vertical, horizontal) dimensions of a detector cell. (default: 1.0)
     :param torch.Tensor angles: Tensor containing angular positions in radii.
-    :param bool is_2d: Boolean specifying if the parameters define a 2d slice or a 3d volume.
+    :param bool is_2d: Boolean specifying if the parameters define a 2D slice or a 3D volume.
     :param dict[str, str], None geometry_parameters: Contains extra parameters specific to certain geometries. When ``geometry_type='fanbeam'`` or  ``'conebeam'``, the dictionnary should contains the keys
 
         - "source_radius" distance between the x-ray source and the rotation axis, denoted :math:`D_{s0}` (default: 80.)
@@ -391,11 +391,11 @@ def create_projection_geometry(
[…]
