### 2025-06-11 13d8af8
correct examples + change objective function + pb batch dim wavelet prior
