+++ deepinv/models/dsccp.py
@@ -10,7 +10,7 @@ class DScCP(nn.Module):
     DScCP denoiser network.
 
     The network architecture is based on the paper
-    `Unfolded proximal neural networks for robust image Gaussian denoising <https://arxiv.org/pdf/2308.03139>`_,
+    `Unfolded proximal neural networks for robust image Gaussian denoising <https://arxiv.org/abs/2308.03139>`_,
     and has an unrolled architecture based on the fast Chambolle-Pock algorithm using strong convexity.
     DScCP stands for Deep Strongly Convex Chambolle Pock.
 
@@ -18,9 +18,9 @@ class DScCP(nn.Module):
     They can be downloaded via setting ``pretrained='download'``.
 
     :param int K: depth i.e. number of convolutional layers.
-    :param int F:  number of channels per convolutional layer.
+    :param int F: number of channels per convolutional layer.
     :param str pretrained: 'download' to download pretrained weights, or path to local weights file.
-    :param str device: 'cuda' or 'cpu'.
+    :param torch.device, str device: 'cuda' or 'cpu'.
     """
 
     def __init__(self, K=20, F=64, pretrained="download", device=None):
