### 2025-06-12 be0197c
Fix formatting of comments in demo_r2r_denoising.py for clarity
