+++ deepinv/tests/test_optim.py
@@ -504,80 +504,94 @@ def get_prior(prior_name, device="cpu"):
     return prior
 
 
+LIST_PRIORS = [
+    "L1Prior",
+    "L12Prior",
+    "Tikhonov",
+    "TVPrior",
+    "WaveletPrior",
+    "WaveletDictPrior",
+]
+
+
+@pytest.mark.parametrize("prior_name", LIST_PRIORS)
 @pytest.mark.parametrize("pnp_algo", ["PGD", "HQS", "DRS", "ADMM", "CP", "FISTA"])
-def test_priors_algo(pnp_algo, imsize, dummy_dataset, device):
-    for prior_name in [
-        "L1Prior",
-        "L12Prior",
-        "Tikhonov",
-        "TVPrior",
-        "WaveletPrior",
-        "WaveletDictPrior",
-    ]:
-        # 1. Generate a dummy dataset
-        dataloader = DataLoader(
-            dummy_dataset, batch_size=1, shuffle=False, num_workers=0
-        )
-        test_sample = next(iter(dataloader)).to(device)
+def test_priors_algo(pnp_algo, imsize, dummy_dataset, device, prior_name):
+    # 1. Generate a dummy dataset
+    dataloader = DataLoader(dummy_dataset, batch_size=1, shuffle=False, num_workers=0)
+    test_sample = next(iter(dataloader)).to(device)
 
-        # 2. Set a physical experiment (here, deblurring)
-        physics = dinv.physics.Blur(
-            dinv.physics.blur.gaussian_blur(sigma=(2, 0.1), angle=45.0),
-            padding="circular",
-            device=device,
-        )
-        y = physics(test_sample)
-        max_iter = 1000
-        # Note: results are better for sigma_denoiser=0.001, but it takes longer to run.
-        # sigma_denoiser = torch.tensor([[0.1]])
-        sigma_denoiser = torch.tensor([[1.0]], device=device)
-        stepsize = 1.0
-        lamb = 1.0
+    # 2. Set a physical experiment (here, deblurring)
+    physics = dinv.physics.Blur(
+        dinv.physics.blur.gaussian_blur(sigma=(2, 0.1), angle=45.0),
+        padding="circular",
+        device=device,
+    )
+    y = physics(test_sample)
+    max_iter = 1000
+    # Note: results are better for sigma_denoiser=0.001, but it takes longer to run.
+    # sigma_denoiser = torch.tensor([[0.1]])
+    sigma_denoiser = torch.tensor([[1.0]], device=device)
+    stepsize = 1.0
+    lamb = 1.0
 
-        data_fidelity = L2()
+    data_fidelity = L2()
 
-        # here the prior model is common for all iterations
-        prior = get_prior(prior_name, device=device)
+    # here the prior model is common for all iterations
+    prior = get_prior(prior_name, device=device)
 
-        stepsize_dual = 1.0 if pnp_algo == "CP" else None
-        params_algo = {
-            "stepsize": stepsize,
-            "g_param": sigma_denoiser,
-            "lambda": lamb,
-            "stepsize_dual": stepsize_dual,
-        }
+    stepsize_dual = 1.0 if pnp_algo == "CP" else None
+    params_algo = {
+        "stepsize": stepsize,
+        "g_param": sigma_denoiser,
+        "lambda": lamb,
+        "stepsize_dual": stepsize_dual,
+    }
 
-        custom_init = custom_init_CP if pnp_algo == "CP" else None
+    custom_init = custom_init_CP if pnp_algo == "CP" else None
 
-        opt_algo = optim_builder(
-            pnp_algo,
-            prior=prior,
-            data_fidelity=data_fidelity,
-            max_iter=max_iter,
[…]
