+++ deepinv/sampling/diffusion_sde.py
@@ -333,7 +333,7 @@ class VariancePreservingDiffusion(DiffusionSDE):
 
     This class is the reverse-time SDE of the VP-SDE, serving as the generation process.
 
-    :param deepinv.models.Denoiser denoiser: a denoiser used to provide an approximation of the score at time :math:`t` : :math:`\nabla \log p_t`.
+    :param deepinv.models.Denoiser denoiser: a denoiser used to provide an approximation of the score at time :math:`t`: :math:`\nabla \log p_t`.
     :param float beta_min: the minimum noise level.
     :param float beta_max: the maximum noise level.
     :param float alpha: the weighting factor of the diffusion term.
