+++ deepinv/optim/prior.py
@@ -361,7 +361,7 @@ class WaveletPrior(Prior):
                 wvdim=self.wvdim,
             )
 
-    def fn(self, x, *args, reduce=True, **kwargs):
+    def fn(self, x, *args, reduce=False, **kwargs):
         r"""
         Computes the regularizer
 
@@ -388,10 +388,8 @@ class WaveletPrior(Prior):
         :return: (:class:`torch.Tensor`) prior :math:`g(x)`.
         """
         list_dec = self.psi(x)
-        print("List of detail coefficients:", len(list_dec))
-        print('Shape of detail coefficients:', [dec.shape for dec in list_dec])
-        list_norm = torch.hstack([torch.norm(dec, p=self.p) for dec in list_dec])
-        print("List norm length:", list_norm.shape)
+        list_norm_dec = torch.stack([torch.norm(dec, p=self.p, dim=-1) for dec in list_dec], dim=-1) # shape: (batch_size, n_dec)
+        list_norm = torch.sum(list_norm_dec, dim=-1)  # shape: (batch_size,)
         if reduce:
             return torch.sum(list_norm)
         else:
