### 2025-06-12 23ed1d8
typo in vp sde (#527)
* typo in vp sde

* Update deepinv/sampling/diffusion_sde.py

Co-authored-by: Julián Tachella <38355592+tachella@users.noreply.github.com>

---------

Co-authored-by: mh-nguyen712 <minh-hai.nguyen@irit.fr>
Co-authored-by: Julián Tachella <38355592+tachella@users.noreply.github.com>
