+++ deepinv/models/guided_diffusion.py
@@ -64,7 +64,7 @@ class ADMUNet(Denoiser):
         attn_resolutions=[32, 16, 8],  # List of resolutions with self-attention.
         dropout=0.10,  # List of resolutions with self-attention.
         label_dropout=0,  # Dropout probability of class labels for classifier-free guidance.
-        pretrained: str = None,
+        pretrained: str = "download",
         pixel_std: float = 0.75,
         device=None,
         *args,
