+++ deepinv/optim/optimizers.py
@@ -522,8 +522,9 @@ def create_iterator(
     if F_fn is None and explicit_prior:
 
         def F_fn(x, data_fidelity, prior, cur_params, y, physics):
-            prior_value = prior(x, cur_params["g_param"], reduce=False)
-            return data_fidelity(x, y, physics) + cur_params["lambda"] * prior_value
+            return data_fidelity(x, y, physics) + cur_params["lambda"] * prior(
+                x, cur_params["g_param"], reduce=False
+            )
 
         has_cost = True  # boolean to indicate if there is a cost function to evaluate along the iterations
     else:
