### 2025-06-13 add7278
pb diffusion NCSPP
