### 2025-06-13 b9c192a
Move example to advanced section. Remove external-libraries section.
