+++ examples/external-libraries/README.rst
@@ -1,2 +1,0 @@
-External libraries
------------------------------------
