### 2025-06-13 4d3374c
add FID computation
