+++ deepinv/loss/metric/perceptual.py
@@ -1,4 +1,4 @@
-from deepinv.loss.metric.metric import import_pyiqa, Metric
+from deepinv.loss.metric.metric import import_pyiqa, import_cleanfid, Metric
 
 
 class LPIPS(Metric):
@@ -42,6 +42,23 @@ class LPIPS(Metric):
 
     def metric(self, x_net, x, *args, **kwargs):
         return self.lpips(x_net, x).squeeze(-1)
+    
+
+class FID(Metric):
+    r"""
+    Frechet Inception Distance (FID) metric.
+
+    Computes the Frechet Inception Distance between two sets of images, which corresponds to the Wasserstein distance between empirical Gaussians in the feature space of a pre-trained Inception network.
+    Uses implementation from `pyiqa <https://pypi.org/project/pyiqa/>`, based on the implementation from `clean-fid <https://github.com/GaParmar/clean-fid?tab=readme-ov-file#computing-fid>`_.
+    """
+    def __init__(self, device="cpu", **kwargs):
+        super().__init__(**kwargs)
+        pyiqa = import_pyiqa()
+        self.device = device
+        self.fid_metric = pyiqa.create_metric("fid").to(device)
+        
+    def metric(self, out_dir, ref_dir = None, dataset_name = None, dataset_res=1024, dataset_split='train', num_workers = 4, batch_size = 8, **kwargs):
+        return self.fid_metric(fdir1 = out_dir, fdir2 = ref_dir, dataset_name=dataset_name, dataset_res=dataset_res, dataset_split=dataset_split, num_workers = num_workers, batch_size = batch_size, device=self.device, **kwargs)
 
 
 class NIQE(Metric):
