### 2025-06-13 bafdc3d
Rename examples/advanced -> examples/external-libraries
