+++ docs/source/conf.py
@@ -190,7 +190,7 @@ sphinx_gallery_conf = {
             "../../examples/patch-priors",
             "../../examples/self-supervised-learning",
             "../../examples/adversarial-learning",
-            "../../examples/advanced",
+            "../../examples/external-libraries",
         ]
     ),
 }
