### 2025-06-13 5ff3e3a
warnings compose physics
