### 2025-06-13 3a9aa39
warning and docstring
