### 2025-06-13 8d4ef2c
Unify argument name by adding aliases (#494)
* change name, add test decorator and deprecated args

* resolve tensor_size inpainting

* resolve img_shape

* resolve in_shape

* add missing files

* randomphase

* changelog

* change inpainting generators

* gan

* gan

* loss

* DIP

* tested

* merged main

* pytest

* docstring

* merged

* alias for function

* alias for function
