### 2025-06-13 393865b
Apply suggestions from code review
Co-authored-by: Andrewwango <andrew.wang27@gmail.com>
