+++ _downloads/0ed8af0d0fd95c7e0dbbaecec30f661a/demo_multioperator_imaging.py
@@ -1,0 +1,207 @@
+r"""
+Self-supervised learning from incomplete measurements of multiple operators.
+====================================================================================================
+
+This example shows you how to train a reconstruction network for an inpainting
+inverse problem on a fully self-supervised way, i.e., using measurement data only.
+
+The dataset consists of pairs :math:`(y_i,A_{g_i})` where :math:`y_i` are the measurements and :math:`A_{g_i}` is a
+binary sampling operator out of :math:`G` (i.e., :math:`g_i\in \{1,\dots,G\}`).
+
+This self-supervised learning approach is presented in `"Unsupervised Learning From Incomplete Measurements for
+Inverse Problems" <https://openreview.net/pdf?id=aV9WSvM6N3>`_, and minimizes the loss function:
+
+.. math::
+
+    \mathcal{L}(\theta) = \sum_{i=1}^{N} \left\|A_{g_i} \hat{x}_{i,\theta} - y_i \right\|_2^2 + \sum_{s=1}^{G}
+    \left\|\hat{x}_{i,\theta} - R_{\theta}(A_s\hat{x}_{i,\theta},A_s) \right\|_2^2
+
+where :math:`R_{\theta}` is a reconstruction network with parameters :math:`\theta`, :math:`y_i` are the measurements,
+:math:`A_s` is a binary sampling operator, and :math:`\hat{x}_{i,\theta} = R_{\theta}(y_i,A_{g_i})`.
+
+"""
+
+from pathlib import Path
+
+import torch
+from torch.utils.data import DataLoader
+from torchvision import datasets, transforms
+
+import deepinv as dinv
+from deepinv.utils.demo import get_data_home
+from deepinv.models.utils import get_weights_url
+
+# %%
+# Setup paths for data loading and results.
+# ---------------------------------------------------------------
+#
+
+BASE_DIR = Path(".")
+DATA_DIR = BASE_DIR / "measurements"
+CKPT_DIR = BASE_DIR / "ckpts"
+ORIGINAL_DATA_DIR = get_data_home()
+
+# Set the global random seed from pytorch to ensure reproducibility of the example.
+torch.manual_seed(0)
+
+device = dinv.utils.get_freer_gpu() if torch.cuda.is_available() else "cpu"
+
+# %%
+# Load base image datasets and degradation operators.
+# ----------------------------------------------------------------------------------
+# In this example, we use the MNIST dataset for training and testing.
+#
+
+transform = transforms.Compose([transforms.ToTensor()])
+
+train_base_dataset = datasets.MNIST(
+    root=ORIGINAL_DATA_DIR, train=True, transform=transform, download=True
+)
+test_base_dataset = datasets.MNIST(
+    root=ORIGINAL_DATA_DIR, train=False, transform=transform, download=True
+)
+
+# %%
+# Generate a dataset of subsampled images and load it.
+# ----------------------------------------------------------------------------------
+# We generate 10 different inpainting operators, each one with a different random mask.
+# If the :func:`deepinv.datasets.generate_dataset` receives a list of physics operators, it
+# generates a dataset for each operator and returns a list of paths to the generated datasets.
+#
+# .. note::
+#
+#   We only use 10 training images per operator to reduce the computational time of this example. You can use the whole
+#   dataset by setting ``n_images_max = None``.
+
+number_of_operators = 10
+
+# defined physics
+physics = [
+    dinv.physics.Inpainting(mask=0.5, img_size=(1, 28, 28), device=device)
+    for _ in range(number_of_operators)
+]
+
+# Use parallel dataloader if using a GPU to reduce training time,
[…]
