### 2025-06-13 a9a7a7e
add Physics.clone
