### 2025-06-13 66dd07c
get rid of dummy variable
