+++ deepinv/physics/tomography.py
@@ -163,7 +163,7 @@ class Tomography(LinearPhysics):
                 output = output * output.shape[-1] ** 2
         else:
             y = self.iradon.filter(y)
-            output = ApplyRadon.apply(x, self.radon, self.iradon, True)
+            output = ApplyRadon.apply(y, self.radon, self.iradon, True)
             if self.normalize:
                 output = output * output.shape[-1]
         return output
