### 2025-06-13 a7bf6dd
add additional tests
