### 2025-06-13 c2709da
improve docstrings
