### 2025-06-13 8bb4cc2
replace deepcopy by Physics.clone
