### 2025-06-14 c8057f3
docs + FBP boundary handling
