+++ deepinv/physics/tomography.py
@@ -37,6 +37,11 @@ class Tomography(LinearPhysics):
     :param bool circle: If ``True`` both forward and backward projection will be restricted to pixels inside a circle
         inscribed in the square image.
     :param bool parallel_computation: if True, all projections are performed in parallel. Requires more memory but is faster on GPUs.
+    :param bool adjoint_via_backprop: if True, the adjoint will be computed via ``deepinv.physics.adjoint_function``. Otherwise the inverse Radon transform is used.
+        The inverse Radon transform is computationally cheaper (particularly in memory), but has a small adjoint mismatch.
+        The backprop adjoint is the exact adjoint, but might break random seeds since it backpropagates through ``torch.nn.functional.grid_sample``, see the note `here <https://docs.pytorch.org/docs/stable/generated/torch.nn.functional.grid_sample.html>`_,
+    :param bool fbp_interpolate_boundary: the FBP (``A_dagger``) usually contains streaking artifacts on the boundary due to padding. For ``fbp_interpolate_boundary=True``
+        these artifacts are corrected by cutting off the outer two pixels of the FBP and recovering them by interpolating the remaining image.
     :param bool normalize: If ``True``, the outputs are normlized by the image size (i.e. it is assumed that the image lives on [0,1]^2 for the computation of the line integrals).
         In this case the operator norm is approximately given by :math:`\|A\|_2^2  \approx \frac{\pi}{2\,\text{angles}}`,
         If ``False``, then it is assumed that the image lives on [0,im_width]^2 for the computation of the line integrals
@@ -70,10 +75,10 @@ class Tomography(LinearPhysics):
         >>> angles = torch.linspace(0, 45, steps=3)
         >>> physics = Tomography(angles=angles, img_width=4, circle=True)
         >>> physics(x)
-        tensor([[[[ 0.1650,  1.2640,  1.6995],
-                  [-0.4860,  0.2674,  0.9971],
-                  [ 0.9002, -0.3856, -0.9360],
-                  [-2.4882, -2.1068, -2.5720]]]])
+        tensor([[[[ 0.0000, -0.1791, -0.1719],
+                  [-0.5713, -0.4521, -0.5177],
+                  [ 0.0340,  0.1448,  0.2334],
+                  [ 0.0000, -0.0448, -0.0430]]]])
 
         Tomography operator with 3 uniformly sampled angles in [0, 360] for 3x3 image:
 
@@ -82,10 +87,10 @@ class Tomography(LinearPhysics):
         >>> x = torch.randn(1, 1, 4, 4)  # Define random 4x4 image
         >>> physics = Tomography(angles=3, img_width=4, circle=True)
         >>> physics(x)
-        tensor([[[[ 0.1650,  1.9493,  1.9897],
-                  [-0.4860,  0.7137, -1.6536],
-                  [ 0.9002, -0.8457, -0.1666],
-                  [-2.4882, -2.7340, -0.9793]]]])
+        tensor([[[[ 0.0000, -0.1806,  0.0500],
+                  [-0.5713, -0.6076, -0.6815],
+                  [ 0.0340,  0.3175,  0.0167],
+                  [ 0.0000, -0.0452,  0.0989]]]])
 
 
     """
@@ -97,6 +102,7 @@ class Tomography(LinearPhysics):
         circle=False,
         parallel_computation=True,
         adjoint_via_backprop=False,
+        fbp_interpolate_boundary=False,
         normalize=False,
         fan_beam=False,
         fan_parameters=None,
@@ -116,6 +122,7 @@ class Tomography(LinearPhysics):
 
         self.fan_beam = fan_beam
         self.adjoint_via_backprop = adjoint_via_backprop
[…]
