### 2025-06-14 6e23332
fix dynamic range PSNR for batches
