### 2025-06-14 8a4740a
fix amin/min and amax/max
