### 2025-06-14 e44021a
batching wavelet prior
