### 2025-06-14 9d8dce6
batching wavelet prior, optimizers flatten reg param
