### 2025-06-14 20bd600
batching sigma of dsccp model
