### 2025-06-14 b6f6b0a
Update docs/source/contributing.rst
Co-authored-by: Andrewwango <andrew.wang27@gmail.com>
