+++ docs/source/contributing.rst
@@ -59,7 +59,7 @@ You can also send an email to any of the main developers with your questions or 
 Contributing new datasets
 --------------------------
 
-Contributing a new dataset can be a little trickier than other contributions, here are a few guidelines to help you with that. The DeepInverse code base is regularly tested on automatic continuous integration (CI) servers in order to ensure that the code works the way it is supposed to. Unfortunately, the CI servers have limited resources and they can generally not host the datasets. This makes testing the implementation of the datasets a bit cumbersome as only the logic can be tested and not the consistency with the actual data.
+In order to contribute a new dataset, you must provide tests alongisde it to check that it functions as expected. The DeepInverse code base is regularly tested on automatic continuous integration (CI) servers in order to ensure that the code works the way it is supposed to. Unfortunately, the CI servers have limited resources and they can generally not host the datasets.
 
 This does not mean that you should not contribute datasets, they are really valuable and make the library more complete! It does not mean that you can simply skip testing the implementation either. Our current guidelines is to first write the tests and the implementation and to make sure that the tests pass locally, on the real data. As it is not possible to do so on the CI servers, the next step is to write `mocking code <https://en.wikipedia.org/wiki/Mock_object>`_, code that intercepts calls to input/output (IO) related functions, e.g. `os.listdir`, and make them return a hard coded value, thereby making execution go as if the data was there. For more details, see `this pull request <https://github.com/deepinv/deepinv/pull/490>`_.
 
