### 2025-06-14 4394775
Update docs/source/contributing.rst
Co-authored-by: Andrewwango <andrew.wang27@gmail.com>
