### 2025-06-14 0b3d83e
catch errors due to unsupported pydicom version
