+++ .github/workflows/test.yml
@@ -35,7 +35,7 @@ jobs:
             extra: dataset,denoisers,test
             name: "windows all dependencies"
     env:
-      DEEPINV_MOCK_TESTS: "True"
+      DEEPINV_MOCK_TESTS: "True" #
     steps:
       - uses: actions/checkout@v3
       - uses: actions/setup-python@v5
