### 2025-06-15 54931df
Make physics classes differentiable (#462)
* start branch

* start caching update_parameters

* change all update_parameters methods

* fixed test device, some numpy to torch and update_parameters

* rm norm cs singlepixel gpu

* physics test passed

* fixed device consistency and test

* checking differentiability, some physics need to be improved

* minor change on physics for autodiff

* revert test physics to origin

* test_physics passed

* coorect tests

* tests passed

* test example

* add example

* done example

* changelog and small device fix in doctest

* suppress download output 
[…]
