+++ _downloads/2face6fea1a6f2a2612271b6280159b3/demo_physics_tour.py
@@ -1,0 +1,258 @@
+r"""
+A tour of forward sensing operators
+===================================================
+
+This example provides a tour of some of the forward operators implemented in DeepInverse.
+We restrict ourselves to operators where the signal is a 2D image. The full list of operators can be found in
+`here <models>`_.
+
+"""
+
+import torch
+
+import deepinv as dinv
+from deepinv.utils.plotting import plot
+from deepinv.utils.demo import load_example
+
+
+# %%
+# Load image from the internet
+# ----------------------------
+#
+# This example uses an image of the CBSD68 dataset.
+
+device = dinv.utils.get_freer_gpu() if torch.cuda.is_available() else "cpu"
+
+x = load_example("CBSD_0010.png", grayscale=False).to(device)
+
+x = torch.tensor(x, device=device, dtype=torch.float)
+x = torch.nn.functional.interpolate(x, size=(64, 64))
+img_size = x.shape[1:]
+# Set the global random seed from pytorch to ensure reproducibility of the example.
+torch.manual_seed(0)
+
+# %%
+# Denoising
+# ---------------------------------------
+#
+# The denoising class :class:`deepinv.physics.Denoising` is associated with an identity operator.
+# In this example we choose a Poisson noise.
+
+physics = dinv.physics.Denoising(dinv.physics.PoissonNoise(0.1))
+
+y = physics(x)
+
+# plot results
+plot([x, y], titles=["signal", "measurement"])
+
+# %%
+# Inpainting
+# ---------------------------------------
+#
+# The inpainting class :class:`deepinv.physics.Inpainting` is associated with a mask operator.
+# The mask is generated at random (unless an explicit mask is provided as input).
+# We also consider Gaussian noise in this example.
+
+sigma = 0.1  # noise level
+physics = dinv.physics.Inpainting(
+    mask=0.5,
+    img_size=x.shape[1:],
+    noise_model=dinv.physics.GaussianNoise(sigma=sigma),
+    device=device,
+)
+
+y = physics(x)
+
+# plot results
+plot([x, y], titles=["signal", "measurement"])
+
+
+# %%
+# Demosaicing
+# ---------------------------------------
+#
+# The demosaicing class :class:`deepinv.physics.Demosaicing` is associated with a Bayer pattern,
+# which is a color filter array used in digital cameras (see `Wikipedia <https://en.wikipedia.org/wiki/Bayer_filter>`_).
+
+physics = dinv.physics.Demosaicing(img_size=(64, 64), device=device)
+
+y = physics(x)
+
+# plot results
+plot([x, y], titles=["signal", "measurement"])
+
+# %%
+# Compressed Sensing
+# ---------------------------------------
+#
+# The compressed sensing class :class:`deepinv.physics.CompressedSensing` is associated with a random Gaussian matrix.
+# Here we take 2048 measurements of an image of size 64x64, which corresponds to a compression ratio of 2.
+
+physics = dinv.physics.CompressedSensing(
+    m=2048,
+    fast=False,
+    channelwise=True,
+    img_size=img_size,
+    compute_inverse=True,
+    device=device,
+)
+
+y = physics(x)
+
+# plot results
+plot([x, physics.A_dagger(y)], titles=["signal", "linear inverse"])
+
+# %%
+# Computed Tomography
+# ---------------------------------------
+#
+# The class :class:`deepinv.physics.Tomography` is associated with the sparse Radon transform.
+# Here we take 20 views of an image of size 64x64, and consider mixed Poisson-Gaussian noise.
+
[…]
