+++ _downloads/4af5e102f52a3fd109e8d7d07dca050c/demo_custom_kernel.py
@@ -1,0 +1,235 @@
+r"""
+Building your custom MCMC sampling algorithm.
+=============================================
+
+This code shows how to build your custom sampling kernel. Here we build a preconditioned Unadjusted Langevin
+Algorithm (PreconULA) that takes advantage of the singular value decomposition of the forward operator
+to accelerate the sampling.
+
+
+"""
+
+import deepinv as dinv
+from deepinv.utils.plotting import plot
+import torch
+from deepinv.sampling import ULA
+import numpy as np
+from deepinv.utils.demo import load_url_image
+
+# %%
+# Load image from the internet
+# ----------------------------
+#
+# This example uses an image of Lionel Messi from Wikipedia.
+
+device = dinv.utils.get_freer_gpu() if torch.cuda.is_available() else "cpu"
+
+url = (
+    "https://upload.wikimedia.org/wikipedia/commons/b/b4/"
+    "Lionel-Messi-Argentina-2022-FIFA-World-Cup_%28cropped%29.jpg"
+)
+x = load_url_image(url=url, img_size=32).to(device)
+
+# %%
+# Define forward operator and noise model
+# ---------------------------------------
+#
+# We use a 5x5 box blur as the forward operator and Gaussian noise as the noise model.
+
+sigma = 0.001  # noise level
+physics = dinv.physics.BlurFFT(
+    img_size=(3, 32, 32),
+    filter=torch.ones((1, 1, 5, 5), device=device) / 25,
+    device=device,
+    noise_model=dinv.physics.GaussianNoise(sigma=sigma),
+)
+
+
+# %%
+# Generate the measurement
+# ------------------------
+# Apply the forward model to generate the noisy measurement.
+
+y = physics(x)
+
+# %%
+# Define the sampling iteration
+# -----------------------------
+#
+# In order to define a custom sampling kernel (possibly a Markov kernel which depends on the previous sample),
+# we only need to define the iterator which takes the current sample and returns the next sample.
+#
+# Here we define a preconditioned ULA iterator (for a Gaussian likelihood),
+# which takes into account the singular value decomposition
+# of the forward operator, :math:`A=USV^{\top}`, in order to accelerate the sampling.
+#
+# We modify the standard ULA iteration (see :class:`deepinv.sampling.ULA`) defined as
+#
+# .. math::
+#
+#        x_{k+1} = x_{k} + \frac{\eta}{\sigma^2} A^{\top}(y-Ax_{k}) +
+#        \eta \alpha \nabla \log p(x_{k}) + \sqrt{2\eta}z_{k+1}
+#
+#
+# by using a matrix-valued step size :math:`\eta = \eta_0 VRV^{\top}` where
+# :math:`R` is a diagonal matrix with entries :math:`R_{i,i} = \frac{1}{S_{i,i}^2 + \epsilon}`.
+# The parameter :math:`\epsilon` is used to avoid numerical issues when :math:`S_{i,i}^2` is close to zero.
+# After some algebra, we obtain the following iteration
+#
+# .. math::
+#
+#        x_{k+1} = x_{k} + \frac{\eta_0}{\sigma^2}  V R S (U^{\top}y - S V^{\top}x_{k})
+#        +\eta_0  \alpha V R V^{\top}\nabla \log p(x_{k}) + \sqrt{2\eta_0}V\sqrt{R}z_{k+1}
+#
+# We exploit the methods of :class:`deepinv.physics.DecomposablePhysics` to compute the matrix-vector products
+# with :math:`V` and :math:`V^{\top}` efficiently. Note that computing the matrix-vector product with :math:`R` and
+# :math:`S` is trivial since they are diagonal matrices.
+
+
+class PULAIterator(torch.nn.Module):
[…]
