+++ _sources/api/stubs/deepinv.optim.BregmanL2.rst.txt
@@ -1,0 +1,15 @@
+﻿BregmanL2
+=========
+
+.. currentmodule:: deepinv.optim
+
+.. autoclass:: BregmanL2
+   :members:
+   :show-inheritance:
+   :no-undoc-members:
+   :special-members: __mul__, __add__, __div__, __neg__, __sub__, __truediv__
+
+.. _sphx_glr_backref_deepinv.optim.BregmanL2:
+
+.. minigallery:: deepinv.optim.BregmanL2
+    :add-heading: Examples using ``BregmanL2``:

[…]
