### 2025-06-15 ccc61b0
Apply suggested docstring changes
Co-authored-by: Andrewwango <andrew.wang27@gmail.com>
