+++ deepinv/sampling/sampling_iterators/ula.py
@@ -110,7 +110,7 @@ class ULAIterator(SamplingIterator):
         :param ScorePrior cur_prior: Score-based prior model for :math:`\nabla \log p(x)`
         :param int iteration: Current iteration number in the sampling process (zero-indexed)
 
-        :return: Dictionary `{"est": x}` containing the next state :math:`x_{t+1}` in the Markov chain.
+        :return: Dictionary `{"x": x}` containing the next state :math:`x_{t+1}` in the Markov chain.
         :rtype: Dict
         """
         x = X["x"]
