### 2025-06-15 beb337d
Move g_statistics to forward in ULA/SKRock
