### 2025-06-15 93a2c1e
Initialize latent variables
