### 2025-06-15 ac06731
refer to example in userguide rather than custom mcmc algorithm for legacy classes
