### 2025-06-16 e3f2aa4
tests and changelog
