### 2025-06-16 b0fd882
Upgrade deprecated types (#501)
* ugprade deprecated types

* update the changelog

* upgrade docstrings and remove unused imports

* add ruff to the ci

* trigger CI

* trigger CI

---------

Co-authored-by: Andrewwango <andrew.wang27@gmail.com>
