+++ deepinv/tests/test_physics.py
@@ -2,7 +2,7 @@ import os
 import shutil
 import copy
 from math import sqrt
-from typing import Optional, List
+from typing import Optional
 import pytest
 import torch
 import numpy as np
@@ -1186,7 +1186,7 @@ def test_mri_fft():
 
         return torch.cat((right, left), dim=dim)
 
-    def roll(x: torch.Tensor, shift: List[int], dim: List[int]) -> torch.Tensor:
+    def roll(x: torch.Tensor, shift: list[int], dim: list[int]) -> torch.Tensor:
         if len(shift) != len(dim):
             raise ValueError("len(shift) must match len(dim)")
 
@@ -1195,7 +1195,7 @@ def test_mri_fft():
 
         return x
 
-    def fftshift(x: torch.Tensor, dim: Optional[List[int]] = None) -> torch.Tensor:
+    def fftshift(x: torch.Tensor, dim: Optional[list[int]] = None) -> torch.Tensor:
         if dim is None:
             # this weird code is necessary for toch.jit.script typing
             dim = [0] * (x.dim())
@@ -1209,7 +1209,7 @@ def test_mri_fft():
 
         return roll(x, shift, dim)
 
-    def ifftshift(x: torch.Tensor, dim: Optional[List[int]] = None) -> torch.Tensor:
+    def ifftshift(x: torch.Tensor, dim: Optional[list[int]] = None) -> torch.Tensor:
         if dim is None:
             # this weird code is necessary for toch.jit.script typing
             dim = [0] * (x.dim())
