### 2025-06-16 fb394fc
Increasing code coverage by mocking datasets (#490)
* poc

* black

* add a non-empty case

* complete coverage for lidc_idri

* mock set14 tests

* mock the tests for Kohler

* fix coverage

* set minimum version for pydicom

* revert pydicom version

* fix lidc idri for python 3.9

* black

* complete the coverage for Kohler

* complete the coverage for Kohler

* complete coverage of Set14HR

* mock FMD tests

* mock tests for LsdirHR

* black

* clean up mess made in test_poc

* test Flickr2kHR

* factor code into get_dummy_pil_png_image

* update changelog

* make mocking controlled by an 
[…]
