### 2025-06-16 d09b859
parametrize test_tomography
