### 2025-06-16 e90ec1a
fix test_inpainting_generator_random_ratio
