### 2025-06-16 d2c4716
black and set fbp_interpolate_boundary to False if circle is True
