### 2025-06-16 78929e6
make rng a test parameter
