### 2025-06-16 b49ff7f
add a cloned physics to the test list
