### 2025-06-16 4b7610c
add tests for requires_grad
