### 2025-06-17 5f0d822
Fix test_inpainting_generator_random_ratio (#542)
* fix test_inpainting_generator_random_ratio

* make rng a test parameter
