+++ docs/source/index.rst
@@ -41,7 +41,7 @@ Github repository: `<https://github.com/deepinv/deepinv>`_.
 
 **Citation**
 
-If you use DeepInverse in your research, please cite the following paper (available in `arXiv <https://arxiv.org/abs/2505.20160>`_):
+If you use DeepInverse in your research, please cite the following paper (available on `arXiv <https://arxiv.org/abs/2505.20160>`_):
 
 .. code-block:: bash
 
