+++ CITATION.cff
@@ -52,8 +52,6 @@ authors:
     given-names: "Victor"
   - family-names: "Schramm"
     given-names: "Georg"
-  - family-names: "Tang"
-    given-names: "Chao"
   - family-names: "Vo"
     given-names: "Romain"
   - family-names: "Weiss"
