### 2025-06-17 eadcfc7
Update README.rst
