### 2025-06-17 0c6f5da
Tomography operator with astra-toolbox (#474)
* Implement astra-wrapper. Started doc.

* Complete docs, fix bugs

* Add astra option to demos

* Add tests, update docs

* black.

* cancel lpd change

* remove unused import

* Fix type hints and docs

* Add external-libraries example. Tighter error tolerances in test_physics. Add pytest parameters to test_tomography_with_astra.

* black.

* Add astra-toolbox to optional dependencies. Small fix on doc code snippet.

* Specify astra-toolbox version in dependencies

* loosened range of package version for astra-toolbox

* Restrict dependency insta
[…]
