### 2025-06-17 8fe18de
fix torch.pi regression
