### 2025-06-17 a134617
update docstrings
