### 2025-06-17 c9b004f
move blind from advanced to basics and remove advanced
