+++ examples/advanced/README.rst
@@ -1,2 +1,0 @@
-Advanced
--------------------------------
