### 2025-06-17 cbec6c6
fix docstring formatting
