### 2025-06-17 3a9291e
remove advanced from conf
