### 2025-06-17 e0049da
Parametrize test_tomography (#541)
* parametrize test_tomography

* black

* fix torch.pi regression

---------

Co-authored-by: Andrewwango <andrew.wang27@gmail.com>
