### 2025-06-18 8b3984b
test attributes other than params and bufffers
