### 2025-06-18 ca79387
Apply suggestions from code review
Co-authored-by: Julián Tachella <38355592+tachella@users.noreply.github.com>
