### 2025-06-18 22449b6
add adjointness test for exact adjoint
