### 2025-06-18 12699f2
dscnn as denoiser
