### 2025-06-19 a982323
Make minor documentation improvements (#549)
* create pr

* fix some formatting issues

* correct typo

* Apply suggestions from code review

Co-authored-by: Julián Tachella <38355592+tachella@users.noreply.github.com>

* fix the link

---------

Co-authored-by: Julián Tachella <38355592+tachella@users.noreply.github.com>
