+++ _downloads/54991abcb2105cbefad71602185c6e3b/demo_unfolded_constrained_LISTA.py
@@ -1,0 +1,316 @@
+r"""
+Unfolded Chambolle-Pock for constrained image inpainting
+====================================================================================================
+
+Image inpainting consists in solving :math:`y = Ax` where :math:`A` is a mask operator.
+This problem can be reformulated as the following minimization problem:
+
+.. math::
+
+    \begin{equation*}
+    \underset{x}{\operatorname{min}} \,\, \iota_{\mathcal{B}_2(y, r)}(Ax) + \regname(x)
+    \end{equation*}
+
+
+where :math:`\iota_{\mathcal{B}_2(y, r)}` is the indicator function of the ball of radius :math:`r` centered at
+:math:`y` for the :math:`\ell_2` norm, and :math:`\regname` is a regularisation. Recall that the indicator function of
+a convex set :math:`\mathcal{C}` is defined as :math:`\iota_{\mathcal{C}}(x) = 0` if :math:`x \in \mathcal{C}` and
+:math:`\iota_{\mathcal{C}}(x) = +\infty` otherwise.
+
+In this example, we unfold the Chambolle-Pock algorithm to solve this problem, and learn the thresholding parameters of
+a wavelet denoiser in a LISTA fashion.
+
+"""
+
+from pathlib import Path
+import torch
+from torch.utils.data import DataLoader
+from torchvision import transforms
+import deepinv as dinv
+from deepinv.utils.demo import load_dataset
+from deepinv.optim.data_fidelity import IndicatorL2
+from deepinv.optim.prior import PnP
+from deepinv.unfolded import unfolded_builder
+
+# %%
+# Setup paths for data loading and results.
+# --------------------------------------------
+#
+
+BASE_DIR = Path(".")
+DATA_DIR = BASE_DIR / "measurements"
+CKPT_DIR = BASE_DIR / "ckpts"
+
+# Set the global random seed from pytorch to ensure reproducibility of the example.
+torch.manual_seed(0)
+
+device = dinv.utils.get_freer_gpu() if torch.cuda.is_available() else "cpu"
+
+# %%
+# Load base image datasets and degradation operators.
+# --------------------------------------------------------------------------------------------
+# In this example, we use the CBSD68 dataset for training and the set3c dataset for testing.
+# We work with images of size 32x32 if no GPU is available, else 128x128.
+
+
+operation = "inpainting"
+train_dataset_name = "CBSD68"
+test_dataset_name = "set3c"
+img_size = 128 if torch.cuda.is_available() else 32
+
+test_transform = transforms.Compose(
+    [transforms.CenterCrop(img_size), transforms.ToTensor()]
+)
+train_transform = transforms.Compose(
+    [transforms.RandomCrop(img_size), transforms.ToTensor()]
+)
+
+train_base_dataset = load_dataset(train_dataset_name, transform=train_transform)
+test_base_dataset = load_dataset(test_dataset_name, transform=test_transform)
+
+
+# %%
+# Define forward operator and generate dataset
+# --------------------------------------------------------------------------------------------
+# We define an inpainting operator that randomly masks pixels with probability 0.5.
+#
+# A dataset of pairs of measurements and ground truth images is then generated using the
+# :func:`deepinv.datasets.generate_dataset` function.
+#
+# Once the dataset is generated, we can load it using the :class:`deepinv.datasets.HDF5Dataset` class.
+
+n_channels = 3  # 3 for color images, 1 for gray-scale images
+probability_mask = 0.5  # probability to mask pixel
+
+# Generate inpainting operator
+physics = dinv.physics.Inpainting(
[…]

[…]
