### 2025-06-19 6fd05ed
make the trainer fail when a collision occurs
