### 2025-06-19 2b511d2
restore mistakenly modified test name
