+++ deepinv/tests/test_trainer.py
@@ -831,8 +831,8 @@ def test_out_dir_collision_detection(dummy_dataset, imsize, device, dummy_model)
 
     timestamp = get_timestamp()
 
-    # NOTE: Due to the way it's imported in the trainer module we need to patch
-    # the importing module instead of the imported module.
+    # NOTE: Due to the way it's imported in the trainer module we need to patch
+    # the importing module instead of the imported module.
     with patch.object(dinv.training.trainer, "get_timestamp", return_value=timestamp):
         with pytest.raises(FileExistsError, match=re.escape(timestamp)):
             # Train twice
