+++ deepinv/tests/test_trainer.py
@@ -17,11 +17,12 @@ import builtins
 import io
 import contextlib
 import re
+import tempfile
+import os
 
 from conftest import no_plot
 
 NO_LEARNING = ["A_dagger", "A_adjoint", "prox_l2", "y"]
-
 
 @pytest.fixture
 def imsize():
@@ -40,8 +41,15 @@ def physics(imsize, device):
     return dinv.physics.BlurFFT(img_size=imsize, filter=filter, device=device)
 
 
+@pytest.fixture
+def save_path():
+    # NOTE: The save path should be unique to avoid collision errors.
+    with tempfile.TemporaryDirectory() as dirname:
+        yield dirname
+
+
 @pytest.mark.parametrize("no_learning", NO_LEARNING)
-def test_nolearning(imsize, physics, model, no_learning, device):
+def test_nolearning(imsize, physics, model, no_learning, device, save_path):
     y = torch.ones((1,) + imsize, device=device)
     trainer = dinv.Trainer(
         model=model,
@@ -51,6 +59,7 @@ def test_nolearning(imsize, physics, model, no_learning, device):
         physics=physics,
         compare_no_learning=True,
         no_learning_method=no_learning,
+        save_path=save_path,
     )
     x_hat = trainer.no_learning_inference(y, physics)
     assert (physics.A(x_hat) - y).pow(2).mean() < 0.1
@@ -126,6 +135,7 @@ def test_get_samples(
     use_physics_generator,
     online_measurements,
     rng,
+    save_path,
 ):
     # Dummy constant GT dataset
     class DummyDataset(Dataset):
@@ -217,6 +227,7 @@ def test_get_samples(
             if online_measurements and physics_generator is not None
             else None
         ),
+        save_path=save_path,
     )
 
     iterator = iter(dataloader)
@@ -532,6 +543,7 @@ def test_dataloader_formats(
     measurements,
     online_measurements,
     rng,
+    save_path,
 ):
     """Test dataloader return formats
 
@@ -606,6 +618,7 @@ def test_dataloader_formats(
         online_measurements=online_measurements,
         train_dataloader=dataloader,
         optimizer=optimizer,
+        save_path=save_path,
     )
     trainer.setup_train()
     x, y, physics = trainer.get_samples([iter(dataloader)], 0)
@@ -658,7 +671,7 @@ def test_dataloader_formats(
 @pytest.mark.parametrize("early_stop", [True, False])
 @pytest.mark.parametrize("max_batch_steps", [3, 100000])
 def test_early_stop(
-    dummy_dataset, imsize, device, dummy_model, early_stop, max_batch_steps
+    dummy_dataset, imsize, device, dummy_model, early_stop, max_batch_steps, save_path
 ):
     torch.manual_seed(0)
     model = dummy_model
@@ -683,6 +696,7 @@ def test_early_stop(
         optimizer=optimizer,
         verbose=False,
         plot_images=True,
+        save_path=save_path,
     )
     with no_plot():
         trainer.train()
@@ -712,7 +726,7 @@ class ConstantLoss(dinv.loss.Loss):
         )
 
 
-def test_total_loss(dummy_dataset, imsize, device, dummy_model):
+def test_total_loss(dummy_dataset, imsize, device, dummy_model, save_path):
     train_data, eval_data = dummy_dataset, dummy_dataset
     dataloader = DataLoader(train_data, batch_size=2)
     eval_dataloader = DataLoader(eval_data, batch_size=2)
@@ -733,6 +747,7 @@ def test_total_loss(dummy_dataset, imsize, device, dummy_model):
         optimizer=torch.optim.AdamW(dummy_model.parameters(), lr=1),
         verbose=False,
         online_measurements=True,
+        save_path=save_path,
     )
 
     trainer.train()
[…]
