### 2025-06-19 3d737d1
add VP model in NCSNPP
