### 2025-06-19 1d8152d
test tensor list dirac_like
