+++ deepinv/tests/test_utils.py
@@ -77,6 +77,30 @@ def test_tensordict_append(tensorlist):
     assert (z1[0] == z[0]).all() and (z1[-1] == z[-1]).all()
 
 
+@pytest.mark.parametrize("shape", [(1, 1, 3, 3), (1, 1, 5, 5)])
+@pytest.mark.parametrize("length", [1, 2, 3, 4, 5])
+def test_dirac_like(shape, length):
+    rng = torch.Generator().manual_seed(0)
+    x = [torch.randn(shape, generator=rng) for _ in range(length)]
+    h = deepinv.utils.dirac_like(x)
+    y = deepinv.utils.TensorList(
+        [
+            deepinv.physics.functional.conv2d(xi, hi, padding="circular")
+            for hi, xi in zip(h, x, strict=True)
+        ]
+    )
+
+    for xi, hi, yi in zip(x, h, y, strict=True):
+        assert (
+            hi.shape == xi.shape
+        ), "Dirac delta should have the same shape as the input tensor."
+
+        if hi.shape[-2] % 2 == 1 and hi.shape[-1] % 2 == 1:
+            assert torch.allclose(
+                xi, yi
+            ), "Convolution with Dirac delta should return the original tensor."
+
+
 def test_plot():
     for c in range(1, 5):
         x = torch.ones((1, c, 2, 2))
