### 2025-06-20 dc34251
improve coverage of plotting functions and fix bugs
