### 2025-06-20 a8d52d8
replace custom fixture by standard fixture
