+++ deepinv/tests/test_models.py
@@ -476,6 +476,22 @@ def test_wavelet_denoiser_ths(
     x_hat = model(y, ths_tensor)
     assert x_hat.shape == y.shape
 
+    
+    # Test the wavelet decomposition and reconstruction
+    x = torch.randn(img_size, dtype=torch.float32).to(device)
+    # 1 decomposition
+    out = model.dwt(x)
+    x_hat = model.iwt(out)
+    assert x_hat.shape == x.shape and torch.allclose(x, x_hat, rtol=1e-5, atol=1e-5)
+    
+    # 2 decomposition
+    cA1, cD1 = model.dwt(x)
+    cA2, cD2 = model.dwt(cA1)
+
+    x_hat = model.iwt((cA2, cD2, cD1))
+    assert torch.allclose(x, x_hat, rtol=1e-5, atol=1e-5)
+
+
 
 def test_drunet_inputs(imsize_1_channel, device):
     f = dinv.models.DRUNet(
