+++ deepinv/models/diffunet.py
@@ -462,9 +462,9 @@ class DiffUNet(Denoiser):
         :return: an `(N, C, ...)` Tensor of outputs.
         """
         if not isinstance(sigma, torch.Tensor):
-            sigma = torch.tensor([sigma]).to(x.device, x.dtype).view(1, 1, 1, 1)
+            sigma = torch.tensor([sigma]).to(x.device, x.dtype).view((1,) * (x.ndim))
         else:
-            sigma = sigma.squeeze().view((sigma.shape + (1,) * (4 - sigma.ndim)))
+            sigma = sigma.squeeze().view((sigma.shape + (1,) * (x.ndim - sigma.ndim)))
         alpha = 1 / (1 + 4 * sigma**2)
         x = alpha.sqrt() * (2 * x - 1)
         sigma = sigma * alpha.sqrt()
