+++ deepinv/tests/test_models.py
@@ -475,6 +475,17 @@ def test_wavelet_denoiser_ths(
     x_hat = model(y, ths_tensor)
     assert x_hat.shape == y.shape
 
+
+@pytest.mark.parametrize("channels", [1, 3])
+@pytest.mark.parametrize("batch_size", [1, 2])
+@pytest.mark.parametrize("dimension", [2, 3])
+def test_wavelet_decomposition(channels, dimension, batch_size, device):
+    model = dinv.models.WaveletDenoiser(level=1, wvdim=dimension).to(device)
+    img_size = (
+        (batch_size, channels, 64, 64)
+        if dimension == 2
+        else (batch_size, channels, 8, 64, 64)
+    )
     # Test the wavelet decomposition and reconstruction
     x = torch.randn(img_size, dtype=torch.float32).to(device)
     # 1 decomposition
