+++ deepinv/utils/parameters.py
@@ -1,4 +1,3 @@
-import numpy as np
 
 
 def get_GSPnP_params(problem, noise_level_img):
