+++ deepinv/optim/optim_iterators/primal_dual_CP.py
@@ -1,4 +1,3 @@
-import torch
 
 from .optim_iterator import OptimIterator, fStep, gStep
 
