+++ deepinv/sampling/__init__.py
@@ -1,12 +1,12 @@
-from .langevin import ULA, MonteCarlo, SKRock
-from .diffusion import DDRM, DiffusionSampler, DiffPIR, DPS
-from . import diffusion_sde, sde_solver
-from .noisy_datafidelity import NoisyDataFidelity, DPSDataFidelity
+from .langevin import ULA as ULA, MonteCarlo as MonteCarlo, SKRock as SKRock
+from .diffusion import DDRM as DDRM, DiffusionSampler as DiffusionSampler, DiffPIR as DiffPIR, DPS as DPS
+from . import diffusion_sde as diffusion_sde, sde_solver as sde_solver
+from .noisy_datafidelity import NoisyDataFidelity as NoisyDataFidelity, DPSDataFidelity as DPSDataFidelity
 from .diffusion_sde import (
-    BaseSDE,
-    DiffusionSDE,
-    VarianceExplodingDiffusion,
-    VariancePreservingDiffusion,
-    PosteriorDiffusion,
+    BaseSDE as BaseSDE,
+    DiffusionSDE as DiffusionSDE,
+    VarianceExplodingDiffusion as VarianceExplodingDiffusion,
+    VariancePreservingDiffusion as VariancePreservingDiffusion,
+    PosteriorDiffusion as PosteriorDiffusion,
 )
-from .sde_solver import SDEOutput, BaseSDESolver, EulerSolver, HeunSolver
+from .sde_solver import SDEOutput as SDEOutput, BaseSDESolver as BaseSDESolver, EulerSolver as EulerSolver, HeunSolver as HeunSolver
