+++ deepinv/loss/mri/__init__.py
@@ -1,6 +1,6 @@
 from .measplit import (
-    Phase2PhaseLoss,
-    Artifact2ArtifactLoss,
-    WeightedSplittingLoss,
-    RobustSplittingLoss,
+    Phase2PhaseLoss as Phase2PhaseLoss,
+    Artifact2ArtifactLoss as Artifact2ArtifactLoss,
+    WeightedSplittingLoss as WeightedSplittingLoss,
+    RobustSplittingLoss as RobustSplittingLoss,
 )
