+++ deepinv/training/__init__.py
@@ -1,3 +1,3 @@
-from .trainer import Trainer, train
-from .testing import test
-from .adversarial import AdversarialTrainer, AdversarialOptimizer
+from .trainer import Trainer as Trainer, train as train
+from .testing import test as test
+from .adversarial import AdversarialTrainer as AdversarialTrainer, AdversarialOptimizer as AdversarialOptimizer
