+++ deepinv/utils/__init__.py
@@ -1,39 +1,39 @@
-from .logger import AverageMeter, ProgressMeter, get_timestamp
-from .metric import cal_psnr, cal_mse, cal_psnr_complex
+from .logger import AverageMeter as AverageMeter, ProgressMeter as ProgressMeter, get_timestamp as get_timestamp
+from .metric import cal_psnr as cal_psnr, cal_mse as cal_mse, cal_psnr_complex as cal_psnr_complex
 from .plotting import (
-    rescale_img,
-    plot,
-    torch2cpu,
-    plot_curves,
-    plot_parameters,
-    plot_inset,
-    plot_videos,
-    save_videos,
-    make_grid,
-    wandb_imgs,
-    wandb_plot_curves,
-    resize_pad_square_tensor,
-    scatter_plot,
-    plot_ortho3D,
+    rescale_img as rescale_img,
+    plot as plot,
+    torch2cpu as torch2cpu,
+    plot_curves as plot_curves,
+    plot_parameters as plot_parameters,
+    plot_inset as plot_inset,
+    plot_videos as plot_videos,
+    save_videos as save_videos,
+    make_grid as make_grid,
+    wandb_imgs as wandb_imgs,
+    wandb_plot_curves as wandb_plot_curves,
+    resize_pad_square_tensor as resize_pad_square_tensor,
+    scatter_plot as scatter_plot,
+    plot_ortho3D as plot_ortho3D,
 )
 from .demo import (
-    load_url_image,
-    load_example,
-    load_image,
-    load_dataset,
-    load_degradation,
-    get_data_home,
-    get_image_url,
-    get_degradation_url,
+    load_url_image as load_url_image,
+    load_example as load_example,
+    load_image as load_image,
+    load_dataset as load_dataset,
+    load_degradation as load_degradation,
+    get_data_home as get_data_home,
+    get_image_url as get_image_url,
+    get_degradation_url as get_degradation_url,
 )
-from .nn import get_freer_gpu
+from .nn import get_freer_gpu as get_freer_gpu
 from .tensorlist import (
-    TensorList,
-    rand_like,
-    zeros_like,
-    randn_like,
-    ones_like,
-    dirac_like,
+    TensorList as TensorList,
+    rand_like as rand_like,
+    zeros_like as zeros_like,
+    randn_like as randn_like,
+    ones_like as ones_like,
+    dirac_like as dirac_like,
 )
-from .phantoms import RandomPhantomDataset, SheppLoganDataset
-from .patch_extractor import patch_extractor
+from .phantoms import RandomPhantomDataset as RandomPhantomDataset, SheppLoganDataset as SheppLoganDataset
+from .patch_extractor import patch_extractor as patch_extractor
