+++ deepinv/physics/generator/__init__.py
@@ -1,26 +1,26 @@
-from .base import PhysicsGenerator, GeneratorMixture
+from .base import PhysicsGenerator as PhysicsGenerator, GeneratorMixture as GeneratorMixture
 from .blur import (
-    MotionBlurGenerator,
-    DiffractionBlurGenerator,
-    PSFGenerator,
-    ProductConvolutionBlurGenerator,
-    DiffractionBlurGenerator3D,
-    ConfocalBlurGenerator3D,
-    bump_function,
+    MotionBlurGenerator as MotionBlurGenerator,
+    DiffractionBlurGenerator as DiffractionBlurGenerator,
+    PSFGenerator as PSFGenerator,
+    ProductConvolutionBlurGenerator as ProductConvolutionBlurGenerator,
+    DiffractionBlurGenerator3D as DiffractionBlurGenerator3D,
+    ConfocalBlurGenerator3D as ConfocalBlurGenerator3D,
+    bump_function as bump_function,
 )
 from .mri import (
-    BaseMaskGenerator,
-    GaussianMaskGenerator,
-    RandomMaskGenerator,
-    EquispacedMaskGenerator,
-    PolyOrderMaskGenerator,
+    BaseMaskGenerator as BaseMaskGenerator,
+    GaussianMaskGenerator as GaussianMaskGenerator,
+    RandomMaskGenerator as RandomMaskGenerator,
+    EquispacedMaskGenerator as EquispacedMaskGenerator,
+    PolyOrderMaskGenerator as PolyOrderMaskGenerator,
 )
-from .noise import SigmaGenerator, GainGenerator
+from .noise import SigmaGenerator as SigmaGenerator, GainGenerator as GainGenerator
 from .inpainting import (
-    BernoulliSplittingMaskGenerator,
-    GaussianSplittingMaskGenerator,
-    Artifact2ArtifactSplittingMaskGenerator,
-    Phase2PhaseSplittingMaskGenerator,
-    MultiplicativeSplittingMaskGenerator,
+    BernoulliSplittingMaskGenerator as BernoulliSplittingMaskGenerator,
+    GaussianSplittingMaskGenerator as GaussianSplittingMaskGenerator,
+    Artifact2ArtifactSplittingMaskGenerator as Artifact2ArtifactSplittingMaskGenerator,
+    Phase2PhaseSplittingMaskGenerator as Phase2PhaseSplittingMaskGenerator,
+    MultiplicativeSplittingMaskGenerator as MultiplicativeSplittingMaskGenerator,
 )
-from .downsampling import DownsamplingGenerator
+from .downsampling import DownsamplingGenerator as DownsamplingGenerator

[…]
