+++ deepinv/training/__init__.py
@@ -1,3 +1,6 @@
 from .trainer import Trainer as Trainer, train as train
 from .testing import test as test
-from .adversarial import AdversarialTrainer as AdversarialTrainer, AdversarialOptimizer as AdversarialOptimizer
+from .adversarial import (
+    AdversarialTrainer as AdversarialTrainer,
+    AdversarialOptimizer as AdversarialOptimizer,
+)
