+++ deepinv/optim/optim_iterators/__init__.py
@@ -1,6 +1,14 @@
-from .optim_iterator import OptimIterator as OptimIterator, fStep as fStep, gStep as gStep
+from .optim_iterator import (
+    OptimIterator as OptimIterator,
+    fStep as fStep,
+    gStep as gStep,
+)
 from .admm import ADMMIteration as ADMMIteration
-from .pgd import PGDIteration as PGDIteration, FISTAIteration as FISTAIteration, PMDIteration as PMDIteration
+from .pgd import (
+    PGDIteration as PGDIteration,
+    FISTAIteration as FISTAIteration,
+    PMDIteration as PMDIteration,
+)
 from .primal_dual_CP import CPIteration as CPIteration
 from .hqs import HQSIteration as HQSIteration
 from .drs import DRSIteration as DRSIteration
