+++ deepinv/sampling/__init__.py
@@ -1,7 +1,15 @@
 from .langevin import ULA as ULA, MonteCarlo as MonteCarlo, SKRock as SKRock
-from .diffusion import DDRM as DDRM, DiffusionSampler as DiffusionSampler, DiffPIR as DiffPIR, DPS as DPS
+from .diffusion import (
+    DDRM as DDRM,
+    DiffusionSampler as DiffusionSampler,
+    DiffPIR as DiffPIR,
+    DPS as DPS,
+)
 from . import diffusion_sde as diffusion_sde, sde_solver as sde_solver
-from .noisy_datafidelity import NoisyDataFidelity as NoisyDataFidelity, DPSDataFidelity as DPSDataFidelity
+from .noisy_datafidelity import (
+    NoisyDataFidelity as NoisyDataFidelity,
+    DPSDataFidelity as DPSDataFidelity,
+)
 from .diffusion_sde import (
     BaseSDE as BaseSDE,
     DiffusionSDE as DiffusionSDE,
@@ -9,4 +17,9 @@ from .diffusion_sde import (
     VariancePreservingDiffusion as VariancePreservingDiffusion,
     PosteriorDiffusion as PosteriorDiffusion,
 )
-from .sde_solver import SDEOutput as SDEOutput, BaseSDESolver as BaseSDESolver, EulerSolver as EulerSolver, HeunSolver as HeunSolver
+from .sde_solver import (
+    SDEOutput as SDEOutput,
+    BaseSDESolver as BaseSDESolver,
+    EulerSolver as EulerSolver,
+    HeunSolver as HeunSolver,
+)
