### 2025-06-20 8ab2ae5
fix the 4 last errors manually
