+++ deepinv/__init__.py
@@ -60,6 +60,7 @@ __all__ += ["training"]
 
 from deepinv.training import train, test, Trainer
 
+__all__ += ["train", "test", "Trainer"]
 
 # GLOBAL PROPERTY
 dtype = torch.float
