### 2025-06-20 652403d
disable ruff for an import with side effects
