### 2025-06-20 ef7b05a
ignore ruff false positive for fixture
