### 2025-06-20 6046a23
test denoiser perf
