+++ deepinv/tests/test_models.py
@@ -4,6 +4,8 @@ import torch
 from torch.utils.data import DataLoader, Dataset
 
 import deepinv as dinv
+from deepinv.loss import PSNR
+
 from dummy import DummyCircles, DummyModel
 
 
@@ -18,7 +20,7 @@ MODEL_LIST_1_CHANNEL = [
     "epll",
     "restormer",
     "ncsnpp",
-    "admunet",
+    "adinv.modelsunet",
 ]
 MODEL_LIST = MODEL_LIST_1_CHANNEL + [
     "bm3d",
@@ -95,8 +97,8 @@ def choose_denoiser(name, imsize):
             img_resolution=imsize[1],
             pretrained=None,
         )
-    elif name == "admunet":
-        out = dinv.models.ADMUNet(
+    elif name == "adinv.modelsunet":
+        out = dinv.models.Adinv.modelsUNet(
             in_channels=imsize[0],
             out_channels=imsize[0],
             img_resolution=imsize[1],
@@ -275,8 +277,8 @@ def test_TV_models_identity():
 
 @pytest.mark.parametrize("denoiser", MODEL_LIST)
 def test_denoiser_color(imsize, device, denoiser):
-    # NCSNpp and ADMUnet only support imsize that divisible by 8
-    if denoiser in ["ncsnpp", "admunet"]:
+    # NCSNpp and Adinv.modelsUnet only support imsize that divisible by 8
+    if denoiser in ["ncsnpp", "adinv.modelsunet"]:
         imsize = (imsize[0], (imsize[1] // 8) * 8, (imsize[2] // 8) * 8)
 
     model = choose_denoiser(denoiser, imsize).to(device)
@@ -294,8 +296,8 @@ def test_denoiser_color(imsize, device, denoiser):
 def test_denoiser_gray(imsize_1_channel, device, denoiser):
     # except scunet and dsccp, all models support 1 channel
     if denoiser not in ["scunet", "dsccp"]:
-        # NCSNpp and ADMUnet only support imsize that divisible by 8
-        if denoiser in ["ncsnpp", "admunet"]:
+        # NCSNpp and Adinv.modelsUnet only support imsize that divisible by 8
+        if denoiser in ["ncsnpp", "adinv.modelsunet"]:
             imsize_1_channel = (
                 imsize_1_channel[0],
                 (imsize_1_channel[1] // 8) * 8,
@@ -367,7 +369,7 @@ def test_equivariant(imsize, device, batch_size):
 
 @pytest.mark.parametrize("denoiser", MODEL_LIST_1_CHANNEL)
 def test_denoiser_1_channel(imsize_1_channel, device, denoiser):
-    if denoiser in ["ncsnpp", "admunet"]:
+    if denoiser in ["ncsnpp", "adinv.modelsunet"]:
         imsize_1_channel = list(imsize_1_channel)
         if imsize_1_channel[1] % 8 > 0:
             imsize_1_channel[1] = 32
@@ -827,7 +829,7 @@ def test_ncsnpp_net(device, image_size, n_channels, batch_size, precond, use_fp1
         pretrained=None,
     )
     if precond:
-        model = dinv.models.EDMPrecond(model, use_fp16=use_fp16).to(device)
+        model = dinv.models.Edinv.modelsPrecond(model, use_fp16=use_fp16).to(device)
     else:
         model = model.to(device)
     x = torch.rand(batch_size, n_channels, image_size, image_size, device=device)
@@ -865,3 +867,99 @@ def test_dsccp_net(device, n_channels):
     x = x.expand(4, -1, -1, -1)
     y = model(x, torch.linspace(0.01, 0.1, 4, device=device))
     assert y.shape == x.shape
+
+
+def test_denoiser_perf(device):
[…]
