### 2025-06-20 8e2012a
add tests for plot_curves
