### 2025-06-20 4a88f92
add test parameters and fix bugs
