+++ deepinv/utils/plotting.py
@@ -425,7 +425,8 @@ def scatter_plot(
     if save_dir:
         plt.savefig(save_dir / "images.png", dpi=1200)
         for i, row_scatter in enumerate(scatters):
-            save_dir_i = Path(save_dir) / Path(titles[i])
+            row_dirname = titles[i] if titles is not None else f"i"
+            save_dir_i = Path(save_dir) / Path(row_dirname)
             save_dir_i.mkdir(parents=True, exist_ok=True)
             for r, img in enumerate(row_scatter):
                 plt.imsave(save_dir_i / Path(str(r) + ".png"), img, cmap=cmap)
