### 2025-06-20 8f898a7
test resize_pad_square_tensor and torch2cpu
