### 2025-06-21 e35c33e
test get_freer_gpu
