### 2025-06-21 59c1bc3
test deprecation of deepinv.utils.metric functions
