### 2025-06-21 849c818
test load_degradation
