### 2025-06-21 fa7d982
Fix data loading in demo_r2r_denoising.py to use the training dataset instead of the test dataset
