+++ examples/self-supervised-learning/demo_r2r_denoising.py
@@ -151,7 +151,7 @@ if noise_name == "poisson":
 #
 #
 
-verbose = True  # print training information
+verbose = True     # print training information
 wandb_vis = False  # plot curves and images in Weight&Bias
 
 train_dataloader = DataLoader(
