### 2025-06-21 6d4eada
test test_get_GSPnP_params
