### 2025-06-21 00e56ac
test tensorlist methods
