### 2025-06-21 2aae932
test load_image
