### 2025-06-21 bfd5354
fix the tests on CPU-only devices
