### 2025-06-21 173db36
improve non-idiomatic code
