### 2025-06-22 36c0709
restore demos
