### 2025-06-23 feecb82
pin sphinx_sitemap to v2.6.0 see #508 and #555
