### 2025-06-23 f7550e9
Remove fetch-depth option from GitHub Actions workflow and pin sphinx_sitemap dependency version
