### 2025-06-23 ecee84c
replace pip with uv in ci
