+++ examples/plug-and-play/demo_ram.py
@@ -6,12 +6,12 @@ This example shows how to use the RAM model method to solve inverse problems. Th
 the following `paper <https://arxiv.org/abs/2503.08915>`_, is a modified DRUNet architecture that is trained on
 a large number of inverse problems.
 """
+
 import torch
 import deepinv as dinv
 from deepinv.models import RAM
 
-import torch
-device = 'cuda' if torch.cuda.is_available() else 'cpu'
+device = "cuda" if torch.cuda.is_available() else "cpu"
 
 # Load the pretrained model
 model = RAM(device=device)
@@ -20,7 +20,12 @@ model = RAM(device=device)
 x = dinv.utils.load_example("butterfly.png").to(device)
 
 # create forward operator
-physics = dinv.physics.Inpainting(tensor_size=(3, 256, 256), mask=.3, noise_model=dinv.physics.GaussianNoise(.05), device=device)
+physics = dinv.physics.Inpainting(
+    tensor_size=(3, 256, 256),
+    mask=0.3,
+    noise_model=dinv.physics.GaussianNoise(0.05),
+    device=device,
+)
 
 # generate measurement
 y = physics(x)
@@ -34,7 +39,14 @@ in_psnr = dinv.metric.PSNR()(x, y).item()
 out_psnr = dinv.metric.PSNR()(x, x_hat).item()
 
 # plot
-dinv.utils.plot([x, y, x_hat], ["Original", "Measurement\n PSNR = {:.2f}dB".format(in_psnr), "Reconstruction\n PSNR = {:.2f}dB".format(out_psnr)])
+dinv.utils.plot(
+    [x, y, x_hat],
+    [
+        "Original",
+        "Measurement\n PSNR = {:.2f}dB".format(in_psnr),
+        "Reconstruction\n PSNR = {:.2f}dB".format(out_psnr),
+    ],
+)
 
 # %%
 # This model is not trained on all degradations, so it may not perform well on all inverse problems.
@@ -43,7 +55,9 @@ dinv.utils.plot([x, y, x_hat], ["Original", "Measurement\n PSNR = {:.2f}dB".form
 
 
 # Define the Demosaicing physics
-physics = dinv.physics.Demosaicing(img_size=(3, 256, 256), noise_model=dinv.physics.GaussianNoise(.05), device=device)
+physics = dinv.physics.Demosaicing(
+    img_size=(3, 256, 256), noise_model=dinv.physics.GaussianNoise(0.05), device=device
+)
 
 # generate measurement
 y = physics(x)
@@ -57,7 +71,14 @@ in_psnr = dinv.metric.PSNR()(x, y).item()
 out_psnr = dinv.metric.PSNR()(x, x_hat).item()
 
 # plot
-dinv.utils.plot([x, y, x_hat], ["Original", "Measurement\n PSNR = {:.2f}dB".format(in_psnr), "0 shot reconstruction\n PSNR = {:.2f}dB".format(out_psnr)])
+dinv.utils.plot(
+    [x, y, x_hat],
+    [
+        "Original",
+        "Measurement\n PSNR = {:.2f}dB".format(in_psnr),
+        "0 shot reconstruction\n PSNR = {:.2f}dB".format(out_psnr),
+    ],
+)
 
 
 # %%
@@ -70,7 +91,11 @@ from deepinv.datasets.utils import UnsupDataset
 
 # Define small physics in the case of no-GPU
 if not torch.cuda.is_available():
-    physics_train = dinv.physics.Demosaicing(img_size=(3, 64, 64), noise_model=dinv.physics.GaussianNoise(.05), device=device)
+    physics_train = dinv.physics.Demosaicing(
+        img_size=(3, 64, 64),
[…]
