### 2025-06-24 1d41446
VP diffusion
