+++ examples/optimization/demo_TV_minimisation.py
@@ -1,190 +1,0 @@
-r"""
-Image deblurring with Total-Variation (TV) prior
-====================================================================================================
-
-This example shows how to use a standard TV prior for image deblurring. The problem writes as :math:`y = Ax + \epsilon`
-where :math:`A` is a convolutional operator and :math:`\epsilon` is the realization of some Gaussian noise. The goal is
-to recover the original image :math:`x` from the blurred and noisy image :math:`y`. The TV prior is used to regularize
-the problem.
-"""
-
-import deepinv as dinv
-from pathlib import Path
-import torch
-from torchvision import transforms
-
-from deepinv.optim.data_fidelity import L2
-from deepinv.optim.optimizers import optim_builder
-from deepinv.utils.demo import load_dataset, load_degradation
-from deepinv.utils.plotting import plot, plot_curves
-
-# %%
-# Setup paths for data loading and results.
-# ----------------------------------------------------------------------------------------
-#
-
-BASE_DIR = Path(".")
-DATA_DIR = BASE_DIR / "measurements"
-RESULTS_DIR = BASE_DIR / "results"
-DEG_DIR = BASE_DIR / "degradations"
-
-
-# %%
-# Load base image datasets and degradation operators.
-# ----------------------------------------------------------------------------------------
-# In this example, we use the Set3C dataset and a motion blur kernel from
-# `Levin et al. (2009) <https://ieeexplore.ieee.org/abstract/document/5206815/>`_.
-#
-
-# Set the global random seed from pytorch to ensure reproducibility of the example.
-torch.manual_seed(0)
-
-device = dinv.utils.get_freer_gpu() if torch.cuda.is_available() else "cpu"
-
-# Set up the variable to fetch dataset and operators.
-dataset_name = "set3c"
-img_size = 256 if torch.cuda.is_available() else 64
-val_transform = transforms.Compose(
-    [transforms.CenterCrop(img_size), transforms.ToTensor()]
-)
-
-# Generate a motion blur operator.
-kernel_index = 1  # which kernel to chose among the 8 motion kernels from 'Levin09.mat'
-kernel_torch = load_degradation("Levin09.npy", DEG_DIR / "kernels", index=kernel_index)
-kernel_torch = kernel_torch.unsqueeze(0).unsqueeze(
-    0
-)  # add batch and channel dimensions
-dataset = load_dataset(dataset_name, transform=val_transform)
-
-
-# %%
-# Generate a dataset of blurred images and load it.
-# --------------------------------------------------------------------------------
-# We use the BlurFFT class from the physics module to generate a dataset of blurred images.
-
-
-noise_level_img = 0.05  # Gaussian Noise standard deviation for the degradation
-n_channels = 3  # 3 for color images, 1 for gray-scale images
-physics = dinv.physics.BlurFFT(
-    img_size=(n_channels, img_size, img_size),
-    filter=kernel_torch,
-    device=device,
-    noise_model=dinv.physics.GaussianNoise(sigma=noise_level_img),
-)
-
-# Select the first image from the dataset
-x = dataset[0][0].unsqueeze(0).to(device)
-
-# Apply the degradation to the image
-y = physics(x)
-
-# %%
-# Exploring the total variation prior.
-# ------------------------------------
-#
-# In this example, we will use the total variation prior, which can be done with the :class:`deepinv.optim.prior.Prior`
-# class. The prior object represents the cost function of the prior (TV in this case), as well as convenient methods,
[…]
