### 2025-06-24 8940209
set correct env for order in nvida-smi
