### 2025-06-24 c07471d
refactor _handle_sigma
