+++ deepinv/models/precond.py
@@ -42,7 +42,7 @@ class EDMPrecond(Denoiser):
 
     def forward(self, x, sigma, class_labels=None, force_fp32=False, **model_kwargs):
         x = x.to(torch.float32)
-        sigma = self._handle_sigma(sigma, torch.float32, x.device, x.size(0))
+        sigma = self._handle_sigma(sigma, batch_size=x.size(0), ndim=x.ndim)
         if class_labels is not None:
             class_labels = class_labels.to(torch.float32)
         dtype = (
@@ -65,22 +65,3 @@ class EDMPrecond(Denoiser):
         assert F_x.dtype == dtype
         D_x = c_skip * x + c_out * F_x.to(torch.float32)
         return D_x
-
-    @staticmethod
-    def _handle_sigma(sigma, dtype, device, batch_size):
-        if isinstance(sigma, torch.Tensor):
-            if sigma.ndim == 0:
-                return sigma[None].to(device, dtype).view(-1, 1, 1, 1)
-            elif sigma.ndim == 1:
-                assert (
-                    sigma.size(0) == batch_size or sigma.size(0) == 1
-                ), "sigma must be a Tensor with batch_size equal to 1 or the batch_size of input images"
-                return sigma.to(device, dtype).view(-1, 1, 1, 1)
-
-            else:
-                raise ValueError(f"Unsupported sigma shape {sigma.shape}.")
-
-        elif isinstance(sigma, (float, int)):
-            return torch.tensor([sigma]).to(device, dtype).reshape(-1, 1, 1, 1)
-        else:
-            raise ValueError("Unsupported sigma type.")
