### 2025-06-24 9a07131
remove refs from first line in class doc
