+++ deepinv/datasets/lsdir.py
@@ -13,7 +13,9 @@ from deepinv.datasets.utils import (
 
 
 class LsdirHR(torch.utils.data.Dataset):
-    """Dataset for `LSDIR <https://data.vision.ee.ethz.ch/yawli/>`_.
+    """Dataset for `LSDIR <https://ofsoundof.github.io/lsdir-data/>`_.
+
+    Published in :footcite:ps:`li2023lsdir`.
 
     A large-scale dataset for image restoration tasks such as image super-resolution (SR),
     image denoising, JPEG deblocking, deblurring, and demosaicking, and real-world SR.
@@ -61,6 +63,8 @@ class LsdirHR(torch.utils.data.Dataset):
             print(val_dataset.verify_split_dataset_integrity())             # check that raw data has been downloaded correctly
             print(len(val_dataset))                                         # check that we have 250 images
 
+    :References:
+        .. footbibliography::
     """
 
     archive_urls = {
