+++ deepinv/loss/augmentation.py
@@ -19,8 +19,7 @@ from deepinv.physics.forward import LinearPhysics
 class AugmentConsistencyLoss(Loss):
     r"""Data augmentation consistency (DAC) loss.
 
-    Performs data augmentation in measurement domain as proposed by
-    `VORTEX: Physics-Driven Data Augmentations Using Consistency Training for Robust Accelerated MRI Reconstruction <https://arxiv.org/abs/2111.02549>`_.
+    Performs data augmentation in measurement domain as proposed by `VORTEX: Physics-Driven Data Augmentations Using Consistency Training for Robust Accelerated MRI Reconstruction <https://arxiv.org/abs/2111.02549>`_ :footcite:ps:`desai2021vortex`.
 
     The loss is defined as follows:
 
@@ -47,6 +46,9 @@ class AugmentConsistencyLoss(Loss):
     :param bool no_grad: if ``True``, only propagate gradients through augmented branch as per original paper,
         if ``False``, propagate through both branches.
     :param torch.Generator rng: torch random number generator to pass to transforms.
+
+    :References:
+        .. footbibliography::
     """
 
     def __init__(
