+++ deepinv/models/GSPnP.py
@@ -16,6 +16,7 @@ class StudentGrad(nn.Module):
 class GSPnP(Denoiser):
     r"""
     Gradient Step module to use a denoiser architecture as a Gradient Step Denoiser.
+    
     See :footcite:ps:`hurault2021gradient`. Code from https://github.com/samuro95/GSPnP.
 
     :param torch.nn.Module denoiser: Denoiser model.
@@ -78,7 +79,9 @@ def GSDRUNet(
     device=torch.device("cpu"),
 ):
     """
-    Gradient Step Denoiser with DRUNet architecture from :footcite:ps:`hurault2021gradient`.
+    Gradient Step Denoiser with DRUNet architecture.
+     
+    Based on the GSPnP method :footcite:ps:`hurault2021gradient`.
 
     :param float alpha: Relaxation parameter
     :param int in_channels: Number of input channels

[…]
