+++ deepinv/models/GSPnP.py
@@ -16,7 +16,7 @@ class StudentGrad(nn.Module):
 class GSPnP(Denoiser):
     r"""
     Gradient Step module to use a denoiser architecture as a Gradient Step Denoiser.
-    
+
     See :footcite:ps:`hurault2021gradient`. Code from https://github.com/samuro95/GSPnP.
 
     :param torch.nn.Module denoiser: Denoiser model.
@@ -80,7 +80,7 @@ def GSDRUNet(
 ):
     """
     Gradient Step Denoiser with DRUNet architecture.
-     
+
     Based on the GSPnP method :footcite:ps:`hurault2021gradient`.
 
     :param float alpha: Relaxation parameter
