### 2025-06-24 07b59cb
correction sampling rst
