### 2025-06-24 f748443
correction sampling rst
