+++ docs/source/user_guide/reconstruction/sampling.rst
@@ -122,8 +122,11 @@ In the case of posterior sampling, we need simply to replace the (unconditional)
 by the conditional score function :math:`\nabla \log p_t(x_t|y)`. The conditional score can be decomposed using the Bayes' rule:
 
 .. math::
-
-    \nabla \log p_t(x_t | y) = \nabla \log p_t(x_t) + \nabla \log p_t \left(y \vert x_t = s(t)x_0 + \right).
+    \begin{align}
+    \nabla_{x_t} \log p_t(x_t | y) &= \nabla_{x_t} \log p_t(x_t) + \nabla_{x_t} \log p_t \left(y \vert x_t \right) \\
+                            &= \nabla_{x_t} \log p_t(x_t) + \frac{1}{s_t} \nabla_{x_t} \log \hat p_t \left(y \vert \frac{x_t}{s(t)} = x_0 + \sigma(t) \omega \right).
+    \end{align}
+where $\hat p_t$ stands for the unscaled distribution, i.e., $\hat p_t(x_0 + \sigma(t) \omega) = p_t(s_t (x_0 + \sigma(t) \omega))$.
 
 The first term is the unconditional score function and can be approximated by using a denoiser as explained previously. 
 The second term is the conditional score function, and can be approximated by the (noisy) data-fidelity term.
