### 2025-06-24 8b16f5b
correction sampling rst
