### 2025-06-24 e588edc
dtype device to _handle_sigma
