### 2025-06-24 01f94ce
Fix some issues related to denoiser (#530)
* fix sigma issue in denoiser

* fixed some denoisers

* wavelet denoiser batch

* wavelet denoiser batch, level, dimension

* fixed epll batch sigma

* fixed wavelet topk

* test wavelet denoiser ths

* test models passed

* pretrained

* fixed tests

* fix demo

* fix ncsnpp

* batching wavelet prior

* batching wavelet prior, optimizers flatten reg param

* batching sigma of dsccp model

* dscnn as denoiser

* fix #548

* fix #507

* Julian comments

* refactor test

* test denoiser perf

* undo auto replace

* refactor _handle_sigma

* dtype devic
[…]
