+++ deepinv/models/GSPnP.py
@@ -1,5 +1,8 @@
 import torch
 import torch.nn as nn
+from torch import Tensor
+
+from typing import Union
 from .utils import get_weights_url
 from .base import Denoiser
 
@@ -21,14 +24,20 @@ class GSPnP(Denoiser):
 
     :param torch.nn.Module denoiser: Denoiser model.
     :param float alpha: Relaxation parameter
+    :param bool detach: If `True`, the denoiser output will be detached from the computation graph.
+        Setting this to `False` allows one to compute the gradient of the denoiser output with respect to the input, it is necessary in training.
+        Default is `True`.
     """
 
-    def __init__(self, denoiser, alpha=1.0):
+    def __init__(self, denoiser, alpha: float = 1.0, detach: bool = True):
         super().__init__()
         self.student_grad = StudentGrad(denoiser)
         self.alpha = alpha
+        self.detach = detach
 
-    def potential(self, x, sigma, *args, **kwargs):
+    def potential(
+        self, x: Tensor, sigma: Union[float, torch.Tensor], *args, **kwargs
+    ) -> Tensor:
         N = self.student_grad(x, sigma)
         return (
             0.5
@@ -36,7 +45,9 @@ class GSPnP(Denoiser):
             * torch.norm((x - N).view(x.shape[0], -1), p=2, dim=-1) ** 2
         )
 
-    def potential_grad(self, x, sigma, *args, **kwargs):
+    def potential_grad(
+        self, x: Tensor, sigma: Union[float, torch.Tensor], *args, **kwargs
+    ) -> Tensor:
         r"""
         Calculate :math:`\nabla g` the gradient of the regularizer :math:`g` at input :math:`x`.
 
@@ -44,16 +55,20 @@ class GSPnP(Denoiser):
         :param float sigma: Denoiser level :math:`\sigma` (std)
         """
         with torch.enable_grad():
-            x = x.float()
+            x = x.to(torch.float32)
             x = x.requires_grad_()
             N = self.student_grad(x, sigma)
             JN = torch.autograd.grad(
-                N, x, grad_outputs=x - N, create_graph=True, only_inputs=True
+                N, x, grad_outputs=x - N, create_graph=False if self.detach else True
             )[0]
+        if self.detach:
+            x = x.detach()
+            JN = JN.detach()
+
         Dg = x - N - JN
         return self.alpha * Dg
 
-    def forward(self, x, sigma):
+    def forward(self, x: Tensor, sigma: Union[float, torch.Tensor]) -> Tensor:
         r"""
         Denoising with Gradient Step Denoiser
 
