### 2025-06-24 7f292c5
curreent state
