+++ deepinv/sampling/noisy_datafidelity.py
@@ -1,107 +1,137 @@
 import torch
-from deepinv.optim import DataFidelity, Distance
+from deepinv.optim import DataFidelity, Distance, Potential
 import deepinv as dinv
 from deepinv.physics import Physics
 from deepinv.models import Denoiser
+from typing import Union
+
+# Question : the base should have default value to automatic differetiation : i.e. not a default distance : so it should be a child of Potential, not DataFidelity ? 
 
 
-class NoisyDataFidelity(DataFidelity):
+class NoisyDataFidelity(Potential):
     r"""
-    Preconditioned data fidelity term for noisy data :math:`- \log p(y|x + \sigma(t) \omega)`
-    with :math:`\omega\sim\mathcal{N}(0,\mathrm{I})`.
+    Base class for implementing the noisy data fidelity term :math:`- \log p(y|x_\sigma = x + \sigma \omega)` with :math:`\omega\sim\mathcal{N}(0,\mathrm{I})`.
 
-    This is a base class for the conditional classes for approximating :math:`\log p_t(y|x_t)` used in diffusion
-    algorithms for inverse problems, in :class:`deepinv.sampling.PosteriorDiffusion`.
+    This class is used to approximate the untractable likelihood term :math:`\log p_t(y|x_t)` in diffusion-based
+    algorithms for inverse problems, as defined in :class:`deepinv.sampling.PosteriorDiffusion`.
 
-    It comes with a `.grad` method computing the score :math:`\nabla_{x_t} \log p_t(y|x_t)`.
+    It comes with a `.grad` method for computing an approximation of the score :math:`\nabla_{x_\sigma} \log p_t(y|x_\sigma)`.
 
-    By default we have
+    You can either define the gradient :math:`\nabla_{x_\sigma} \log p_t(y|x_\sigma)` by overwriting the :meth:`grad` method, 
+    or you can define the noisy data fidelity function :math:`- \log p(y|x_\sigma)` by overwriting the :meth:`fn` method. 
+    In this case, the gradient will be computed automatically using automatic differentiation.
 
+    By default the class matches the :class:`deepinv.optim.DataFidelity` class, i.e. we make the approximation
+ 
     .. math::
 
         \begin{equation*}
-            \nabla_{x_t} \log p(y|x + \sigma(t) \omega) = P(\forw{x_t'}-y),
+            -\log p(y| x_\sigma) \approx d(A x, y)
         \end{equation*}
 
-
-    where :math:`P` is a preconditioner and :math:`x_t'` is an estimation of the image :math:`x`.
-    By default, :math:`P` is defined as :math:`A^\top`, :math:`x_t' = x_t` and this class matches the
-    :class:`deepinv.optim.DataFidelity` class.
-
-    :param deepinv.optim.Distance d: Distance metric to use for the data fidelity term. Default to :class:`deepinv.optim.L2Distance`.
-    :param float weight: Weighting factor for the data fidelity term. Default to 1.
+    By default, the distance function :math:`d` is the L2 distance.
     """
 
-    def __init__(self, d: Distance = None, weight=1.0, *args, **kwargs):
+    def __init__(self, *args, **kwargs):
         super().__init__()
-        if d is not None:
-            self.d = Distance(d)
-        else:
-            self.d = dinv.optim.L2Distance()
-        self.weight = weight
 
-    def precond(
-        self, u: torch.Tensor, physics: Physics, *args, **kwargs
-    ) -> torch.Tensor:
+    def fn(self, x, y, physics, sigma=None, *args, **kwargs):
         r"""
[…]
