+++ docs/source/refs.bib
@@ -929,7 +929,8 @@ journal = {ACM Trans. Graph.}
 @article{scanvic2025scale,
   title={Scale-Equivariant Imaging: Self-Supervised Learning for Image Super-Resolution and Deblurring},
   author={Scanvic, J{\'e}r{\'e}my and Davies, Mike and Abry, Patrice and Tachella, Juli{\'a}n},
-  year={2025}
+  year={2025},
+  journal={arXiv preprint arXiv:2312.11232},
 }
 
 @article{jin2017deep,
@@ -958,4 +959,11 @@ journal = {ACM Trans. Graph.}
   volume={35},
   pages={26565--26577},
   year={2022}
+}
+
+@inproceedings{zhang2018poisson,
+    title={A Poisson-Gaussian Denoising Dataset with Real Fluorescence Microscopy Images},
+    author={Yide Zhang and Yinhao Zhu and Evan Nichols and Qingfei Wang and Siyuan Zhang and Cody Smith and Scott Howard},
+    booktitle={CVPR},
+    year={2019}
 }
