+++ deepinv/sampling/noisy_datafidelity.py
@@ -5,7 +5,7 @@ from deepinv.physics import Physics
 from deepinv.models import Denoiser
 from typing import Union
 
-# Question : the base should have default value to automatic differetiation : i.e. not a default distance : so it should be a child of Potential, not DataFidelity ? 
+# Question : the base should have default value to automatic differetiation : i.e. not a default distance : so it should be a child of Potential, not DataFidelity ?
 
 
 class NoisyDataFidelity(Potential):
@@ -17,12 +17,12 @@ class NoisyDataFidelity(Potential):
 
     It comes with a `.grad` method for computing an approximation of the score :math:`\nabla_{x_\sigma} \log p_t(y|x_\sigma)`.
 
-    You can either define the gradient :math:`\nabla_{x_\sigma} \log p_t(y|x_\sigma)` by overwriting the :meth:`grad` method, 
-    or you can define the noisy data fidelity function :math:`- \log p(y|x_\sigma)` by overwriting the :meth:`fn` method. 
+    You can either define the gradient :math:`\nabla_{x_\sigma} \log p_t(y|x_\sigma)` by overwriting the :meth:`grad` method,
+    or you can define the noisy data fidelity function :math:`- \log p(y|x_\sigma)` by overwriting the :meth:`fn` method.
     In this case, the gradient will be computed automatically using automatic differentiation.
 
     By default the class matches the :class:`deepinv.optim.DataFidelity` class, i.e. we make the approximation
- 
+
     .. math::
 
         \begin{equation*}
@@ -42,23 +42,22 @@ class NoisyDataFidelity(Potential):
         :param torch.Tensor x: Variable :math:`x_\sigma` at which the data fidelity term is computed.
         :param torch.Tensor y: Data :math:`y`.
         :param deepinv.physics.Physics physics: physics model.
-        :param Union[torch.Tensor, float] sigma: Standard deviation :math:`\sigma` of the noise in :math:`x_\sigma`. 
+        :param Union[torch.Tensor, float] sigma: Standard deviation :math:`\sigma` of the noise in :math:`x_\sigma`.
         :return: (:class:`torch.Tensor`) data fidelity term.
         """
         return super().fn(x, y, physics, sigma, *args, **kwargs)
 
-    def grad(self, x, y, physics, sigma = None, *args, **kwargs):
+    def grad(self, x, y, physics, sigma=None, *args, **kwargs):
         r"""
-        Calculates the gradient of the noisy data fidelity term, approximating :math:`\nabla_{x_\sigma} \log p_t(y|x_\sigma)`. 
+        Calculates the gradient of the noisy data fidelity term, approximating :math:`\nabla_{x_\sigma} \log p_t(y|x_\sigma)`.
 
         :param torch.Tensor x: Variable :math:`x_\sigma` at which the gradient is computed.
         :param torch.Tensor y: Data :math:`y`.
         :param deepinv.physics.Physics physics: physics model.
-        :param Union[torch.Tensor, float] sigma: Standard deviation :math:`\sigma` of the noise in :math:`x_\sigma`. 
+        :param Union[torch.Tensor, float] sigma: Standard deviation :math:`\sigma` of the noise in :math:`x_\sigma`.
         :return: (:class:`torch.Tensor`) gradient :math:`\nabla_x \datafid{x}{y}`, computed in :math:`x`.
         """
         return super().grad(x, y, physics, sigma, *args, **kwargs)
-
 
 
 class L2(NoisyDataFidelity):
[…]
