+++ examples/unfolded/demo_unfolded_constrained_LISTA.py
@@ -1,45 +1,34 @@
 r"""
-Unfolded Chambolle-Pock for constrained image inpainting
+Learned Iterative Soft-Thresholding Algorithm (LISTA) for compressed sensing
 ====================================================================================================
 
-Image inpainting consists in solving :math:`y = Ax` where :math:`A` is a mask operator.
-This problem can be reformulated as the following minimization problem:
-
-.. math::
-
-    \begin{equation*}
-    \underset{x}{\operatorname{min}} \,\, \iota_{\mathcal{B}_2(y, r)}(Ax) + \regname(x)
-    \end{equation*}
-
-
-where :math:`\iota_{\mathcal{B}_2(y, r)}` is the indicator function of the ball of radius :math:`r` centered at
-:math:`y` for the :math:`\ell_2` norm, and :math:`\regname` is a regularisation. Recall that the indicator function of
-a convex set :math:`\mathcal{C}` is defined as :math:`\iota_{\mathcal{C}}(x) = 0` if :math:`x \in \mathcal{C}` and
-:math:`\iota_{\mathcal{C}}(x) = +\infty` otherwise.
-
-In this example, we unfold the Chambolle-Pock algorithm to solve this problem, and learn the thresholding parameters of
-a wavelet denoiser in a LISTA fashion.
+This example shows how to implement the `LISTA <http://yann.lecun.com/exdb/publis/pdf/gregor-icml-10.pdf>`_ algorithm
+for a compressed sensing problem. In a nutshell, LISTA is an unfolded proximal gradient algorithm involving a
+soft-thresholding proximal operator with learnable thresholding parameters.
 
 """
 
 from pathlib import Path
 import torch
-from torch.utils.data import DataLoader
+from torchvision import datasets
 from torchvision import transforms
+
 import deepinv as dinv
-from deepinv.utils.demo import load_dataset
-from deepinv.optim.data_fidelity import IndicatorL2
-from deepinv.optim.prior import PnP
-from deepinv.optim import PrimalDualCP
+from torch.utils.data import DataLoader
+from deepinv.optim.data_fidelity import L2
+from deepinv.unfolded import unfolded_builder
+from deepinv.utils.demo import get_data_home
 
 # %%
 # Setup paths for data loading and results.
-# --------------------------------------------
+# -----------------------------------------
 #
 
 BASE_DIR = Path(".")
 DATA_DIR = BASE_DIR / "measurements"
+RESULTS_DIR = BASE_DIR / "results"
 CKPT_DIR = BASE_DIR / "ckpts"
+ORIGINAL_DATA_DIR = get_data_home()
 
 # Set the global random seed from pytorch to ensure reproducibility of the example.
 torch.manual_seed(0)
@@ -48,71 +37,169 @@ device = dinv.utils.get_freer_gpu() if torch.cuda.is_available() else "cpu"
 
 # %%
 # Load base image datasets and degradation operators.
-# --------------------------------------------------------------------------------------------
-# In this example, we use the CBSD68 dataset for training and the set3c dataset for testing.
-# We work with images of size 32x32 if no GPU is available, else 128x128.
+# ----------------------------------------------------------------------------------------
+# In this example, we use MNIST as the base dataset.
 
+img_size = 28
+n_channels = 1
+operation = "compressed-sensing"
+train_dataset_name = "MNIST_train"
 
-operation = "inpainting"
-train_dataset_name = "CBSD68"
-test_dataset_name = "set3c"
-img_size = 128 if torch.cuda.is_available() else 32
-
-test_transform = transforms.Compose(
-    [transforms.CenterCrop(img_size), transforms.ToTensor()]
[…]
