+++ docs/source/community.rst
@@ -3,8 +3,8 @@
 Community
 =========
 
-Deepinv contributors includes researchers from multiple institutions around the world,
-such as the Univ. of Edinburgh, EPFL, Inria, ENS Lyon, ENS Paris, Univ. Toulouse, UCL, etc.
+DeepInverse contributors include researchers from multiple institutions around the world,
+such as the University of Edinburgh, EPFL, Inria, ENS Lyon, ENS Paris, Univ. Toulouse, UCL, etc.
 We organize hackathons and summer schools on an annual basis,
 and we welcome contributions from the community.
 
@@ -13,13 +13,12 @@ and we welcome contributions from the community.
    :width: 100%
    :alt: deepinverse Hackathon 2024 at CIRM, Marseille, France
 
-Future events
-~~~~~~~~~~~~~
-- June 4-6, 2025: Summer school tutorial at `International Symposium on Computational Sensing <https://www.iscs2025.com/>`_ in Luxembourg.
-- Feb 24-26 2025: Doctoral school tutorial at `ICMS <https://www.icms.org.uk/>`_ in Edinburgh, UK.
 
 Past events
 ~~~~~~~~~~~
+- June 4-6, 2025: Summer school tutorial at `International Symposium on Computational Sensing <https://www.iscs2025.com/>`_ in Luxembourg.
+- Apr 15-17, 2025: Software demo at `International Symposium on Biomedical Imaging <https://biomedicalimaging.org/2025/>`_ in Houston, Texas, USA.
+- Feb 24-26 2025: Doctoral school tutorial at `ICMS <https://www.icms.org.uk/>`_ in Edinburgh, UK.
 - Nov 26, 2024: the library obtained the documentation award at the `French Open Source Software for Science Awards <https://www.ouvrirlascience.fr/deepinverse>`_.
 - Oct 28-30, 2024: Hackathon at `CIRM <https://conferences.cirm-math.fr/3396.html>`_ in Marseille, France.
 - June 10-11, 2024: Summer school tutorial at `Mathematics and Machine Learning for Image Analysis <https://site.unibo.it/mathematical-ml-imaging/en/overview>`_ in Bologna, Italy.
