+++ MAINTAINERS.md
@@ -1,0 +1,52 @@
+# Maintainers
+
+**Contents**
+
+- [Current Maintainers](#current-maintainers)
+- [Emeritus](#emeritus)
+- [Responsibilities](#responsibilities)
+
+This document contains a list of maintainers in this repo. See [below](#responsibilities) for the maintainer responsibilities.
+
+If you're a contributor and are interested in becoming a maintainer, please reach out to one of us below!
+
+## Current Maintainers
+
+| Maintainer | Website | Affiliation | Email |
+|------|------|-------|-------|
+| Julián Tachella | [tachella.github.io](https://tachella.github.io/) | CNRS, ENS Lyon | [📧](mailto:julian.tachella@cnrs.fr) |
+| Matthieu Terris | [matthieutrs.github.io](https://matthieutrs.github.io) | Université Paris-Saclay, Inria, CEA | [📧](mailto:matthieu.terris@gmail.com) |
+| Samuel Hurault | [github.com/samuro95](https://github.com/samuro95/) | CNRS, ENS Paris | [📧]() |
+| Andrew Wang | [andrewwango.github.io](https://andrewwango.github.io/about) | University of Edinburgh | [📧](mailto:andrew.wang@ed.ac.uk) |
+| Dongdong Chen | [dongdongchen.com](http://dongdongchen.com/) | Heriot-Watt University | [📧](mailto:d.chen@hw.ac.uk) |
+| Minh-Hai Nguyen | [mh-nguyen712.github.io](https://mh-nguyen712.github.io/) | IRIT, CBI, CNRS, Université de Toulouse | [📧](mailto:nguyenhai7120qh@gmail.com) |
+| Jérémy Scanvic | [jeremyscanvic.com](https://jeremyscanvic.com/) | CNRS, ENS Lyon | [📧]() |
+
+## Emeritus
+
+| Maintainer | Website | Affiliation | Email |
+|------|------|-------|-------|
+
+## Responsibilities
+
+Maintainers are active and visible members of the community, and have [maintain-level permissions on a repository](https://docs.github.com/en/organizations/managing-access-to-your-organizations-repositories/repository-permission-levels-for-an-organization). Maintainers serve the community using this privilege, by having the below day-to-day responsibilities. Read more maintainer best practices [here](https://opensource.guide/best-practices/).
+
+- **Uphold code of conduct**: model the behavior set forward by the [code of conduct](CODE_OF_CONDUCT.md) and raise any violations to other maintainers.
+
+- **Prioritise security**: security is your number one priority. Maintainer's GitHub keys must be password protected securely and any reported security vulnerabilities are addressed before features or bugs, following the [security guidelines](SECURITY.md).
+
+- **Pull requests**: ensure PRs adhere to our [contributing guidelines](https://deepinv.github.io/deepinv/contributing.html) and have completed the checklist in the [PR template](.github/pull_request_template.md). Triage the PR and assign them to maintainers for review. Provide actionable feedback to guide the PR towards a conclusion. In cases of uncertainty, or where contributed code is out of your domain of expertise, reach out to additional maintainers or contributors.
+
+- **Issues**: manage labels, review issues regularly, and triage by labelling them, e.g. `bug`, `documentation`, `duplicate`, `enhancement`, `good first issue`, `help wanted`, `blocker`, `invalid`, `question`, `wontfix`, and `untriaged`. Request further information from a submitter if an issue is not clear. See the [issue template](.github/ISSUE_TEMPLATE.md) for an example.
+
+- **Responsiveness**: promptly respond to [issues](https://github.com/deepinv/deepinv/issues) and [discussions](https://github.com/deepinv/deepinv/discussions). Allocate time for reviewing, commenting and continuing conversations.
+
[…]

[…]
