### 2025-06-25 feb1c2d
add philosophy to maintainers
