+++ README.rst
@@ -115,22 +115,6 @@ Maintainers
 
 View our current and former maintainers, and how to get in touch, at [MAINTAINERS](https://github.com/deepinv/deepinv/blob/main/MAINTAINERS.md).
 
-.. |Black| image:: https://img.shields.io/badge/code%20style-black-000000.svg
-    :target: https://github.com/psf/black
-.. |Test Status| image:: https://github.com/deepinv/deepinv/actions/workflows/test.yml/badge.svg
-   :target: https://github.com/deepinv/deepinv/actions/workflows/test.yml
-.. |Docs Status| image:: https://github.com/deepinv/deepinv/actions/workflows/documentation.yml/badge.svg
-   :target: https://github.com/deepinv/deepinv/actions/workflows/documentation.yml
-.. |Python Version| image:: https://img.shields.io/badge/python-3.9%2B-blue
-   :target: https://www.python.org/downloads/release/python-390/
-.. |codecov| image:: https://codecov.io/gh/deepinv/deepinv/branch/main/graph/badge.svg?token=77JRvUhQzh
-   :target: https://codecov.io/gh/deepinv/deepinv
-.. |discord| image:: https://dcbadge.vercel.app/api/server/qBqY5jKw3p?style=flat
-   :target: https://discord.gg/qBqY5jKw3p
-.. |colab| image:: https://colab.research.google.com/assets/colab-badge.svg
-   :target: https://colab.research.google.com/drive/1XhCO5S1dYN3eKm4NEkczzVU7ZLBuE42J
-
-
 Citation
 --------
 If you use DeepInverse in your research, please cite the following paper (available on `arXiv <https://arxiv.org/abs/2505.20160>`_):
@@ -147,3 +131,19 @@ If you use DeepInverse in your research, please cite the following paper (availa
           primaryClass={eess.IV},
           url={https://arxiv.org/abs/2505.20160},
     }
+
+
+.. |Black| image:: https://img.shields.io/badge/code%20style-black-000000.svg
+    :target: https://github.com/psf/black
+.. |Test Status| image:: https://github.com/deepinv/deepinv/actions/workflows/test.yml/badge.svg
+   :target: https://github.com/deepinv/deepinv/actions/workflows/test.yml
+.. |Docs Status| image:: https://github.com/deepinv/deepinv/actions/workflows/documentation.yml/badge.svg
+   :target: https://github.com/deepinv/deepinv/actions/workflows/documentation.yml
+.. |Python Version| image:: https://img.shields.io/badge/python-3.9%2B-blue
+   :target: https://www.python.org/downloads/release/python-390/
+.. |codecov| image:: https://codecov.io/gh/deepinv/deepinv/branch/main/graph/badge.svg?token=77JRvUhQzh
+   :target: https://codecov.io/gh/deepinv/deepinv
+.. |discord| image:: https://dcbadge.vercel.app/api/server/qBqY5jKw3p?style=flat
+   :target: https://discord.gg/qBqY5jKw3p
+.. |colab| image:: https://colab.research.google.com/assets/colab-badge.svg
+   :target: https://colab.research.google.com/drive/1XhCO5S1dYN3eKm4NEkczzVU7ZLBuE42J
