### 2025-06-25 c232937
Add Physics.clone (#534)
* add Physics.clone

* get rid of dummy variable

* improve the comments

* get rid of bleeding-edge function

* add additional tests

* improve docstrings

* replace deepcopy by Physics.clone

* put __future__ back up

* add a cloned physics to the test list

* add tests for requires_grad

* test autograd

* trigger ci

* add a dummy parameter to hit coverage

* test attributes other than params and bufffers
