### 2025-06-25 1dfee87
add start history and views
