### 2025-06-25 63b872f
installation counts
