### 2025-06-25 79e0902
Test Discord integration (#567)
