+++ README.rst
@@ -125,7 +125,7 @@ If you have any questions or suggestions, please join the conversation in our
 .. |colab| image:: https://colab.research.google.com/assets/colab-badge.svg
    :target: https://colab.research.google.com/drive/1XhCO5S1dYN3eKm4NEkczzVU7ZLBuE42J
 
-
+ 
 Citation
 --------
 If you use DeepInverse in your research, please cite the following paper (available on `arXiv <https://arxiv.org/abs/2505.20160>`_):
