+++ CHANGELOG.rst
@@ -22,14 +22,10 @@ Fixed
 ^^^^^
 - Fix the total loss reported by the trainer (:gh:`515` by `Jérémy Scanvic`_)
 - Fix the gradient norm reported by the trainer (:gh:`520` by `Jérémy Scanvic`_)
-<<<<<<< tomography
 - Fix that the max_pixel option in PSNR and SSIM and add analgous min_pixel option (:gh:`535` by `Johannes Hertrich`_)
-
-
-=======
 - Fix some issues related to denoisers: ICNN grad not working inside torch.no_grad(), batch of image and batch of sigma for some denoisers (DiffUNet, BM3D, TV, Wavemet), EPLL error when batch size > 1 (:gh:`530` by `Minh Hai Nguyen`_)  
 - Batching WaveletPrior and fix iwt(:gh:`530` by `Minh Hai Nguyen`_)  
->>>>>>> main
+
 v0.3.1
 ----------------
 
