### 2025-06-25 bd4050d
fix test_trainer_physics_generator_params
