### 2025-06-25 4efef3e
fix test_trainer_multidatasets
