### 2025-06-25 f735d63
fix test_trainer_identity
