+++ deepinv/tests/test_trainer.py
@@ -364,6 +364,7 @@ def test_trainer_identity(imsize, rng, device):
             return self.dummy_param * y
 
     dummy_model = DummyModel()
+    dummy_model.to(device)
     optimizer = torch.optim.Adam(dummy_model.parameters(), lr=1e-2, weight_decay=0.0)
 
     trainer = Trainer(
