### 2025-06-25 b2504f2
fix test_metrics
