+++ deepinv/tests/test_loss.py
@@ -99,7 +99,7 @@ def test_jacobian_spectral_values(toymatrix, reduction):
     assert torch.allclose(regfnel2, reg_fne_target, rtol=1e-3)
 
 
-def choose_loss(loss_name, rng=None):
+def choose_loss(loss_name, rng=None, device="cpu"):
     loss = []
     if loss_name == "mcei":
         loss.append(dinv.loss.MCLoss())
@@ -114,7 +114,7 @@ def choose_loss(loss_name, rng=None):
             "installed with `pip install kornia`",
         )
         loss.append(dinv.loss.MCLoss())
-        loss.append(dinv.loss.EILoss(dinv.transform.Homography()))
+        loss.append(dinv.loss.EILoss(dinv.transform.Homography(device=device)))
     elif loss_name == "splittv":
         loss.append(dinv.loss.SplittingLoss(split_ratio=0.25))
         loss.append(dinv.loss.TVLoss())
@@ -290,7 +290,7 @@ def test_notraining(physics, tmp_path, imsize, device):
 @pytest.mark.parametrize("loss_name", LOSSES)
 def test_losses(loss_name, tmp_path, dataset, physics, imsize, device, rng):
     # choose training losses
-    loss = choose_loss(loss_name, rng)
+    loss = choose_loss(loss_name, rng, device=device)
 
     save_dir = tmp_path / "dataset"
     # choose backbone denoiser
@@ -432,6 +432,7 @@ def test_measplit(device, loss_name, rng):
             dinv.physics.generator.BernoulliSplittingMaskGenerator(
                 imsize, 0.5, device=device, rng=rng
             ),
+            device=device,
         )
         loss = dinv.loss.mri.WeightedSplittingLoss(
             mask_generator=gen, physics_generator=physics.gen
@@ -442,6 +443,7 @@ def test_measplit(device, loss_name, rng):
             dinv.physics.generator.BernoulliSplittingMaskGenerator(
                 imsize, 0.5, device=device, rng=rng
             ),
+            device=device,
         )
         loss = dinv.loss.mri.RobustSplittingLoss(
             mask_generator=gen,
