### 2025-06-25 a385ada
fix test_wavelet_decomposition
