+++ deepinv/utils/nn.py
@@ -4,6 +4,14 @@ from warnings import warn
 import numpy as np
 import torch
 
+_warn_msg_cuda_order = (
+    "Your attempt to choose GPU was not reliable because the environment variable "
+    "'CUDA_DEVICE_ORDER' either is not set or not equal to 'PCI_BUS_ID' (see also the doc). "
+    "Before loading Pytorch you can set the variable in your shell using the command: "
+    "'export CUDA_DEVICE_ORDER=PCI_BUS_ID' "
+    "or use IPython's 'magic' command '%env CUDA_DEVICE_ORDER=PCI_BUS_ID'."
+)
+
 
 def get_freer_gpu(verbose=True):
     """
@@ -13,7 +21,14 @@ def get_freer_gpu(verbose=True):
     Attempts to use ``nvidia-smi`` with ``bash``, if these don't exist then uses torch commands to get free memory.
 
     :param bool verbose: print selected GPU index and memory
-    :return torch.device device: selected torch cuda device.
+    :return torch.device device: selected torch cuda device
+
+    .. warning::
+        GPU indices in ``nvidia-smi`` may not match those in Pytorch if in your environment ``CUDA_DEVICE_ORDER``
+        is not set to ``PCI_BUS_ID``. If the above variable is not set, the call will generate a warning but will
+        still return a device. We strongly recommend setting this variable as above prior to loading Pytorch's
+        runtime API.
+
     """
     try:
         pipeline = "nvidia-smi -q -d Memory | grep -A5 GPU | grep Free"
@@ -29,6 +44,9 @@ def get_freer_gpu(verbose=True):
         idx, mem = np.argmax(memory_available), np.max(memory_available)
         device = torch.device(f"cuda:{idx}")
 
+        env = os.environ.copy()
+        if "CUDA_DEVICE_ORDER" not in env or env["CUDA_DEVICE_ORDER"] != "PCI_BUS_ID":
+            warn(_warn_msg_cuda_order)
     except:
         if torch.cuda.device_count() == 0:
             warn("Couldn't find free GPU")
