+++ README.rst
@@ -4,7 +4,7 @@
    :align: center
 
 
-|Test Status| |Docs Status| |Python Version| |Black| |codecov| |views| |install| |discord| |colab|   
+|Test Status| |Docs Status| |Python Version| |Black| |codecov| |views| |pip install| |discord| |colab|   
 
 
 Introduction
@@ -121,7 +121,7 @@ If you have any questions or suggestions, please join the conversation in our
 .. |codecov| image:: https://codecov.io/gh/deepinv/deepinv/branch/main/graph/badge.svg?token=77JRvUhQzh
    :target: https://codecov.io/gh/deepinv/deepinv
 .. |views| image:: https://komarev.com/ghpvc/?username=deepinv
-.. |install| image:: https://img.shields.io/pypi/dm/deepinv.svg?logo=pypi&label=install&color=fedcba
+.. |pip install| image:: https://img.shields.io/pypi/dm/deepinv.svg?logo=pypi&label=pip%20install&color=fedcba
    :target: https://pypistats.org/packages/deepinv
 .. |discord| image:: https://dcbadge.vercel.app/api/server/qBqY5jKw3p?style=flat
    :target: https://discord.gg/qBqY5jKw3p
