### 2025-06-26 997ad82
fix test_noise_model
