+++ deepinv/physics/noise.py
@@ -519,10 +519,23 @@ class PoissonNoise(NoiseModel):
         self.to(x.device)
         gain = self.gain[(...,) + (None,) * (x.dim() - 1)]
 
-        y = torch.poisson(
-            torch.clip(x / gain, min=0.0) if self.clip_positive else x / gain,
-            generator=self.rng,
-        )
+        if self.clip_positive:
+            z = torch.clip(x / gain, min=0.0)
+        else:
+            # NOTE: PyTorch operations are generally run asynchronously on CUDA
+            # devices and the underlying CUDA kernel under
+            # torch.poisson typically raises a CUDA-level assertion error
+            # when its input has negative entries. Those errors can't be
+            # recovered from using Python's exception system due to their
+            # asynchronous nature. For this reason we add a manual check if the
+            # RNG is on a CUDA device.
+            if self.rng.device.type == "cuda":
+                assert gain > 0, "Gain must be positive"
+                assert torch.all(x >= 0), "Input tensor must be non-negative"
+
+            z = x / gain
+
+        y = torch.poisson(z, generator=self.rng)
         if self.normalize:
             y = y * gain
         return y
@@ -626,6 +639,17 @@ class PoissonGaussianNoise(NoiseModel):
         if self.clip_positive:
             y = torch.poisson(torch.clip(x / gain, min=0.0), generator=self.rng) * gain
         else:
+            # NOTE: PyTorch operations are generally run asynchronously on CUDA
+            # devices and the underlying CUDA kernel under
+            # torch.poisson typically raises a CUDA-level assertion error
+            # when its input has negative entries. Those errors can't be
+            # recovered from using Python's exception system due to their
+            # asynchronous nature. For this reason we add a manual check if the
+            # RNG is on a CUDA device.
+            if self.rng.device.type == "cuda":
+                assert gain > 0, "Gain must be positive"
+                assert torch.all(x >= 0), "Input tensor must be non-negative"
+
             y = torch.poisson(x / gain, generator=self.rng) * gain
 
         y = y + self.randn_like(x) * sigma
