### 2025-06-26 f67bbf3
fix None-related bug
