+++ deepinv/physics/noise.py
@@ -529,7 +529,7 @@ class PoissonNoise(NoiseModel):
             # recovered from using Python's exception system due to their
             # asynchronous nature. For this reason we add a manual check if the
             # RNG is on a CUDA device.
-            if self.rng.device.type == "cuda":
+            if self.rng is not None and self.rng.device.type == "cuda":
                 assert gain > 0, "Gain must be positive"
                 assert torch.all(x >= 0), "Input tensor must be non-negative"
 
@@ -646,7 +646,7 @@ class PoissonGaussianNoise(NoiseModel):
             # recovered from using Python's exception system due to their
             # asynchronous nature. For this reason we add a manual check if the
             # RNG is on a CUDA device.
-            if self.rng.device.type == "cuda":
+            if self.rng is not None and self.rng.device.type == "cuda":
                 assert gain > 0, "Gain must be positive"
                 assert torch.all(x >= 0), "Input tensor must be non-negative"
 
