### 2025-06-26 9ff3188
added str literal, changed doc sentences; changed set/no set behaviour
