### 2025-06-27 145449e
Update MAINTAINERS.md
Co-authored-by: Julián Tachella <38355592+tachella@users.noreply.github.com>
