### 2025-06-27 18c7067
Add an error on trainer directory collisions instead of overwriting files silently (#550)
* create pr

* make the trainer fail when a collision occurs

* black

* fix the mocking

* restore mistakenly modified test name

* black

* use unique directory paths for testing

* black

* replace custom fixture by standard fixture
