+++ deepinv/tests/test_trainer.py
@@ -4,6 +4,7 @@ import torch
 from torch.utils.data import DataLoader, Dataset
 
 import deepinv as dinv
+from deepinv.utils import get_timestamp
 from dummy import DummyCircles, DummyModel
 from deepinv.training.trainer import Trainer
 from deepinv.physics.generator.base import PhysicsGenerator
@@ -16,6 +17,8 @@ import builtins
 import io
 import contextlib
 import re
+import tempfile
+import os
 
 from conftest import no_plot
 
@@ -40,7 +43,7 @@ def physics(imsize, device):
 
 
 @pytest.mark.parametrize("no_learning", NO_LEARNING)
-def test_nolearning(imsize, physics, model, no_learning, device):
+def test_nolearning(imsize, physics, model, no_learning, device, tmpdir):
     y = torch.ones((1,) + imsize, device=device)
     trainer = dinv.Trainer(
         model=model,
@@ -50,6 +53,7 @@ def test_nolearning(imsize, physics, model, no_learning, device):
         physics=physics,
         compare_no_learning=True,
         no_learning_method=no_learning,
+        save_path=tmpdir,
     )
     x_hat = trainer.no_learning_inference(y, physics)
     assert (physics.A(x_hat) - y).pow(2).mean() < 0.1
@@ -125,6 +129,7 @@ def test_get_samples(
     use_physics_generator,
     online_measurements,
     rng,
+    tmpdir,
 ):
     # Dummy constant GT dataset
     class DummyDataset(Dataset):
@@ -216,6 +221,7 @@ def test_get_samples(
             if online_measurements and physics_generator is not None
             else None
         ),
+        save_path=tmpdir,
     )
 
     iterator = iter(dataloader)
@@ -531,6 +537,7 @@ def test_dataloader_formats(
     measurements,
     online_measurements,
     rng,
+    tmpdir,
 ):
     """Test dataloader return formats
 
@@ -605,6 +612,7 @@ def test_dataloader_formats(
         online_measurements=online_measurements,
         train_dataloader=dataloader,
         optimizer=optimizer,
+        save_path=tmpdir,
     )
     trainer.setup_train()
     x, y, physics = trainer.get_samples([iter(dataloader)], 0)
@@ -657,7 +665,7 @@ def test_dataloader_formats(
 @pytest.mark.parametrize("early_stop", [True, False])
 @pytest.mark.parametrize("max_batch_steps", [3, 100000])
 def test_early_stop(
-    dummy_dataset, imsize, device, dummy_model, early_stop, max_batch_steps
+    dummy_dataset, imsize, device, dummy_model, early_stop, max_batch_steps, tmpdir
 ):
     torch.manual_seed(0)
     model = dummy_model
@@ -682,6 +690,7 @@ def test_early_stop(
         optimizer=optimizer,
         verbose=False,
         plot_images=True,
+        save_path=tmpdir,
     )
     with no_plot():
         trainer.train()
@@ -711,7 +720,7 @@ class ConstantLoss(dinv.loss.Loss):
         )
 
 
-def test_total_loss(dummy_dataset, imsize, device, dummy_model):
+def test_total_loss(dummy_dataset, imsize, device, dummy_model, tmpdir):
     train_data, eval_data = dummy_dataset, dummy_dataset
     dataloader = DataLoader(train_data, batch_size=2)
     eval_dataloader = DataLoader(eval_data, batch_size=2)
@@ -732,6 +741,7 @@ def test_total_loss(dummy_dataset, imsize, device, dummy_model):
         optimizer=torch.optim.AdamW(dummy_model.parameters(), lr=1),
         verbose=False,
         online_measurements=True,
+        save_path=tmpdir,
     )
[…]
