### 2025-06-27 812706c
add summer school
