+++ deepinv/physics/blur.py
@@ -132,6 +132,8 @@ class Downsampling(LinearPhysics):
 
         Otherwise, it computes it using the conjugate gradient algorithm which can be slow if applied many times.
 
+        |sep|
+        
         :References:
             .. footbibliography::
         """
