+++ deepinv/models/swinir.py
@@ -825,6 +825,8 @@ class SwinIR(Denoiser):
     :param int pretrained_noise_level: The noise level of the pretrained model to be downloaded (in 0-255 scale). This
         value is directly concatenated to the download url; should be chosen in the set {15, 25, 50}. Default: 15.
 
+    |sep|
+    
     :References:
     .. footbibliography::
     """
