+++ deepinv/sampling/utils.py
@@ -7,7 +7,7 @@ class Welford:
     Welford's algorithm :footcite:ps:`welford1962note`for calculating mean and variance.
 
     |sep|
-    
+
     :References:
         .. footbibliography::
     """
