+++ deepinv/datasets/lidc_idri.py
@@ -73,7 +73,7 @@ class LidcIdriSliceDataset(torch.utils.data.Dataset):
             batch = next(iter(dataloader))
             print(batch.shape)
 
-            
+
     :References:
     .. footbibliography::
     """
