+++ deepinv/models/drunet.py
@@ -38,7 +38,7 @@ class DRUNet(Denoiser):
      :param str device: gpu or cpu.
 
     |sep|
-    
+
     :References:
      .. footbibliography::
     """
