### 2025-06-27 ecbfd23
add star history and views (#566)
* add start history and views

* installation counts

* change color

* add to index

* rm view

---------

Co-authored-by: mh-nguyen712 <minh-hai.nguyen@irit.fr>
