### 2025-06-27 cc19a04
revert old noisy data fid
