+++ docs/source/community.rst
@@ -17,10 +17,10 @@ and we welcome contributions from the community.
 Past events
 ~~~~~~~~~~~
 
-- June 9-10, 2025: Summer school tutorial at `Mathematics and Machine Learning for Image Analysis <https://site.unibo.it/mml-imaging/en>`_ in Bologna, Italy.
+- June 9-10, 2025: Summer school tutorial at `Mathematics and Machine Learning for Image Analysis <https://site.unibo.it/mml-imaging/en>`__ in Bologna, Italy.
 - June 4-6, 2025: Summer school tutorial at `International Symposium on Computational Sensing <https://www.iscs2025.com/>`_ in Luxembourg.
 - Apr 15-17, 2025: Software demo at `International Symposium on Biomedical Imaging <https://biomedicalimaging.org/2025/>`_ in Houston, Texas, USA.
 - Feb 24-26 2025: Doctoral school tutorial at `ICMS <https://www.icms.org.uk/>`_ in Edinburgh, UK.
 - Nov 26, 2024: the library obtained the documentation award at the `French Open Source Software for Science Awards <https://www.ouvrirlascience.fr/deepinverse>`_.
 - Oct 28-30, 2024: Hackathon at `CIRM <https://conferences.cirm-math.fr/3396.html>`_ in Marseille, France.
-- June 10-11, 2024: Summer school tutorial at `Mathematics and Machine Learning for Image Analysis <https://site.unibo.it/mathematical-ml-imaging/en/overview>`_ in Bologna, Italy.
+- June 10-11, 2024: Summer school tutorial at `Mathematics and Machine Learning for Image Analysis <https://site.unibo.it/mathematical-ml-imaging/en/overview>`__ in Bologna, Italy.
