### 2025-06-27 f881df2
Backward Paths for Tomography (#535)
* backward paths for tomography

* fix

* fix

* fix

* fix

* fix

* docs + FBP boundary handling

* fix dynamic range PSNR for batches

* fix amin/min and amax/max

* update docs

* tests and changelog

* parametrize test_tomography

* changes

* black

* black and set fbp_interpolate_boundary to False if circle is True

* changes

* Apply suggestions from code review

Co-authored-by: Julián Tachella <38355592+tachella@users.noreply.github.com>

* black

* update docstrings

* fix docstring formatting

* add adjointness test for exact adjoint

* fix merge
[…]
