### 2025-06-27 7b4b63e
add meeting responsibility
