+++ examples/basics/demo_microscopy_3d.py
@@ -215,4 +215,5 @@ plot_ortho3D(
 
 # %%
 # :References:
+# 
 #    .. footbibliography::
