+++ examples/plug-and-play/demo_PnP_DPIR_deblur.py
@@ -161,4 +161,5 @@ test(
 
 # %%
 # :References:
+# 
 #    .. footbibliography::
