+++ examples/optimization/demo_TV_minimisation.py
@@ -190,4 +190,5 @@ if plot_convergence_metrics:
 
 # %%
 # :References:
+# 
 #    .. footbibliography::
