+++ examples/basics/demo_custom_prior.py
@@ -191,4 +191,5 @@ test(
 
 # %%
 # :References:
+# 
 #    .. footbibliography::
