+++ examples/unfolded/demo_LISTA.py
@@ -274,4 +274,5 @@ dinv.utils.plotting.plot_parameters(
 
 # %%
 # :References:
+# 
 #    .. footbibliography::
