+++ examples/self-supervised-learning/demo_equivariant_imaging.py
@@ -200,5 +200,5 @@ trainer.test(test_dataloader)
 
 # %%
 # :References:
-# 
+#
 #    .. footbibliography::
