+++ examples/self-supervised-learning/demo_artifact2artifact.py
@@ -245,5 +245,5 @@ trainer.test(test_dataloader)
 
 # %%
 # :References:
-# 
+#
 #    .. footbibliography::
