+++ examples/sampling/demo_ddrm.py
@@ -147,5 +147,5 @@ plot(
 
 # %%
 # :References:
-# 
+#
 #    .. footbibliography::
