+++ examples/patch-priors/demo_epll.py
@@ -102,5 +102,5 @@ plot(
 
 # %%
 # :References:
-# 
+#
 #    .. footbibliography::
