+++ examples/self-supervised-learning/demo_multioperator_imaging.py
@@ -206,5 +206,5 @@ trainer.test(test_dataloader)
 
 # %%
 # :References:
-# 
+#
 #    .. footbibliography::
