+++ examples/basics/demo_tour_mri.py
@@ -490,5 +490,5 @@ dinv.utils.plot(
 
 # %%
 # :References:
-# 
+#
 #    .. footbibliography::
