+++ examples/self-supervised-learning/demo_unsure.py
@@ -199,5 +199,5 @@ trainer.test(test_dataloader=test_dataloader)
 
 # %%
 # :References:
-# 
+#
 #    .. footbibliography::
