+++ examples/optimization/demo_3D_wavelets.py
@@ -253,5 +253,5 @@ dinv.utils.plot(
 
 # %%
 # :References:
-# 
+#
 #    .. footbibliography::
