### 2025-06-27 3c709a8
add auto_release
