+++ .github/scripts/get_version.py
@@ -1,0 +1,7 @@
+try:
+    import tomllib
+except ImportError:
+    import tomli as tomllib
+    
+data = tomllib.load("pyproject.toml")
+print(data["project"]["version"])
