+++ .github/scripts/get_version.py
@@ -2,6 +2,6 @@ try:
     import tomllib
 except ImportError:
     import tomli as tomllib
-    
+
 data = tomllib.load("pyproject.toml")
 print(data["project"]["version"])
