### 2025-06-27 d7c9565
copilot fix
