### 2025-06-27 c5cc2a8
fix get version
