### 2025-06-28 726b0e5
fix maintainer link + pad library schematic
