### 2025-06-28 a3842c4
Update noise.py
