### 2025-06-29 b4d6920
Update deepinv/__about__.py
Co-authored-by: Andrewwango <andrew.wang27@gmail.com>
