### 2025-06-29 140b4ed
Update docs/source/conf.py
Co-authored-by: Andrewwango <andrew.wang27@gmail.com>
