+++ .github/scripts/get_version.py
@@ -3,7 +3,12 @@ try:
 except ImportError:
     import tomli as tomllib
 
-with open("pyproject.toml", "rb") as f:
-    metadata = tomllib.load(f)["project"]
 
-print(metadata["version"])
+def get_version_from_toml(filename):
+    with open(filename, "rb") as f:
+        metadata = tomllib.load(f)["project"]
+    return metadata["version"]
+
+if __name__ == "__main__":
+    version = get_version_from_toml("../../pyproject.toml")
+    print(version)
