+++ _downloads/dc2bd81d16a856e0c8854be186e39b9b/demo_spc.py
@@ -1,0 +1,223 @@
+r"""
+Pattern Ordering in a Compressive Single Pixel Camera
+========================================================
+
+This demo illustrates the impact of different Hadamard pattern ordering algorithms in the Single Pixel Camera (SPC),
+a computational imaging system that uses a single photodetector to capture images by projecting the
+scene onto a series of patterns. The SPC is implemented in the DeepInverse library.
+
+In this example, we explore the following ordering algorithms for the Hadamard transform:
+
+1. **Sequency Ordering**:
+   Rows are ordered based on the number of sign changes (sequency pattern).
+   Reference: https://en.wikipedia.org/wiki/Walsh_matrix#Sequency_ordering
+
+2. **Cake Cutting Ordering**:
+   Rows are ordered based on the number of blocks in the 2D resized Hadamard matrix.
+   Reference: https://doi.org/10.3390/s19194122
+
+3. **Zig-Zag Ordering**:
+   Rows are ordered in a zig-zag pattern from the top-left to the bottom-right of the matrix.
+   Reference: https://doi.org/10.1364/OE.451656
+
+4. **XY Ordering**:
+   Rows are ordered in a circular pattern starting from the top-left to the bottom-right.
+   Reference: https://doi.org/10.48550/arXiv.2209.04449
+
+We compare the reconstructions obtained using these orderings and visualize their corresponding masks
+and Hadamard spectrum.
+
+"""
+
+import torch
+from pathlib import Path
+import matplotlib.pyplot as plt
+import deepinv as dinv
+from deepinv.utils.demo import get_image_url, load_url_image
+from deepinv.utils.plotting import plot
+from deepinv.loss.metric import PSNR
+from deepinv.physics.singlepixel import hadamard_2d_shift
+
+# %%
+# General Setup
+# -------------
+BASE_DIR = Path(".")
+RESULTS_DIR = BASE_DIR / "results"
+
+# Set the global random seed for reproducibility.
+torch.manual_seed(0)
+device = dinv.utils.get_freer_gpu() if torch.cuda.is_available() else "cpu"
+
+# %%
+# Configuration
+# -------------
+# Define the image size, noise level, number of measurements, and ordering algorithms.
+img_size_H = 128  # Size of the image (128x128 pixels)
+img_size_W = 128
+noise_level_img = 0.0  # Noise level in the image
+n = img_size_H * img_size_W  # Total number of pixels in the image
+m = 5000  # Number of measurements
+orderings = [
+    "sequency",
+    "cake_cutting",
+    "zig_zag",
+    "xy",
+]  # Ordering algorithms to compare
+
+# %%
+# Load Image
+# ----------
+# Load a grayscale image from the internet and resize it to the desired size.
+url = get_image_url("barbara.jpeg")
+x = load_url_image(
+    url=url,
+    img_size=(img_size_H, img_size_W),
+    grayscale=True,
+    resize_mode="resize",
+    device=device,
+)
+
+# %%
+# Create Physics Models
+# ---------------------
+# Instantiate Single Pixel Camera models with different ordering algorithms.
+physics_list = [
+    dinv.physics.SinglePixelCamera(
+        m=m,
+        img_size=(1, img_size_H, img_size_W),
+        noise_model=dinv.physics.GaussianNoise(sigma=noise_level_img),
+        device=device,
+        ordering=ordering,
+    )
+    for ordering in orderings
+]
+
+# %%
+# Generate Measurements and Reconstructions
+# -----------------------------------------
[…]
