+++ _downloads/2a714a8332e6f0a73ab837f9f0817118/demo_tour_mri.py
@@ -1,0 +1,491 @@
+r"""
+Tour of MRI functionality in DeepInverse
+========================================
+
+This example presents the various datasets, forward physics and models
+available in DeepInverse for Magnetic Resonance Imaging (MRI) problems:
+
+-  Physics: :class:`deepinv.physics.MRI`,
+   :class:`deepinv.physics.MultiCoilMRI`,
+   :class:`deepinv.physics.DynamicMRI`
+-  Datasets: the full `FastMRI <https://fastmri.med.nyu.edu>`__ dataset
+   :class:`deepinv.datasets.FastMRISliceDataset` and a lightweight,
+   easy-to-use subset
+   :class:`deepinv.datasets.SimpleFastMRISliceDataset`
+-  Models: :class:`deepinv.models.VarNet`
+   (`VarNet <https://onlinelibrary.wiley.com/doi/full/10.1002/mrm.26977>`__/`E2E-VarNet <https://arxiv.org/abs/2004.06688>`__),
+   :class:`deepinv.models.MoDL` (a simple
+   `MoDL <https://ieeexplore.ieee.org/document/8434321>`__ unrolled
+   model)
+
+Contents:
+
+1. Get started with FastMRI (singlecoil + multicoil)
+2. Train an accelerated MRI with neural networks
+3. Load raw FastMRI data (singlecoil + multicoil)
+4. Train using raw data
+5. Explore 3D MRI
+6. Explore dynamic MRI
+
+"""
+
+import deepinv as dinv
+import torch, torchvision
+from torch.utils.data import DataLoader
+
+device = dinv.utils.get_freer_gpu() if torch.cuda.is_available() else "cpu"
+rng = torch.Generator(device=device).manual_seed(0)
+
+
+# %%
+# 1. Get started with FastMRI
+# ~~~~~~~~~~~~~~~~~~~~~~~~~~~
+#
+# You can get started with our simple
+# `FastMRI <https://fastmri.med.nyu.edu>`__ mini slice subsets which provide
+# quick, easy-to-use, in-memory datasets which can be used for simulation
+# experiments.
+#
+# .. important::
+#
+#    By using this dataset, you confirm that you have agreed to and signed the `FastMRI data use agreement <https://fastmri.med.nyu.edu/>`_.
+#
+# .. seealso::
+#
+#   Datasets :class:`deepinv.datasets.FastMRISliceDataset` :class:`deepinv.datasets.SimpleFastMRISliceDataset`
+#       We provide convenient datasets to easily load both raw and reconstructed FastMRI images.
+#       You can download more data on the `FastMRI site <https://fastmri.med.nyu.edu/>`_.
+#
+# Load mini demo knee and brain datasets (original data is 320x320 but we resize to
+# 128 for speed):
+#
+
+transform = torchvision.transforms.Resize(128)
+knee_dataset = dinv.datasets.SimpleFastMRISliceDataset(
+    dinv.utils.get_data_home(),
+    anatomy="knee",
+    transform=transform,
+    train=True,
+    download=True,
+)
+brain_dataset = dinv.datasets.SimpleFastMRISliceDataset(
+    dinv.utils.get_data_home(),
+    anatomy="brain",
+    transform=transform,
+    train=True,
+    download=True,
+)
+
+img_size = knee_dataset[0].shape[-2:]  # (128, 128)
+dinv.utils.plot({"knee": knee_dataset[0], "brain": brain_dataset[0]})
+
+
+# %%
+# Let's start with single-coil MRI. We can define a constant Cartesian 4x
+# undersampling mask by sampling once from a physics generator. The mask,
+# data and measurements will all be of shape ``(B, C, H, W)`` where
[…]
