### 2025-06-29 412aafe
add automatic release message + automatic update of CHANGELOG + safeguard for not releasing when there has been no changes
