+++ .github/scripts/get_latest_changelog.py
@@ -1,0 +1,86 @@
+import re
+import sys
+from get_version import get_version_from_toml
+
+def rst_to_md(rst_text):
+    lines = rst_text.strip().splitlines()
+    md_lines = []
+    skip_next = False
+
+    for i, line in enumerate(lines):
+        if skip_next:
+            skip_next = False
+            continue
+
+        # Convert section headers
+        if line.strip() in {"New Features", "Changed", "Fixed"}:
+            md_lines.append(f"### {line.strip()}")
+            skip_next = True  # Skip the underline in the next line
+            continue
+
+        # Remove underline-only lines (that slipped through)
+        if re.match(r"^[-=^~]{3,}$", line.strip()):
+            continue
+
+        # Convert references
+        line = re.sub(r":gh:`(\d+)`", r"[#\1](https://github.com/deepinv/deepinv/pull/\1)", line)
+        line = re.sub(r"`([^`]+)`_", r"\1", line)
+        line = re.sub(r":class:`([^`]+)`", r"`\1`", line)
+
+        md_lines.append(line)
+
+    return "\n".join(md_lines).strip()
+
+
+def update_changelog(filename):
+    r"""
+    Function for automatic release changelog generation.
+    """
+
+    with open(filename, "r", encoding="utf-8") as f:
+        changelog = f.read()
+
+    # Extract the "Current" section
+    match = re.search(
+        r"(?P<header>^Current\s*-+\s*\n)(?P<body>.*?)(?=^v?\d+\.\d+\.\d+\s*-+\s*\n)",
+        changelog,
+        re.DOTALL | re.MULTILINE,
+    )
+
+    if not match:
+        print("Could not find a 'Current' section.")
+        sys.exit(1)
+
+    header = match.group("header")
+    body = match.group("body").rstrip()
+
+    # Extract current version from pyproject.toml or bump script
+    version = get_version_from_toml("../../pyproject.toml")
+    version_header = f"v{version}\n{'-' * len(version)}\n"
+
+    # Write extracted changes to changelog.txt
+    with open("changelog.txt", "w", encoding="utf-8") as f:
+        f.write(rst_to_md(body).strip())
+
+    # Prepare a new empty "Current" section
+    new_current = (
+        "Current\n"
+        "-------\n\n"
+        "New Features\n^^^^^^^^^^^^\n\n"
+        "Changed\n^^^^^^^\n\n"
+        "Fixed\n^^^^^\n\n"
+    )
+
+    # Replace the old "Current" section with the empty one and insert the versioned section below it
+    updated_changelog = changelog.replace(header + body, new_current + "\n" + version_header + body.strip() + "\n\n", 1)
+
+    with open(filename, "w", encoding="utf-8") as f:
+        f.write(updated_changelog)
+
+
+if __name__ == "__main__":
+    filename = "../../CHANGELOG.rst"
+    update_changelog(filename)
+    print("CHANGELOG.rst updated.")
+
+
