+++ pyproject.toml
@@ -4,7 +4,7 @@ build-backend = "setuptools.build_meta"
 
 [project]
 name = "deepinv"
-authors = [{name="Julian Tachella", email="tachellajulian@gmail.com"}]
+authors = [{name="DeepInverse maintainers (see https://github.com/deepinv/deepinv/blob/main/MAINTAINERS.md)"}]
 license = {text = "BSD 3-Clause"}
 description = "Pytorch library for solving inverse problems with deep learning"
 classifiers = [
