### 2025-06-30 af5b06e
LISTA exampl
