### 2025-06-30 343f50b
move to tomlkit + add manual minor/major release option
