+++ .github/scripts/get_version.py
@@ -1,5 +1,6 @@
 import tomlkit
 
+
 def get_version_from_toml(filename):
     with open(filename, "r", encoding="utf-8") as f:
         # Load the TOML file
@@ -10,10 +11,11 @@ def get_version_from_toml(filename):
 
 if __name__ == "__main__":
     import sys
+
     if len(sys.argv) != 2:
         print("Usage: python get_version.py path/to/pyproject.toml")
         sys.exit(1)
-    
+
     file_path = sys.argv[1]
     version = get_version_from_toml(file_path)
     print(version)
