+++ .github/workflows/release.yml
@@ -69,7 +69,7 @@ jobs:
         id: bump
         run: |
           python .github/scripts/bump_version.py pyproject.toml ${{github.event.inputs.bump_type}}$
-          echo "version=$(python .github/scripts/get_version.py)" >> "$GITHUB_OUTPUT"
+          echo "version=$(python .github/scripts/get_version.py pyproject.toml)" >> "$GITHUB_OUTPUT"
 
       - name: Build the package
         run: python -m build
