### 2025-06-30 1bef440
add toml as doc dependency
