### 2025-06-30 bccfa9f
move tomlkit to strict dependency + fix about
