+++ pyproject.toml
@@ -32,6 +32,7 @@ dependencies = [
     "wandb",
     "scipy",
     "natsort",
+    "tomlkit",
 ]
 
 [tool.setuptools]
@@ -66,7 +67,6 @@ doc = [
     "sphinx_autoapi",
     "sphinx-design",
     "sphinx_sitemap<2.7.0",
-    "tomlkit",
 ]
 
 # optional dependencies for specific denoisers
