+++ .github/scripts/bump_version.py
@@ -4,7 +4,10 @@ import tomlkit
 
 
 def increment_version(version, increment="patch"):
-    """Increment the patch version number (last digit)."""
+    """Increment the version number.
+    :param str version: The current version string in the format "major".minor.patch".
+    :param str increment: The part of the version to increment, can be "major", "minor", or "patch".
+    """
     parts = version.split(".")
 
     try:
