### 2025-06-30 c56acbe
change bib style
