+++ docs/source/refs.bib
@@ -960,3 +960,42 @@ journal = {ACM Trans. Graph.}
     booktitle={CVPR},
     year={2019}
 }
+
+@article{millard2023theoretical,
+  title={A theoretical framework for self-supervised MR image reconstruction using sub-sampling via variable density Noisier2Noise},
+  author={Millard, Charles and Chiew, Mark},
+  journal={IEEE transactions on computational imaging},
+  volume={9},
+  pages={707--720},
+  year={2023},
+  publisher={IEEE}
+}
+
+@article{millard2024clean,
+  title={Clean self-supervised MRI reconstruction from noisy, sub-sampled training data with Robust SSDU},
+  author={Millard, Charles and Chiew, Mark},
+  journal={Bioengineering},
+  volume={11},
+  number={12},
+  pages={1305},
+  year={2024},
+  publisher={MDPI}
+}
+
+@article{eldeniz2021phase2phase,
+  title={Phase2Phase: respiratory motion-resolved reconstruction of free-breathing magnetic resonance imaging using deep learning without a ground truth for improved liver imaging},
+  author={Eldeniz, Cihat and Gan, Weijie and Chen, Sihao and Fraum, Tyler J and Ludwig, Daniel R and Yan, Yan and Liu, Jiaming and Vahle, Thomas and Krishnamurthy, Uday and Kamilov, Ulugbek S and others},
+  journal={Investigative Radiology},
+  volume={56},
+  number={12},
+  pages={809--819},
+  year={2021},
+  publisher={LWW}
+}
+
+@inproceedings{pajot2019unsupervised,
+  title={Unsupervised adversarial image reconstruction},
+  author={Pajot, Arthur and De B{\'e}zenac, Emmanuel and Gallinari, Patrick},
+  booktitle={International conference on learning representations},
+  year={2019}
+}
