+++ deepinv/datasets/div2k.py
@@ -14,7 +14,7 @@ from deepinv.datasets.utils import (
 class DIV2K(torch.utils.data.Dataset):
     """Dataset for `DIV2K Image Super-Resolution Challenge <https://data.vision.ee.ethz.ch/cvl/DIV2K>`.
 
-    The DIV2K dataset :footcite:ps:`agustsson2017ntire` is a high-quality image dataset originally built for image super-resolution tasks.
+    The DIV2K dataset :footcite:t:`agustsson2017ntire` is a high-quality image dataset originally built for image super-resolution tasks.
 
     Images have varying sizes with up to 2040 vertical pixels, and 2040 horizontal pixels.
 
@@ -53,10 +53,7 @@ class DIV2K(torch.utils.data.Dataset):
         100
         >>> shutil.rmtree("DIV2K")                                    # remove raw data from disk
 
-    |sep|
 
-    :References:
-        .. footbibliography::
     """
 
     # https://data.vision.ee.ethz.ch/cvl/DIV2K/
