+++ docs/source/refs.bib
@@ -982,17 +982,6 @@ journal = {ACM Trans. Graph.}
   publisher={MDPI}
 }
 
-@article{eldeniz2021phase2phase,
-  title={Phase2Phase: respiratory motion-resolved reconstruction of free-breathing magnetic resonance imaging using deep learning without a ground truth for improved liver imaging},
-  author={Eldeniz, Cihat and Gan, Weijie and Chen, Sihao and Fraum, Tyler J and Ludwig, Daniel R and Yan, Yan and Liu, Jiaming and Vahle, Thomas and Krishnamurthy, Uday and Kamilov, Ulugbek S and others},
-  journal={Investigative Radiology},
-  volume={56},
-  number={12},
-  pages={809--819},
-  year={2021},
-  publisher={LWW}
-}
-
 @inproceedings{pajot2019unsupervised,
   title={Unsupervised adversarial image reconstruction},
   author={Pajot, Arthur and De B{\'e}zenac, Emmanuel and Gallinari, Patrick},
