+++ .github/workflows/release.yml
@@ -68,7 +68,7 @@ jobs:
       - name: Bump version
         id: bump
         run: |
-          python .github/scripts/bump_version.py pyproject.toml ${{github.event.inputs.bump_type}}$
+          python .github/scripts/bump_version.py pyproject.toml ${{github.event.inputs.bump_type}}
           echo "version=$(python .github/scripts/get_version.py pyproject.toml)" >> "$GITHUB_OUTPUT"
 
       - name: Build the package
@@ -120,6 +120,6 @@ jobs:
         run: |
           git config user.name "github-actions"
           git config user.email "github-actions@github.com"
-          git add pyproject.toml CHANGELOG.rst
+          git add pyproject.toml CHANGELOG.rst ${{github.event.inputs.bump_type}}
           git commit -m "Release v${{ steps.bump.outputs.version }}: update version, changelog"
           git push origin HEAD
