### 2025-06-30 6f76785
fix badges (#576)
