### 2025-06-30 f8b3ea5
add bump type in the right place for changelog
