### 2025-06-30 acd49e0
Update .github/scripts/get_latest_changelog.py
Co-authored-by: Andrewwango <andrew.wang27@gmail.com>
