### 2025-06-30 2407217
add auto_release (#573)
* add auto_release

* check if CI is passing

* black

* black github files

* copilot fix

* better patch bump

* fix get version

* even better bump

* fix maintainer link + pad library schematic

* Update deepinv/__about__.py

Co-authored-by: Andrewwango <andrew.wang27@gmail.com>

* Update docs/source/conf.py

Co-authored-by: Andrewwango <andrew.wang27@gmail.com>

* fixes

* black

* add automatic release message + automatic update of CHANGELOG + safeguard for not releasing when there has been no changes

* black

* fix import place + description

* move to tomlkit +
[…]
