+++ docs/source/conf.py
@@ -12,6 +12,8 @@ import os
 from pathlib import Path
 from sphinx.util import logging
 
+import tomlkit
+
 logger = logging.getLogger(__name__)
 
 import doctest
@@ -19,15 +21,18 @@ import doctest
 basedir = os.path.abspath(os.path.join(os.path.dirname(__file__), "..", ".."))
 sys.path.insert(0, basedir)
 
+
 # -- Project information -----------------------------------------------------
 # https://www.sphinx-doc.org/en/master/usage/configuration.html#project-information
 
-project = "deepinverse"
-copyright = "2024, deepinverse contributors"
-author = (
-    "Julian Tachella, Matthieu Terris, Samuel Hurault, Dongdong Chen and Andrew Wang"
-)
-release = "0.3"
+with open(os.path.join(basedir, "pyproject.toml"), "r", encoding="utf-8") as f:
+    metadata = tomlkit.parse(f.read())["project"]
+
+project = str(metadata["name"])
+copyright = "deepinverse contributors 2025"
+
+author = ", ".join(str(auth["name"]) for auth in metadata["authors"])
+release = str(metadata["version"])
 
 # -- General configuration ---------------------------------------------------
 # https://www.sphinx-doc.org/en/master/usage/configuration.html#general-configuration
