+++ deepinv/sampling/sampling.py
@@ -167,10 +167,13 @@ class BaseSampling(Reconstructor):
 
         Example:
             >>> from deepinv.sampling import BaseSampling, ULAIterator
+            >>> # Define iterator
+            >>> iterator = ULAIterator(...)
+            >>> 
             >>> # Basic usage with default settings
-            >>> sampler = BaseSampling(ULAIterator, data_fidelity, prior)
+            >>> sampler = BaseSampling(iterator, data_fidelity, prior)
             >>> mean, var = sampler.sample(measurements, forward_operator)
-
+            >>> 
             >>> # Using multiple statistics
             >>> sampler = BaseSampling(
             ...     iterator, data_fidelity, prior,
@@ -303,7 +306,7 @@ class BaseSampling(Reconstructor):
 
         Example:
             >>> from deepinv.sampling import BaseSampling, SamplingIterator
-            >>> sampler = BaseSampling(iterator, data_fidelity, prior, history_size=5)
+            >>> sampler = BaseSampling(SamplingIterator(...), data_fidelity, prior, history_size=5)
             >>> _ = sampler(measurements, forward_operator)
             >>> history = sampler.get_chain()
             >>> latest_state = history[-1]  # Get most recent state dictionary
