+++ deepinv/sampling/sampling.py
@@ -169,11 +169,11 @@ class BaseSampling(Reconstructor):
             >>> from deepinv.sampling import BaseSampling, ULAIterator
             >>> # Define iterator
             >>> iterator = ULAIterator(...)
-            >>> 
+            >>>
             >>> # Basic usage with default settings
             >>> sampler = BaseSampling(iterator, data_fidelity, prior)
             >>> mean, var = sampler.sample(measurements, forward_operator)
-            >>> 
+            >>>
             >>> # Using multiple statistics
             >>> sampler = BaseSampling(
             ...     iterator, data_fidelity, prior,
