+++ .github/workflows/release.yml
@@ -21,24 +21,6 @@ jobs:
     runs-on: ubuntu-latest
 
     steps:
-      - name: Check for recent release commit
-        id: check_release
-        run: |
-          LAST_COMMIT_MSG=$(git log -1 --pretty=%s)
-          echo "Last commit: $LAST_COMMIT_MSG"
-          if [[ "$LAST_COMMIT_MSG" =~ ^Release\ v[0-9]+\.[0-9]+\.[0-9]+ ]]; then
-            echo "Latest commit is a release. Skipping release process."
-            echo "skip=true" >> $GITHUB_OUTPUT
-          else
-            echo "skip=false" >> $GITHUB_OUTPUT
-          fi
-
-      - name: Exit if no new changes
-        if: steps.check_release.outputs.skip == 'true'
-        run: |
-          echo "No new changes since last release. Skipping workflow."
-          exit 0
-
       - name: Check CI is passing
         run: |
             gh auth setup-git
@@ -56,6 +38,24 @@ jobs:
       - name: Checkout code
         uses: actions/checkout@v4
 
+      - name: Check for recent release commit
+        id: check_release
+        run: |
+          LAST_COMMIT_MSG=$(git log -1 --pretty=%s)
+          echo "Last commit: $LAST_COMMIT_MSG"
+          if [[ "$LAST_COMMIT_MSG" =~ ^Release\ v[0-9]+\.[0-9]+\.[0-9]+ ]]; then
+            echo "Latest commit is a release. Skipping release process."
+            echo "skip=true" >> $GITHUB_OUTPUT
+          else
+            echo "skip=false" >> $GITHUB_OUTPUT
+          fi
+
+      - name: Exit if no new changes
+        if: steps.check_release.outputs.skip == 'true'
+        run: |
+          echo "No new changes since last release. Skipping workflow."
+          exit 0
+          
       - name: Set up Python
         uses: actions/setup-python@v5
         with:
