### 2025-07-01 ba19423
ruff UP006
