### 2025-07-01 291ce55
better doc references + langevin deprecation warnings
