### 2025-07-01 f068a8a
fix project metadata
