+++ docs/source/refs.bib
@@ -642,13 +642,6 @@ doi = {10.1287/moor.2016.0817}
   year={2018}
 }
 
-@inproceedings{pajot2019unsupervised,
-  title={Unsupervised adversarial image reconstruction},
-  author={Pajot, Arthur and De B{\'e}zenac, Emmanuel and Gallinari, Patrick},
-  booktitle={International conference on learning representations},
-  year={2019}
-}
-
 @article{aghabiglou2024r2d2,
   title={The R2D2 deep neural network series paradigm for fast precision imaging in radio astronomy},
   author={Aghabiglou, Amir and San Chu, Chung and Dabbech, Arwa and Wiaux, Yves},
