### 2025-07-01 4ef2891
docstring guidelines
