### 2025-07-01 0e78794
fix sampling doctest
