### 2025-07-01 8dfbe74
fix doctests
