### 2025-07-01 569ba32
Fix project metadata (#581)
* fix project metadata

* remove tomlkit
