### 2025-07-01 6b07abe
Refactor Sampling (#397)
* rstring + uv files

* base of new sampling classes added

* create utils.py

* ULA + Skrock

* formatting

* untracked oldsampling + formatting

* Added the old sampling modules back in + merged __init__.py

* Update deepinv/optim/optim_iterators/pgd.py

fix small typo

Co-authored-by: Andrewwango <andrew.wang27@gmail.com>

* Added doc strings to sampling + moved BaseSample to nn.Module

* raise an exception if an iterator doesn't implement forward

* typo in SKRock

* added clipping

* added sampling_build wrapper

* Uncertainty quantification example + Custom Kerne
[…]
