+++ deepinv/sampling/__init__.py
@@ -1,5 +1,13 @@
-from .langevin import ULA, MonteCarlo, SKRock
+from .sampling import BaseSampling, sampling_builder
+from .langevin import ULA, SKRock
 from .diffusion import DDRM, DiffusionSampler, DiffPIR, DPS
+from .sampling_iterators import (
+    ULAIterator,
+    SKRockIterator,
+    SamplingIterator,
+    SKROCKIterator,
+    DiffusionIterator,
+)
 from . import diffusion_sde, sde_solver
 from .noisy_datafidelity import NoisyDataFidelity, DPSDataFidelity
 from .diffusion_sde import (
