### 2025-07-01 61e7cd2
more fixes to release
