### 2025-07-02 f542ae7
add debugging yml
