### 2025-07-02 a111518
fixes release
