+++ .github/workflows/test_release.yml
@@ -1,115 +1,0 @@
-name: Test Auto Publish to PyPI
-
-on:
-  push:
-    branches:
-      - tachella-patch-4
-  workflow_dispatch:        # manual trigger
-    inputs:
-      bump_type:
-        description: 'Type of version bump'
-        required: true
-        default: 'dev'
-        type: choice
-        options:
-          - dev # for debugging purposes, not a real version bump
-
-permissions:
-  contents: write
-  
-jobs:
-  build-test-publish:
-    name: Auto version, test & publish
-    runs-on: ubuntu-latest
-
-    steps:
-      - name: Check CI is passing
-        env:
-          GH_TOKEN: ${{ secrets.GITHUB_TOKEN }}
-        run: |
-          REPO="deepinv/deepinv"
-
-          get_status () {
-            curl -s -H "Authorization: token $GH_TOKEN" \
-              "https://api.github.com/repos/$REPO/actions/workflows/$1/runs?branch=main&status=completed&per_page=1" \
-              | jq -r '.workflow_runs[0].conclusion'
-          }
-
-          STATUS_TEST=$(get_status "test.yml")
-          STATUS_DOC=$(get_status "documentation.yml")
-
-          echo "Test: $STATUS_TEST"
-          echo "Docs: $STATUS_DOC"
-
-          if [[ "$STATUS_TEST" != "success" || "$STATUS_DOC" != "success" ]]; then
-            echo "❌ One or both workflows failed — aborting."
-            exit 1
-          fi
-
-      - name: Checkout code
-        uses: actions/checkout@v4
-
-      - name: Check for recent release commit
-        id: check_release
-        run: |
-          LAST_COMMIT_MSG=$(git log -1 --pretty=%s)
-          echo "Last commit: $LAST_COMMIT_MSG"
-          if [[ "$LAST_COMMIT_MSG" =~ ^Release\ v[0-9]+\.[0-9]+\.[0-9]+ ]]; then
-            echo "Latest commit is a release. Skipping release process."
-            echo "skip=true" >> $GITHUB_OUTPUT
-          else
-            echo "skip=false" >> $GITHUB_OUTPUT
-          fi
-
-      - name: Exit if no new changes
-        if: steps.check_release.outputs.skip == 'true'
-        run: |
-          echo "No new changes since last release. Skipping workflow."
-          exit 1
-          
-      - name: Set up Python
-        uses: actions/setup-python@v5
-        with:
-          python-version: '3.9'
-
-      - name: Install tools
-        run: |
-          pip install --upgrade build twine tomlkit
-
-      - name: Update changelog 
-        id: changelog
-        run: |
-          python .github/scripts/get_latest_changelog.py CHANGELOG.rst dev
-
-      - name: Bump version
-        id: bump
-        run: |
-          python .github/scripts/bump_version.py pyproject.toml dev
-          echo "version=$(python .github/scripts/get_version.py pyproject.toml)" >> "$GITHUB_OUTPUT"
-
-      - name: Build the package
-        run: python -m build
-
-      - name: Commit version bump and tag
-        run: |
-          git config user.name "github-actions"
-          git config user.email "github-actions@github.com"
-          git checkout -b releasetest3
-          git add pyproject.toml CHANGELOG.rst 
-          git commit -m "Release version ${{ steps.bump.outputs.version }}"
-          git push origin HEAD
-          
-      - name: Publish to TestPyPI
-        env:
-          TWINE_USERNAME: __token__
-          TWINE_PASSWORD: ${{ secrets.TEST_PYPI_TOKEN }}
-        run: |
-          python -m twine check dist/*
-          python -m twine upload --verbose --repository testpypi dist/*
-
[…]
