### 2025-07-02 b2932f0
add github release notes
