+++ .github/workflows/release.yml
@@ -127,4 +127,11 @@ jobs:
           TWINE_PASSWORD: ${{ secrets.PYPI_TOKEN }}
         run: |
           python -m twine --verbose upload dist/*
-          
+
+      - name: Create Github release
+        uses: softprops/action-gh-release@v2
+        with:
+          tag_name: v${{ steps.bump.outputs.version }}
+          body_path: changelog.txt
+          name: Release v${{ steps.bump.outputs.version }}
+          files: dist/*
