### 2025-07-02 38ef7d6
more fixes to release (#584)
* more fixes to release

* Update release.yml

* Update release.yml

* Update release.yml

* Update release.yml

* add debugging yml

* black + change title of test

* test

* fixes release

* try again

* onemoretime

* v3

* final fixes

* add github release notes
