+++ .github/workflows/release.yml
@@ -14,7 +14,11 @@ on:
           - patch
           - minor
           - major
+          - dev # for debugging
 
+permissions:
+  contents: write
+  
 jobs:
   build-test-publish:
     name: Auto version, test & publish
@@ -22,18 +26,27 @@ jobs:
 
     steps:
       - name: Check CI is passing
+        env:
+          GH_TOKEN: ${{ secrets.GITHUB_TOKEN }}
         run: |
-            gh auth setup-git
-            # Check latest run for "Test" workflow
-            STATUS_TEST=$(gh run list --repo deepinv/deepinv --workflow test.yml --limit 1 --json conclusion --jq '.[0].conclusion')
-            echo "Test workflow conclusion: $STATUS_TEST"
-            # Check "documentation"
-            STATUS_DOC=$(gh run list --repo deepinv/deepinv --workflow documentation.yml --limit 1 --json conclusion --jq '.[0].conclusion')
-            echo "Doc workflow conclusion: $STATUS_DOC"
-            if [ "$STATUS_TEST" != "SUCCESS" ] || [ "$STATUS_DOC" != "SUCCESS" ]; then
-            echo "❌ Tests not passing — aborting release of new version."
+          REPO="deepinv/deepinv"
+
+          get_status () {
+            curl -s -H "Authorization: token $GH_TOKEN" \
+              "https://api.github.com/repos/$REPO/actions/workflows/$1/runs?branch=main&status=completed&per_page=1" \
+              | jq -r '.workflow_runs[0].conclusion'
+          }
+
+          STATUS_TEST=$(get_status "test.yml")
+          STATUS_DOC=$(get_status "documentation.yml")
+
+          echo "Test: $STATUS_TEST"
+          echo "Docs: $STATUS_DOC"
+
+          if [[ "$STATUS_TEST" != "success" || "$STATUS_DOC" != "success" ]]; then
+            echo "❌ One or both workflows failed — aborting."
             exit 1
-            fi
+          fi
 
       - name: Checkout code
         uses: actions/checkout@v4
@@ -54,7 +67,7 @@ jobs:
         if: steps.check_release.outputs.skip == 'true'
         run: |
           echo "No new changes since last release. Skipping workflow."
-          exit 0
+          exit 1
           
       - name: Set up Python
         uses: actions/setup-python@v5
@@ -65,61 +78,60 @@ jobs:
         run: |
           pip install --upgrade build twine tomlkit
 
+      - name: Build the package
+        run: python -m build
+
+      - name: Update changelog
+        id: changelog
+        run: |
+          python .github/scripts/get_latest_changelog.py CHANGELOG.rst ${{github.event.inputs.bump_type}}
+
       - name: Bump version
         id: bump
         run: |
           python .github/scripts/bump_version.py pyproject.toml ${{github.event.inputs.bump_type}}
           echo "version=$(python .github/scripts/get_version.py pyproject.toml)" >> "$GITHUB_OUTPUT"
 
-      - name: Build the package
-        run: python -m build
+      - name: Stop if not on main
+        run: |
+          if [ "${{ github.ref_name }}" != "main" ]; then
+            echo "❌ Not on main branch — skipping release."
+            exit 1
+          fi
 
+      - name: Commit version bump
+        run: |
+          git config user.name "github-actions"
+          git config user.email "github-actions@github.com"
+          git add pyproject.toml CHANGELOG.rst 
+          git commit -m "Release version ${{ steps.bump.outputs.version }}"
+          git push origin HEAD
+          
       - name: Publish to TestPyPI
         env:
           TWINE_USERNAME: __token__
           TWINE_PASSWORD: ${{ secrets.TEST_PYPI_TOKEN }}
         run: |
-          python -m twine upload --repository testpypi dist/*
+          python -m twine check dist/*
[…]
