+++ .github/workflows/upload_pypi.yml
@@ -12,6 +12,21 @@ jobs:
       github.event.pull_request.head.ref == 'release-branch'
     runs-on: ubuntu-latest
     steps:
+      - name: Checkout code
+        uses: actions/checkout@v4
+
+      - name: Set up Python
+        uses: actions/setup-python@v5
+        with:
+          python-version: '3.9'
+
+      - name: Install tools
+        run: |
+          pip install --upgrade build twine
+
+      - name: Build the package
+        run: python -m build
+
       - name: Publish to TestPyPI
         env:
           TWINE_USERNAME: __token__
@@ -31,3 +46,5 @@ jobs:
           TWINE_PASSWORD: ${{ secrets.PYPI_TOKEN }}
         run: |
           python -m twine --verbose upload dist/*
+        
+      
