### 2025-07-03 2fbe196
fix release one last time (#585)
* fix release one last time

* divide into two actions

* easier

* fix release
