+++ .github/workflows/release.yml
@@ -1,8 +1,6 @@
 name: Auto Publish to PyPI
 
 on:
-  schedule:
-    - cron: '0 0 */21 * *'  # every 21 days at 00:00 UTC
   workflow_dispatch:        # manual trigger
     inputs:
       bump_type:
@@ -47,27 +45,6 @@ jobs:
             echo "❌ One or both workflows failed — aborting."
             exit 1
           fi
-
-      - name: Checkout code
-        uses: actions/checkout@v4
-
-      - name: Check for recent release commit
-        id: check_release
-        run: |
-          LAST_COMMIT_MSG=$(git log -1 --pretty=%s)
-          echo "Last commit: $LAST_COMMIT_MSG"
-          if [[ "$LAST_COMMIT_MSG" =~ ^Release\ v[0-9]+\.[0-9]+\.[0-9]+ ]]; then
-            echo "Latest commit is a release. Skipping release process."
-            echo "skip=true" >> $GITHUB_OUTPUT
-          else
-            echo "skip=false" >> $GITHUB_OUTPUT
-          fi
-
-      - name: Exit if no new changes
-        if: steps.check_release.outputs.skip == 'true'
-        run: |
-          echo "No new changes since last release. Skipping workflow."
-          exit 1
           
       - name: Set up Python
         uses: actions/setup-python@v5
@@ -76,10 +53,10 @@ jobs:
 
       - name: Install tools
         run: |
-          pip install --upgrade build twine tomlkit
+          pip install --upgrade tomlkit
 
-      - name: Build the package
-        run: python -m build
+      - name: Checkout code
+        uses: actions/checkout@v4
 
       - name: Update changelog
         id: changelog
@@ -92,43 +69,24 @@ jobs:
           python .github/scripts/bump_version.py pyproject.toml ${{github.event.inputs.bump_type}}
           echo "version=$(python .github/scripts/get_version.py pyproject.toml)" >> "$GITHUB_OUTPUT"
 
-      - name: Stop if not on main
-        run: |
-          if [ "${{ github.ref_name }}" != "main" ]; then
-            echo "❌ Not on main branch — skipping release."
-            exit 1
-          fi
-
-      - name: Commit version bump
-        run: |
-          git config user.name "github-actions"
-          git config user.email "github-actions@github.com"
-          git add pyproject.toml CHANGELOG.rst 
-          git commit -m "Release version ${{ steps.bump.outputs.version }}"
-          git push origin HEAD
-          
-      - name: Publish to TestPyPI
-        env:
-          TWINE_USERNAME: __token__
-          TWINE_PASSWORD: ${{ secrets.TEST_PYPI_TOKEN }}
-        run: |
-          python -m twine check dist/*
-          python -m twine upload --verbose --repository testpypi dist/*
-
-      - name: Install from TestPyPI and test
-        run: |
-          pip install --index-url https://test.pypi.org/simple/ deepinv --extra-index-url https://pypi.org/simple
-          python -c "import deepinv"
-          
-      - name: Publish to PyPI
-        if: ${{ success() }}
-        env:
-          TWINE_USERNAME: __token__
-          TWINE_PASSWORD: ${{ secrets.PYPI_TOKEN }}
-        run: |
-          python -m twine --verbose upload dist/*
-
+      - name: create PR with changes
+        uses: peter-evans/create-pull-request@v7
+        with:
+          branch: release-branch
+          title: "Release v${{ steps.bump.outputs.version }}"
+          body: |
+            This automatic PR updates the version to v${{ steps.bump.outputs.version }} and updates the changelog.
+            Please merge this PR to finalize the release.
+          commit-message: "Release v${{ steps.bump.outputs.version }}"
+          add-paths: |
[…]
