### 2025-07-03 35f54c6
Release v0.3.2 (#586)
