+++ .github/workflows/upload_pypi.yml
@@ -45,6 +45,6 @@ jobs:
           TWINE_USERNAME: __token__
           TWINE_PASSWORD: ${{ secrets.PYPI_TOKEN }}
         run: |
-          python -m twine --verbose upload dist/*
+          python -m twine upload dist/*
         
       
