### 2025-07-03 35b0745
fixes release and upload to pypi (#588)
* Update upload_pypi.yml

* Update release.yml

* Update release.yml
