+++ .github/workflows/release.yml
@@ -16,6 +16,7 @@ on:
 
 permissions:
   contents: write
+  pull-requests: write
   
 jobs:
   build-test-publish:
