+++ _downloads/d606302330961a578850ed4f90bda982/demo_DEQ.py
@@ -1,0 +1,236 @@
+r"""
+Deep Equilibrium (DEQ) algorithms for image deblurring
+====================================================================================================
+
+This a toy example to show you how to use DEQ to solve a deblurring problem.
+Note that this is a small dataset for training. For optimal results, use a larger dataset.
+For visualizing the training, you can use Weight&Bias (wandb) by setting ``wandb_vis=True``.
+
+For now DEQ is only possible with PGD, HQS and GD optimization algorithms.
+
+"""
+
+import deepinv as dinv
+from pathlib import Path
+import torch
+from torch.utils.data import DataLoader
+from deepinv.models import DnCNN
+from deepinv.optim.data_fidelity import L2
+from deepinv.optim.prior import PnP
+from deepinv.unfolded import DEQ_builder
+from deepinv.training import train, test
+from torchvision import transforms
+from deepinv.utils.demo import load_dataset, load_degradation
+
+
+# %%
+# Setup paths for data loading and results.
+# ----------------------------------------------------------------------------------------
+#
+
+BASE_DIR = Path(".")
+DATA_DIR = BASE_DIR / "measurements"
+RESULTS_DIR = BASE_DIR / "results"
+CKPT_DIR = BASE_DIR / "ckpts"
+DEG_DIR = BASE_DIR / "degradations"
+
+# Set the global random seed from pytorch to ensure reproducibility of the example.
+torch.manual_seed(0)
+
+device = dinv.utils.get_freer_gpu() if torch.cuda.is_available() else "cpu"
+
+# %%
+# Load base image datasets and degradation operators.
+# ----------------------------------------------------------------------------------------
+# In this example, we use the CBSD500 dataset and the Set3C dataset for testing.
+
+img_size = 32
+n_channels = 3  # 3 for color images, 1 for gray-scale images
+operation = "deblurring"
+# For simplicity, we use a small dataset for training.
+# To be replaced for optimal results. For example, you can use the larger "drunet" dataset.
+train_dataset_name = "CBSD500"
+test_dataset_name = "set3c"
+# Generate training and evaluation datasets in HDF5 folders and load them.
+test_transform = transforms.Compose(
+    [transforms.CenterCrop(img_size), transforms.ToTensor()]
+)
+train_transform = transforms.Compose(
+    [transforms.RandomCrop(img_size), transforms.ToTensor()]
+)
+train_base_dataset = load_dataset(train_dataset_name, transform=train_transform)
+test_base_dataset = load_dataset(test_dataset_name, transform=test_transform)
+
+
+# %%
+# Generate a dataset of low resolution images and load it.
+# ----------------------------------------------------------------------------------------
+# We use the Downsampling class from the physics module to generate a dataset of low resolution images.
+
+
+# Use parallel dataloader if using a GPU to fasten training, otherwise, as all computes are on CPU, use synchronous
+# dataloading.
+num_workers = 4 if torch.cuda.is_available() else 0
+
+# Degradation parameters
+noise_level_img = 0.03
+
+# Generate a motion blur operator.
+kernel_index = 1  # which kernel to chose among the 8 motion kernels from 'Levin09.mat'
+kernel_torch = load_degradation("Levin09.npy", DEG_DIR / "kernels", index=kernel_index)
+kernel_torch = kernel_torch.unsqueeze(0).unsqueeze(
+    0
+)  # add batch and channel dimensions
+
+# Generate the gaussian blur downsampling operator.
+physics = dinv.physics.BlurFFT(
[…]
