### 2025-07-04 287fba4
try shorten to ~2000 words
