+++ paper/paper.md
@@ -113,14 +113,13 @@ bibliography: paper.bib
 
 # Summary
 
-DeepInverse is an open-source PyTorch-based library for imaging inverse problems.
-The library covers all crucial steps in image reconstruction from the efficient implementation of forward operators (e.g., optics, MRI, tomography), 
-to the definition and resolution of variational problems and the design and training of advanced neural network architectures. 
+[DeepInverse](https://deepinv.github.io/) is an open-source PyTorch-based library for imaging inverse problems. DeepInverse implements all crucial steps in image reconstruction from imaging forward operators across a wide set of domains (medical imaging, astronomical imaging, remote sensing, computational photography, compressed sensing, and more)
+to reconstruction algorithms and advanced neural network training. 
 
 # Statement of Need
 
-Deep neural networks have become ubiquitous in various imaging inverse problems, from computational photography to astronomical and medical imaging. Despite the ever-increasing research effort in the field, most learning-based algorithms are built from scratch, are hard to generalize beyond their specific training setting, and the reported results are often hard to reproduce. `deepinv` overcomes these limitations by providing a unified framework for defining imaging operators and solvers. It leverages the popular PyTorch deep learning library [@paszke2019pytorch], making most modules compatible with auto-differentiation.
-The target audience of this library is both researchers in inverse problems (experts in optimization, machine learning, etc.) and practitioners (biologists, physicists, etc.). `deepinv` has the following objectives:
+Deep neural networks have become ubiquitous in various imaging inverse problems. Despite the ever-increasing research effort in the field, most learning-based algorithms are built from scratch, are hard to generalize beyond their specific training setting, and the reported results are often hard to reproduce. `deepinv` overcomes these limitations by providing a modular unified framework, leveraging the popular PyTorch deep learning library [@paszke2019pytorch], making most modules compatible with auto-differentiation.
+The target audience of this library is both researchers in inverse problems (experts in optimization, machine learning etc.), practitioners (biologists, physicists etc.) and imaging software engineers. `deepinv` has the following objectives:
 
 1. **Accelerate research** by enabling efficient testing, deployment and transfer
 of new ideas across imaging domains;
@@ -128,12 +127,11 @@ of new ideas across imaging domains;
 3. Enhance **research reproducibility** via a common framework for imaging operators, reconstruction
 methods, datasets, and metrics for inverse problems.
 
-While other Python computational imaging libraries exist, to the best of our knowledge, `deepinv` is the only one with a strong focus on learning-based methods, providing a larger set of realistic imaging operators.
-SCICO [@balke2022scico] and Pyxu [@simeoni2022pyxu] are Python libraries whose main focus is variational optimization and/or plug-and-play reconstruction methods. These libraries do not provide specific tools for training reconstruction models such as trainers and custom loss functions, and do not cover non optimization-based solvers including diffusion methods, adversarial methods or unrolling networks.
-CUQIpy [@riis2024cuqipy] is a library focusing on Bayesian uncertainty quantification methods for inverse problems.
-TIGRE [@biguri2025tigre], ODL [@adler2018odl] and CIL [@jorgensen2021core] mostly focus on computed tomography and do not cover deep learning pipelines for inverse solvers. 
-There are also multiple libraries focusing on specific inverse problems: ASTRA [@van2016astra] and the related pytomography [@polson2025pytomography] define advanced tomography operators, sigpy [@ong2019sigpy] provides magnetic resonance imaging (MRI) operators without deep learning, and PyLops [@ravasi2019pylops] provides a linear operator class and many built-in linear operators.
[…]
