+++ paper/paper.md
@@ -127,11 +127,11 @@ of new ideas across imaging domains;
 3. Enhance **research reproducibility** via a common framework for imaging operators, reconstruction
 methods, datasets, and metrics for inverse problems.
 
-To the best of our knowledge, `deepinv` is the only library with a strong focus and wide set of modern learning-based methods, while covering many imaging domains.
-SCICO [@balke2022scico] and Pyxu [@simeoni2022pyxu] focus only on optimization-based methods, and do not include neural network training, diffusion methods, adversarial methods nor unrolling networks.
-CUQIpy [@riis2024cuqipy] focuses on Bayesian uncertainty quantification methods for inverse problems.
-There are also multiple libraries focusing on specific inverse problems, and we encourage `deepinv` to be used alongside any such library: ASTRA [@van2016astra], pytomography [@polson2025pytomography], TIGRE [@biguri2025tigre], ODL [@adler2018odl] and CIL [@jorgensen2021core] for tomography, sigpy [@ong2019sigpy] for magnetic resonance imaging (without deep learning), and PyLops [@ravasi2019pylops] for certain linear operators. 
-Advanced libraries for inverse problems also exist in MATLAB, including GlobalBioIm [@soubies2019pocket] or IR Tools [@gazzola2019ir], but they are restricted to handcrafted reconstruction methods without automatic differentiation.
+To the best of our knowledge, `deepinv` is the only library with a strong focus and wide set of modern learning-based methods, covering many imaging domains.
+SCICO [@balke2022scico] and Pyxu [@simeoni2022pyxu] focus only on optimization-based methods.
+CUQIpy [@riis2024cuqipy] focuses on Bayesian uncertainty quantification methods.
+ASTRA [@van2016astra], pytomography [@polson2025pytomography], TIGRE [@biguri2025tigre], ODL [@adler2018odl] and CIL [@jorgensen2021core] focus only on tomography, sigpy [@ong2019sigpy] on magnetic resonance imaging, and PyLops [@ravasi2019pylops] on certain linear operators. 
+MATLAB libraries such as GlobalBioIm [@soubies2019pocket] or IR Tools [@gazzola2019ir] are restricted to handcrafted reconstruction methods without automatic differentiation.
 
 
 ![Schematic of the library.\label{fig:schematic}](../docs/source/figures/deepinv_schematic.png)
@@ -144,16 +144,12 @@ y = N_{\sigma}(A_{\xi}(x))
 \end{equation}
 where $x\in\mathcal{X}$ is an image, $y\in\mathcal{Y}$ are the measurements, $A_{\xi}\colon\mathcal{X}\mapsto\mathcal{Y}$ is a
 deterministic (linear or non-linear) operator capturing the physics of the acquisition and
-$N_{\sigma}\colon\mathcal{Y}\mapsto \mathcal{Y}$ is a mapping that characterizes the noise affecting the 
-measurements (Gaussian, Poisson etc.) parameterized by $\sigma$ (e.g. the noise level and/or gain). The forward operation is written in `deepinv` as `x = physics(y, **params)` where `params` is a dictionary with
-optional parameters $\xi$ (blur kernels, noise levels, MRI sampling masks, tomographic projection angles etc.). This framework unifies the wide variety of forward operators across various domains. The ever-expanding library of physics is enumerated in the [documentation](https://deepinv.github.io/deepinv/user_guide/physics/physics.html).
[…]
