+++ deepinv/datasets/utils.py
@@ -181,22 +181,3 @@ class ToComplex(Module):
         :param torch.Tensor x: image tensor of shape (..., H, W)
         """
         return stack([x, zeros_like(x)], dim=-3)
-
-
-class UnsupDataset(torch.utils.data.Dataset):
-    r"""
-    Dataset for unsupervised learning tasks.
-
-    This dataset is used to return only the data without any labels.
-
-    :param torch.Tensor data: Input data tensor of shape (N, ...), where N is the number of samples and ... represents the data dimensions.
-    """
-
-    def __init__(self, data):
-        self.data = data
-
-    def __len__(self):
-        return self.data.size(0)
-
-    def __getitem__(self, idx):
-        return torch.nan, self.data[idx]
