+++ deepinv/models/ram.py
@@ -5,24 +5,20 @@ import torch.nn.functional as F
 import deepinv as dinv
 from deepinv.physics import MultiScaleLinearPhysics, Pad
 from deepinv.utils.tensorlist import TensorList
-from .base import Reconstructor
-
-from huggingface_hub import hf_hub_download
-
-cuda = True if torch.cuda.is_available() else False
-Tensor = torch.cuda.FloatTensor if cuda else torch.FloatTensor
+from deepinv.models.base import Reconstructor
 
 
 class RAM(Reconstructor):
     r"""
-    RAM model.
+    Reconstruct Anything Model.
 
     This model (proposed in `this paper <https://arxiv.org/abs/2503.08915>`_) is a convolutional neural network (CNN)
     designed for image reconstruction tasks.
 
-    :param in_channels: Number of input channels. If a list is provided, the model will have separate heads for each channel.
-    :param device: Device to which the model should be moved. If None, the model will be created on the default device.
-    :param pretrained: If True, the model will be initialized with pretrained weights.
+    :param list in_channels: Number of input channels. If a list is provided, the model will have separate heads for each channel.
+    :param str device: Device to which the model should be moved. If None, the model will be created on the default device.
+    :param bool pretrained: If True, the model will be initialized with pretrained weights.
+    :param float sigma_threshold: Threshold (minimum value) for the noise level. Default is 1e-3.
     """
 
     def __init__(
@@ -30,6 +26,7 @@ class RAM(Reconstructor):
         in_channels=[1, 2, 3],
         device=None,
         pretrained=True,
+        sigma_threshold=1e-3,
     ):
         super(RAM, self).__init__()
 
@@ -78,14 +75,12 @@ class RAM(Reconstructor):
 
         self.m_tail = OutTail(nc[0], in_channels)
 
+        self.sigma_threshold = sigma_threshold
+
         # load pretrained weights from hugging face
         if pretrained:
-            self.load_state_dict(
-                torch.load(
-                    hf_hub_download(repo_id="mterris/ram", filename="ram.pth.tar"),
-                    map_location=device,
-                )
-            )
+            url_download = "https://huggingface.co/mterris/ram/resolve/main/ram.pth.tar"
+            self.load_state_dict(torch.hub.load_state_dict_from_url(url_download))
 
         if device is not None:
             self.to(device)
@@ -94,8 +89,8 @@ class RAM(Reconstructor):
         r"""
         Converts a constant value to a map of the same size as the input tensor x.
 
-        :params float value: constant value
-        :params torch.Tensor x: input tensor
+        :param float value: constant value
+        :param torch.Tensor x: input tensor
         """
         if isinstance(value, torch.Tensor):
             if value.ndim > 0:
@@ -122,9 +117,9 @@ class RAM(Reconstructor):
         Realign the input x based on the measurements y and the physics model.
         Applies the proximity operator of the L2 norm with respect to the physics model.
 
-        :params torch.Tensor x: Input tensor
-        :params deepinv.physics.Physics physics: Physics model
-        :params torch.Tensor y: Measurements
+        :param torch.Tensor x: Input tensor
+        :param deepinv.physics.Physics physics: Physics model
+        :param torch.Tensor y: Measurements
         """
         if hasattr(physics, "factor"):
             f = physics.factor
@@ -172,11 +167,11 @@ class RAM(Reconstructor):
         r"""
         Forward pass of the UNet model.
 
-        :params torch.Tensor x0: init image
-        :params float sigma: Gaussian noise level
-        :params float gamma: Poisson noise gain
-        :params deepinv.physics.Physics physics: physics measurement operator
-        :params torch.Tensor y: measurements
[…]
