+++ deepinv/tests/test_models.py
@@ -4,6 +4,7 @@ import torch
 from torch.utils.data import DataLoader, Dataset
 
 import deepinv as dinv
+from deepinv.loss import PSNR
 from dummy import DummyCircles, DummyModel
 
 from deepinv.tests.test_physics import find_operator
@@ -684,10 +685,14 @@ def test_varnet(varnet_type, device):
     assert psnr(x_init, x) < psnr(x_hat, x)
 
 
+LIST_IMAGE_WHSIZE = [(32, 37), (25, 129)]
+
+
+@pytest.mark.parametrize("whsize", LIST_IMAGE_WHSIZE)
 @pytest.mark.parametrize("model", REST_MODEL_LIST)
 @pytest.mark.parametrize("physics_name", LINEAR_OPERATORS)
 @pytest.mark.parametrize("channels", CHANNELS)
-def test_restoration_model(device, model, physics_name, channels, rng):
+def test_restoration_model(device, model, physics_name, channels, rng, whsize):
 
     # skip test if channel is 1 and physics_name is in ["demosaicing", "MRI"]
     if channels == 1 and physics_name in ["demosaicing", "MRI"]:
@@ -704,20 +709,36 @@ def test_restoration_model(device, model, physics_name, channels, rng):
     model = choose_restoration_model(model).to(device)
     torch.manual_seed(0)
 
-    imsize = (
-        channels,
-        32,
-        37,
-    )  # minimal size for this model is (3, 64, 64), but we want to test non-square images
+    imsize = (channels, whsize[0], whsize[1])
+
     physics, imsize, _, dtype = find_operator(physics_name, device, imsize=imsize)
 
-    x = torch.randn(imsize, device=device, dtype=dtype, generator=rng).unsqueeze(0)
+    if hasattr(physics, "noise_model"):
+        if hasattr(physics.noise_model, "sigma"):
+            physics.noise_model.sigma = torch.tensor(
+                [max(physics.noise_model.sigma, 0.01)]
+            )
+        else:
+            physics.noise_model = dinv.physics.GaussianNoise(0.01, rng=rng)
+    else:
+        physics.noise_model = dinv.physics.GaussianNoise(0.01, rng=rng)
+
+    x = DummyCircles(imsize=imsize, samples=1)[0].unsqueeze(0)
+
     y = physics(x)
 
     with torch.no_grad():
         x_hat = model(y, physics)
 
+    psnr_fn = PSNR(max_pixel=1)
     assert x_hat.shape == x.shape
+
+    if not (
+        physics_name == "super_resolution_circular" and channels == 2
+    ):  # suboptimal performance in this case
+        psnr_in = psnr_fn(physics.A_dagger(y), x)
+        psnr_out = psnr_fn(x_hat, x)
+        assert torch.all(psnr_out > psnr_in)
 
 
 def test_pannet():
