+++ paper/paper.md
@@ -71,13 +71,13 @@ authors:
     affiliation: 8  
 
 affiliations:
-  - name: CNRS, ENS de Lyon, Univ Lyon, Lyon, France
+  - name: CNRS, ENS de Lyon, France
     index: 1
   - name: Université Paris-Saclay, Inria, CEA, Palaiseau, France
     index: 2
-  - name: CNRS, ENS Paris, PSL, Paris, France
+  - name: CNRS, ENS Paris, PSL, France
     index: 3
-  - name: University of Edinburgh, Edinburgh, UK
+  - name: University of Edinburgh, UK
     index: 4
   - name: Heriot-Watt University, Edinburgh, UK
     index: 5
@@ -85,15 +85,15 @@ affiliations:
     index: 6
   - name: EPFL, Lausanne, Switzerland
     index: 7
-  - name: IRIT, CBI, CNRS, Université de Toulouse, Toulouse, France. 
+  - name: IRIT, CBI, CNRS, Université de Toulouse, France
     index: 8
-  - name: IMT, CNRS, Université de Toulouse, Toulouse, France
+  - name: IMT, CNRS, Université de Toulouse, France
     index: 9
-  - name: Inria, ENS de Lyon, Univ Lyon, Lyon, France
+  - name: Inria, ENS de Lyon, France
     index: 10
-  - name: INSA de Lyon, Lyon, France
+  - name: INSA de Lyon, France
     index: 11
-  - name: University of Texas at Austin, Austin, USA
+  - name: University of Texas at Austin, USA
     index: 12
   - name: IRFU, CEA, Université Paris-Saclay, Gif-sur-Yvette, France
     index: 13
@@ -103,7 +103,7 @@ affiliations:
     index: 15
   - name: Chemnitz University of Technology, Chemnitz, Germany
     index: 16
-  - name: Department of Imaging and Pathology, KU Leuven, Leuven, Belgium
+  - name: Department of Imaging and Pathology, KU Leuven, Belgium
     index: 17
 
 date: 15 May 2025
@@ -113,26 +113,23 @@ bibliography: paper.bib
 
 # Summary
 
-[DeepInverse](https://deepinv.github.io/) is an open-source PyTorch-based library for imaging inverse problems. DeepInverse implements all crucial steps in image reconstruction from imaging forward operators across a wide set of domains (medical imaging, astronomical imaging, remote sensing, computational photography, compressed sensing, and more)
-to reconstruction algorithms and advanced neural network training. 
+[DeepInverse](https://deepinv.github.io/) is an open-source PyTorch-based library for imaging inverse problems. DeepInverse implements all the steps to reconstruct images using deep learning for a wide set of domains (medical imaging, astronomical imaging, remote sensing, computational photography, compressed sensing and more).
 
 # Statement of Need
 
-Deep neural networks have become ubiquitous in various imaging inverse problems. Despite the ever-increasing research effort in the field, most learning-based algorithms are built from scratch, are hard to generalize beyond their specific training setting, and the reported results are often hard to reproduce. `deepinv` overcomes these limitations by providing a modular unified framework, leveraging the popular PyTorch deep learning library [@paszke2019pytorch], making most modules compatible with auto-differentiation.
-The target audience of this library is both researchers in inverse problems (experts in optimization, machine learning etc.), practitioners (biologists, physicists etc.) and imaging software engineers. `deepinv` has the following objectives:
[…]
