+++ paper/paper.bib
@@ -11,7 +11,8 @@
   author={Chung, Hyungjin and Kim, Jeongsol and Kim, Sehui and Ye, Jong Chul},
   booktitle={Proceedings of the IEEE/CVF Conference on Computer Vision and Pattern Recognition},
   pages={6059--6069},
-  year={2023}
+  year={2023},
+  doi={10.1109/cvpr52729.2023.00587}
 }
 
 @article{debarnot2024deep,
@@ -21,7 +22,8 @@
   volume={4},
   pages={e13},
   year={2024},
-  publisher={Cambridge University Press}
+  publisher={Cambridge University Press},
+  doi={10.1017/s2633903x24000096}
 }
 
 @article{terris2023meta,
@@ -39,7 +41,8 @@
   number={2},
   pages={1314--1346},
   year={2024},
-  publisher={SIAM}
+  publisher={SIAM},
+  doi={10.1137/23m1545628}
 }
 
 @article{lazarus2019sparkling,
@@ -61,7 +64,8 @@
   number={7},
   pages={734--740},
   year={2020},
-  publisher={Nature Publishing Group US New York}
+  publisher={Nature Publishing Group US New York},
+  doi={10.1038/s41592-020-0853-5}
 }
 
 
@@ -73,7 +77,8 @@
   number={7},
   pages={3142--3155},
   year={2017},
-  publisher={IEEE}
+  publisher={IEEE},
+  doi={10.1109/tip.2017.2662206}
 }
 
 @article{soubies2019pocket,
@@ -84,7 +89,8 @@
   number={10},
   pages={104006},
   year={2019},
-  publisher={IOP Publishing}
+  publisher={IOP Publishing},
+  doi={10.1088/1361-6420/ab2ae9}
 }
 
 @article{gazzola2019ir,
@@ -95,7 +101,8 @@
   number={3},
   pages={773--811},
   year={2019},
-  publisher={Springer}
+  publisher={Springer},
+  doi={10.1007/s11075-018-0570-7},
 }
 
 
@@ -107,7 +114,8 @@
   number={4},
   pages={1804--1844},
   year={2017},
-  publisher={SIAM}
+  publisher={SIAM},
+  doi={10.1137/16m1102884}
 }
 @inproceedings{blau2018perception,
   title={The perception-distortion tradeoff},
@@ -125,7 +133,8 @@
   number={4},
   pages={2040--2071},
   year={2023},
-  publisher={SIAM}
+  publisher={SIAM},
+  doi={10.1137/22m1506122}
 }
 
 @inproceedings{zhang2017learning,
@@ -133,7 +142,8 @@
   author={Zhang, Kai and Zuo, Wangmeng and Gu, Shuhang and Zhang, Lei},
   booktitle={Proceedings of the IEEE conference on computer vision and pattern recognition},
   pages={3929--3938},
-  year={2017}
+  year={2017},
+  doi={10.1109/cvpr.2017.300}
 }
 
 @inproceedings{zhang2018unreasonable,
@@ -141,7 +151,8 @@
   author={Zhang, Richard and Isola, Phillip and Efros, Alexei A and Shechtman, Eli and Wang, Oliver},
   booktitle={Proceedings of the IEEE conference on computer vision and pattern recognition},
   pages={586--595},
-  year={2018}
+  year={2018},
+  doi={10.1109/cvpr.2018.00068}
 }
 
 @article{yeganeh2012objective,
@@ -163,7 +174,8 @@
   number={3},
   pages={209--212},
   year={2012},
-  publisher={IEEE}
+  publisher={IEEE},
+  doi={10.1109/lsp.2012.2227726}
 }
 
 @article{wang2004image,
@@ -174,7 +186,8 @@
   number={4},
[…]
