+++ paper/paper.bib
@@ -199,6 +199,19 @@
   doi={10.1109/iccv48922.2021.00434}
 }
 
+
+@article{chen2023imaging,
+	title = {Imaging {With} {Equivariant} {Deep} {Learning}: {From} unrolled network design to fully unsupervised learning},
+	volume = {40},
+	doi = {10.1109/MSP.2022.3205430},
+	journal = {IEEE Signal Processing Magazine},
+	author = {Chen, Dongdong and Davies, Mike and Ehrhardt, Matthias J. and Schönlieb, Carola-Bibiane and Sherry, Ferdia and Tachella, Julián},
+	month = jan,
+	year = {2023},
+	pages = {134--147},
+}
+
+
 @article{adler2018learned,
   title={Learned primal-dual reconstruction},
   author={Adler, Jonas and {\"O}ktem, Ozan},
