### 2025-07-07 36e91a3
docs + tests
