+++ deepinv/models/ram.py
@@ -1,3 +1,5 @@
+from collections import OrderedDict
+
 import torch
 import torch.nn as nn
 import torch.nn.functional as F
@@ -215,8 +217,10 @@ class RAM(Reconstructor):
     def forward(self, y=None, physics=None):
         r"""
         Reconstructs a signal estimate from measurements y
+
         :param torch.Tensor y: measurements
         :param deepinv.physics.Physics physics: forward operator
+        :return: torch.Tensor: reconstructed signal estimate
         """
         if physics is None:
             physics = dinv.physics.Denoising(
@@ -245,8 +249,20 @@ class RAM(Reconstructor):
         return out
 
 
-### --------------- MODEL ---------------
 class BaseEncBlock(nn.Module):
+    r"""
+    Base encoding block for the RAM model.
+
+    This block consists of multiple convolutional residual blocks.
+
+    :param int in_channels: Number of input channels.
+    :param int out_channels: Number of output channels.
+    :param bool bias: Whether to use bias in the convolution.
+    :param int nb: Number of residual blocks in the encoding block.
+    :param int, list[int] img_channels: Number of input channels. If a list is provided, the model will have separate heads for each channel.
+    :param int decode_upscale: Upscaling factor for the decoding convolution.
+    """
+
     def __init__(
         self,
         in_channels,
@@ -288,6 +304,7 @@ def krylov_embeddings(y, p, factor, v=None, N=4, x_init=None):
     :param torch.Tensor v: Precomputed values to subtract from Krylov sequence. Defaults to None.
     :param int N: Number of Krylov iterations. Defaults to 4.
     :param torch.Tensor x_init: Initial guess. Defaults to None.
+    :return: torch.Tensor: a stacked tensor over the channel dimension containing Krylov embeddings.
     """
 
     if x_init is None:
@@ -554,6 +571,21 @@ class OutTail(torch.nn.Module):
 
 
 class Heads(torch.nn.Module):
+    r"""
+    General heads module for the RAM model.
+
+    :param list[int] in_channels_list: List of input channels for each head.
+    :param int out_channels: Number of output channels for the convolution.
+    :param int depth: Depth of the head block.
+    :param int scale: Scale factor for the downsampling or upsampling.
+    :param bool bias: Whether to use bias in the convolution.
+    :param str mode: Mode for the upsampling, e.g., "bilinear".
+    :param int c_mult: Multiplier for the number of channels.
+    :param int c_add: Additional channels to add to the input.
+    :param bool relu_in: If True, applies ReLU activation after the input convolution.
+    :param bool skip_in: If True, applies a skip connection from the input to the output.
+    """
+
     def __init__(
         self,
         in_channels_list,
@@ -617,6 +649,20 @@ class Heads(torch.nn.Module):
 
 
 class Tails(torch.nn.Module):
+    r"""
+    General tails module for the RAM model.
+
+    :param int in_channels: Number of input channels.
+    :param list[int] out_channels_list: List of output channels for each tail.
+    :param int depth: Depth of the tail block.
+    :param int scale: Scale factor for the upsampling.
+    :param bool bias: Whether to use bias in the convolution.
+    :param str mode: Mode for the upsampling, e.g., "bilinear".
+    :param int c_mult: Multiplier for the number of channels.
+    :param bool relu_in: If True, applies ReLU activation after the input convolution.
+    :param bool skip_in: If True, applies a skip connection from the input to the output.
+    """
+
     def __init__(
         self,
         in_channels,
@@ -678,6 +724,20 @@ class Tails(torch.nn.Module):
 
 
 class HeadBlock(torch.nn.Module):
+    r"""
[…]
