### 2025-07-08 4d6ae15
demo LISTA constrained
