+++ deepinv/optim/prior.py
@@ -407,15 +407,23 @@ class WaveletPrior(Prior):
         else:
             return list_norm
 
-    def prox(self, x, *args, gamma=1.0, **kwargs):
+    def prox(self, x, ths, *args, gamma=0.1, **kwargs):
         r"""Compute the proximity operator of the wavelet prior with the denoiser :class:`~deepinv.models.WaveletDenoiser`.
         Only detail coefficients are thresholded.
 
         :param torch.Tensor x: Variable :math:`x` at which the proximity operator is computed.
-        :param float gamma: stepsize of the proximity operator.
+        :param int, float, torch.Tensor ths: thresholding parameter :math:`\gamma`.
+            If `ths` is a tensor, it should be of shape
+            ``(B,)`` (same coefficent for all levels), ``(B, n_levels-1)`` (one coefficient per level),
+            or ``(B, n_levels-1, 3)`` (one coefficient per subband and per level). `B` should be the same as the batch size of the input or `1`.
+            If ``non_linearity`` equals ``"soft"`` or ``"hard"``, ``ths`` serves as a (soft or hard)
+            thresholding parameter for the wavelet coefficients. If ``non_linearity`` equals ``"topk"``,
+            ``ths`` can indicate the number of wavelet coefficients
+            that are kept (if ``int``) or the proportion of coefficients that are kept (if ``float``).
+        :param float gamma: proximal operator stepsize.
         :return: (:class:`torch.Tensor`) proximity operator at :math:`x`.
         """
-        out = self.WaveletDenoiser(x, ths=gamma)
+        out = self.WaveletDenoiser(x, ths=ths * gamma)
         if self.clamp_min is not None:
             out = torch.clamp(out, min=self.clamp_min)
         if self.clamp_max is not None:
