+++ examples/unfolded/demo_vanilla_unfolded.py
@@ -14,7 +14,7 @@ import torch
 from torch.utils.data import DataLoader
 from deepinv.optim.data_fidelity import L2
 from deepinv.optim.prior import PnP
-from deepinv.optim import DRS
+from deepinv.unfolded import unfolded_builder
 from torchvision import transforms
 from deepinv.utils.demo import load_dataset
 
@@ -114,33 +114,35 @@ data_fidelity = L2()
 
 # Set up the trainable denoising prior
 # Here the prior model is common for all iterations
-prior = PnP(denoiser=dinv.models.DnCNN(depth=7, pretrained=None).to(device))
+prior = PnP(denoiser=dinv.models.DnCNN(depth=20, pretrained="download").to(device))
 
 # The parameters are initialized with a list of length max_iter, so that a distinct parameter is trained for each iteration.
 stepsize = [1] * max_iter  # stepsize of the algorithm
 sigma_denoiser = [0.01] * max_iter  # noise level parameter of the denoiser
 beta = 1  # relaxation parameter of the Douglas-Rachford splitting
+params_algo = {  # wrap all the restoration parameters in a 'params_algo' dictionary
+    "stepsize": stepsize,
+    "g_param": sigma_denoiser,
+    "beta": beta,
+}
 trainable_params = [
     "g_param",
     "stepsize",
     "beta",
-]  # define which parameters are trainable
+]  # define which parameters from 'params_algo' are trainable
 
 # Logging parameters
 verbose = True
 wandb_vis = False  # plot curves and images in Weight&Bias
 
 # Define the unfolded trainable model.
-model = DRS(
-    prior=prior,
+model = unfolded_builder(
+    iteration="DRS",
+    params_algo=params_algo.copy(),
+    trainable_params=trainable_params,
     data_fidelity=data_fidelity,
     max_iter=max_iter,
-    verbose=verbose,
-    stepsize=stepsize,
-    g_param=sigma_denoiser,
-    beta=beta,
-    unfold=True,
-    trainable_params=trainable_params,
+    prior=prior,
 )
 
 # %%
@@ -150,7 +152,7 @@ model = DRS(
 
 
 # training parameters
-epochs = 10 if torch.cuda.is_available() else 2
+epochs = 0 if torch.cuda.is_available() else 2
 learning_rate = 5e-4
 train_batch_size = 32 if torch.cuda.is_available() else 1
 test_batch_size = 3
@@ -223,11 +225,9 @@ dinv.utils.plot(
 # ------------------------------------
 #
 # We now plot the weights of the network that were learned and check that they are different from their initialization
-# values. Note that ``g_param`` corresponds to the noise parameter :math:`\sigma` of the denoiser.
+# values. Note that ``g_param`` corresponds to :math:`\lambda` in the proximal gradient algorithm.
 #
 
 dinv.utils.plotting.plot_parameters(
-    model,
-    init_params={"g_param": sigma_denoiser, "stepsize": stepsize, "beta": beta},
-    save_dir=RESULTS_DIR / "unfolded_pgd" / operation,
+    model, init_params=params_algo, save_dir=RESULTS_DIR / "unfolded_pgd" / operation
 )

[…]
