+++ paper/paper.md
@@ -117,19 +117,19 @@ bibliography: paper.bib
 
 # Statement of Need
 
-Deep neural networks have become ubiquitous in various imaging inverse problems. Despite the ever-increasing research effort, most learning-based algorithms are built from scratch, are hard to generalize beyond their specific training setting, and the reported results are often hard to reproduce. `deepinv` overcomes these limitations by providing a modular unified framework, leveraging the popular PyTorch deep learning library [@paszke2019pytorch].
-For our audience of researchers (experts in optimization, deep learning etc.), practitioners (biologists, physicists etc.) and imaging software engineers, `deepinv` is:
+Deep neural networks have become ubiquitous in various imaging inverse problems. Despite the ever-increasing research effort, most learning-based algorithms are built from scratch, are hard to generalize beyond their specific training setting, and the reported results are often hard to reproduce. DeepInverse overcomes these limitations by providing a modular unified framework, leveraging the popular PyTorch deep learning library [@paszke2019pytorch].
+For our audience of researchers (experts in optimization, deep learning etc.), practitioners (biologists, physicists etc.) and imaging software engineers, DeepInverse is:
 
 1. **Accelerating research** by enabling efficient testing, deployment and transfer
 of new ideas across imaging domains;
 2. Enlarging the **adoption of deep learning in inverse problems** by lowering the entrance bar to new users;
 3. Enhancing **research reproducibility** via a common modular framework of problems and algorithms.
 
-To the best of our knowledge, `deepinv` is the only library with a strong focus and wide set of modern learning-based methods across domains.
+To the best of our knowledge, DeepInverse is the only library with a strong focus and wide set of modern learning-based methods across domains.
 SCICO [@balke2022scico] and Pyxu [@simeoni2022pyxu] focus on optimization-based methods.
 CUQIpy [@riis2024cuqipy] focuses on Bayesian uncertainty quantification.
-ASTRA [@van2016astra], pytomography [@polson2025pytomography], TIGRE [@biguri2025tigre], ODL [@adler2018odl] and CIL [@jorgensen2021core] focus on tomography, sigpy [@ong2019sigpy] on magnetic resonance imaging, and PyLops [@ravasi2019pylops] on certain operators. 
-MATLAB libraries ([@soubies2019pocket], IR Tools [@gazzola2019ir]) are restricted to handcrafted methods without automatic differentiation.
+ASTRA [@van2016astra], pytomography [@polson2025pytomography], TIGRE [@biguri2025tigre], ODL [@adler2018odl] and CIL [@jorgensen2021core] focus on tomography, sigpy [@ong2019sigpy] on magnetic resonance imaging, and PyLops [@ravasi2019pylops] on certain linear operators. 
+MATLAB libraries [@soubies2019pocket] [@gazzola2019ir] are restricted to handcrafted methods without automatic differentiation.
 
 ![Schematic of the modular DeepInverse framework.\label{fig:schematic}](../docs/source/figures/deepinv_schematic.png)
 
@@ -143,15 +143,15 @@ where $x\in\mathcal{X}$ is an image, $y\in\mathcal{Y}$ are the measurements, $A_
 deterministic (linear or non-linear) operator capturing the physics of the acquisition and
[…]
