### 2025-07-08 26fe4de
add automatically the ref to examples
