+++ examples/plug-and-play/demo_RED_GSPnP_SR.py
@@ -186,8 +186,7 @@ test(
     plot_convergence_metrics=plot_convergence_metrics,
     verbose=True,
 )
-
 # %%
 # :References:
 #
-#    .. footbibliography::
+# .. footbibliography::
