+++ examples/plug-and-play/demo_PnP_DPIR_deblur.py
@@ -2,10 +2,7 @@
 DPIR method for PnP image deblurring.
 ====================================================================================================
 
-This example shows how to use the DPIR method to solve a PnP image deblurring problem. The DPIR method is described in
-the following paper:
-Zhang, K., Zuo, W., Gu, S., & Zhang, L. (2017).
-Learning deep CNN denoiser prior for image restoration.
+This example shows how to use the DPIR method to solve a PnP image deblurring problem. The DPIR method is described in :footcite:t:`zhang2021plug`.
 In Proceedings of the IEEE conference on computer vision and pattern recognition (pp. 3929-3938).
 """
 
