+++ docs/source/conf.py
@@ -182,6 +182,36 @@ cuda_available = torch.cuda.is_available()
 
 #############################
 
+def add_references_block_to_examples():
+    print("🔧 add_references_block_to_examples() called")
+    for root, _, files in os.walk('../../examples'):
+        for fname in files:
+            if not fname.endswith('.py'):
+                continue
+            full_path = os.path.join(root, fname)
+
+            with open(full_path, 'r', encoding='utf-8') as f:
+                content = f.read()
+
+            # Skip if already has a bibliography block
+            if 'footbibliography' in content:
+                continue
+
+            # Add References block if footcite appears
+            if ':footcite:' in content:
+                references_block = (
+                    "\n# %%\n"
+                    "# :References:\n"
+                    "#\n"
+                    "# .. footbibliography::\n"
+                )
+                content += references_block
+
+                with open(full_path, 'w', encoding='utf-8') as f:
+                    f.write(content)
+
+add_references_block_to_examples()
+
 templates_path = ["_templates"]
 exclude_patterns = ["_build", "Thumbs.db", ".DS_Store"]
 add_module_names = True  # include the module path in the function name
@@ -218,7 +248,7 @@ sphinx_gallery_conf = {
             "../../examples/adversarial-learning",
             "../../examples/external-libraries",
         ]
-    ),
+    )
 }
 
 # how to define macros: https://docs.mathjax.org/en/latest/input/tex/macros.html
@@ -298,3 +328,4 @@ napoleon_custom_sections = [
     ("Reference", "params_style"),  # Sphinx ≥ 3.5
     # ("Reference", "Parameters"),   # fallback syntax for very old Sphinx (<3.5)
 ]
+
