### 2025-07-09 fcc7413
remove REF in examples
