+++ examples/optimization/demo_3D_wavelets.py
@@ -251,7 +251,4 @@ dinv.utils.plot(
 #
 
 
-# %%
-# :References:
-#
-#    .. footbibliography::
+ 
