+++ examples/sampling/demo_sampling.py
@@ -146,7 +146,4 @@ plot(
     titles=["measurement", "ground truth", "post. mean", "post. std", "abs. error"],
 )
 
-# %%
-# :References:
-#
-#    .. footbibliography::
+ 
