+++ examples/plug-and-play/demo_RED_GSPnP_SR.py
@@ -186,7 +186,3 @@ test(
     plot_convergence_metrics=plot_convergence_metrics,
     verbose=True,
 )
-# %%
-# :References:
-#
-# .. footbibliography::
