+++ examples/sampling/demo_diffpir.py
@@ -409,7 +409,4 @@ plot(
 #       xhat = model(y, physics)
 #
 
-# %%
-# :References:
-#
-#    .. footbibliography::
+ 
