+++ examples/self-supervised-learning/demo_n2n_denoising.py
@@ -188,7 +188,4 @@ model = trainer.train()
 
 trainer.test(test_dataloader)
 
-# %%
-# :References:
-#
-#    .. footbibliography::
+ 
