+++ examples/self-supervised-learning/demo_r2r_denoising.py
@@ -191,7 +191,4 @@ model = trainer.train()
 
 trainer.test(test_dataloader)
 
-# %%
-# :References:
-#
-#    .. footbibliography::
+ 
