+++ examples/unfolded/demo_learned_primal_dual.py
@@ -286,7 +286,4 @@ model = trainer.train()
 
 trainer.test(test_dataloader)
 
-# %%
-# :References:
-#
-#    .. footbibliography::
+ 
