+++ examples/self-supervised-learning/demo_equivariant_imaging.py
@@ -197,5 +197,3 @@ model = trainer.train()
 #
 
 trainer.test(test_dataloader)
-
- 
