+++ examples/sampling/demo_ddrm.py
@@ -144,5 +144,3 @@ plot(
         "abs. error",
     ],
 )
-
- 
