+++ examples/patch-priors/demo_patch_priors_CT.py
@@ -265,5 +265,3 @@ plot(
     ],
     ["Ground truth", "Filtered Backprojection", "EPLL", "PatchNR"],
 )
-
- 
