+++ examples/self-supervised-learning/demo_multioperator_imaging.py
@@ -203,5 +203,3 @@ model = trainer.train()
 #
 
 trainer.test(test_dataloader)
-
- 
