+++ examples/unfolded/demo_LISTA.py
@@ -271,5 +271,3 @@ dinv.utils.plot(
 dinv.utils.plotting.plot_parameters(
     model, init_params=params_algo, save_dir=RESULTS_DIR / "unfolded_pgd" / operation
 )
-
- 
