+++ examples/basics/demo_lidar.py
@@ -98,6 +98,3 @@ for i in range(2):
 
 plt.tight_layout()
 plt.show()
-
-
- 
