+++ examples/basics/demo_custom_prior.py
@@ -187,6 +187,3 @@ test(
     plot_convergence_metrics=plot_convergence_metrics,
     verbose=verbose,
 )
-
-
- 
