+++ examples/sampling/demo_sampling.py
@@ -145,5 +145,3 @@ plot(
     imgs,
     titles=["measurement", "ground truth", "post. mean", "post. std", "abs. error"],
 )
-
- 
