+++ examples/unfolded/demo_learned_primal_dual.py
@@ -285,5 +285,3 @@ model = trainer.train()
 #
 
 trainer.test(test_dataloader)
-
- 
