+++ examples/self-supervised-learning/demo_r2r_denoising.py
@@ -190,5 +190,3 @@ model = trainer.train()
 #
 
 trainer.test(test_dataloader)
-
- 
