+++ examples/adversarial-learning/demo_gan_imaging.py
@@ -368,6 +368,3 @@ G = trainer.train()
 #
 
 trainer.test(test_dataloader)
-
-
- 
