+++ paper/paper.md
@@ -113,7 +113,7 @@ bibliography: paper.bib
 
 # Summary
 
-[DeepInverse](https://deepinv.github.io/) is an open-source PyTorch-based library for imaging inverse problems. DeepInverse implements all the steps to simulate acquisition and perform image reconstruction using deep learning for a wide set of domains (medical imaging, astronomical imaging, remote sensing, computational photography, compressed sensing and more).
+[DeepInverse](https://deepinv.github.io/) is an open-source PyTorch-based library for imaging inverse problems. DeepInverse implements all image reconstruction steps, including the efficient implementation of forward operators, the definition and resolution of variational problems and the design, training of advanced neural network architectures, for a wide set of domains (medical imaging, astronomical imaging, remote sensing, computational photography, compressed sensing and more).
 
 # Statement of Need
 
@@ -143,7 +143,7 @@ where $x\in\mathcal{X}$ is an image, $y\in\mathcal{Y}$ are the measurements, $A_
 deterministic (linear or non-linear) operator capturing the physics of the acquisition and
 $N_{\sigma}\colon\mathcal{Y}\mapsto \mathcal{Y}$ is a noise model parameterized by $\sigma$. The [`physics` module](https://deepinv.github.io/deepinv/user_guide/physics/intro.html) provides a scalable and modular framework, writing the forward operation as `y = physics(x, **params)`, unifying the wide variety of forward operators across various domains. 
 
-The library crucially introduces optional physics `params` $\xi$, allowing for advanced problems, including calibration,
+The library crucially introduces optional physics `params` $(\xi,\sigma)$, allowing for advanced problems, including calibration,
 blind inverse problems [@debarnot2024deep] [@chung2023parallel], co-design [@lazarus2019sparkling] [@nehme2020deepstorm3d], and robust training [@gossard2024training] [@terris2023meta].
 
 The current implemented physics, noise models, parameters $\xi$ and tools for manipulating them are enumerated in the [documentation](https://deepinv.github.io/deepinv/user_guide/physics/physics.html).
@@ -159,7 +159,7 @@ where $\operatorname{R}_{\theta}$ is a reconstruction algorithm with optional tr
 
 - **Optimization-based** methods [@chambolle2016introduction] solve
 \begin{equation} \label{eq:var}
-\operatorname{R}_{\theta}(y, A_{\xi}, \sigma) \in \operatorname{argmin}_{x} f_{\sigma}(y,A_{\xi}(x)) + g(x).
+\operatorname{R}_{\theta}(y, A_{\xi}, \sigma) \in \underset{x}{\operatorname{argmin}} f_{\sigma}(y,A_{\xi}(x)) + g(x).
 \end{equation}
 
   The [`optim` module](https://deepinv.github.io/deepinv/user_guide/reconstruction/optimization.html) implements classical data fidelity terms $f_{\sigma}\colon\mathcal{Y} \times \mathcal{Y} \mapsto \mathbb{R}$ and a variety of regularization priors $g\colon\mathcal{X}\mapsto\mathbb{R}$, including:
@@ -174,22 +174,22 @@ where $\operatorname{R}_{\theta}$ is a reconstruction algorithm with optional tr
 
 - **Sampling-based** methods defined by differential equations:
 \begin{equation}
-x_{t+1} \sim p(x_{t+1}|x_t, y, \operatorname{R}_{\theta}, A_{\xi}, \sigma) \text{ for } t=0,\dots,T-1,
[…]
