### 2025-07-09 542f730
Remove the size parameter in calls to plt.suptitle and set_title
