### 2025-07-09 670478c
change citation style GAN example
