### 2025-07-09 daef81e
pb doc optim_builder
