### 2025-07-09 5d8912f
Other errors in the examples, related to the device when a GPU is available, mix up between cpu and cuda.
