+++ examples/basics/demo_tour_mri.py
@@ -96,7 +96,7 @@ physics = dinv.physics.MRI(mask=mask, img_size=img_size, device=device)
 
 dinv.utils.plot(
     {
-        "x": (x := next(iter(DataLoader(knee_dataset)))),
+        "x": (x := next(iter(DataLoader(knee_dataset))).to(device)),
         "mask": mask,
         "y": physics(x).clamp(-1, 1),
     }
