### 2025-07-09 d8228aa
Add the ansmath package to the tex environment of matlplotlib so that lines containing \text{} compile correctly
