### 2025-07-09 c93736f
Further device errors in examples when building with a GPU available
