+++ examples/sampling/demo_diffpir.py
@@ -289,8 +289,8 @@ plt.show()
 # In the algorithm, we will only use sub-sampled versions of the noise and regularization schedules. Let's visualize
 # those.
 
-list_sigmas_algo = [sigmas[seq[i]] for i in range(max_iter)]
-list_rhos_algo = [rhos[seq[i]] for i in range(max_iter)]
+list_sigmas_algo = [sigmas[seq[i]].cpu().item() for i in range(max_iter)]
+list_rhos_algo = [rhos[seq[i]].cpu().item() for i in range(max_iter)]
 
 plt.figure(figsize=(6, 3))
 plt.rcParams.update({"font.size": 9})
@@ -350,7 +350,7 @@ with torch.no_grad():
             # 3. Sampling step
             next_sigma = sigmas[T - 1 - seq[i + 1]].cpu().numpy()
             t_im1 = find_nearest(
-                sigmas, next_sigma
+                sigmas.cpu().numpy(), next_sigma
             )  # time step associated with the next noise level
 
             eps = (x - alphas_cumprod[t_i].sqrt() * x0) / torch.sqrt(
