### 2025-07-09 b46c932
Call to config_matplotlib in conf.py + remove default fontsize value in plot parameters
