+++ examples/basics/demo_lidar.py
@@ -10,11 +10,6 @@ import deepinv as dinv
 import torch
 import matplotlib.pyplot as plt
 
-# Use matplotlib config from deepinv to get nice plots
-from deepinv.utils.plotting import config_matplotlib
-
-config_matplotlib()
-
 
 # %%
 # Create forward model
