### 2025-07-10 99402b1
Set figure title fontsize and axes title fontsize explicitely in ``config_matplotlib``
When using seaborn, fontsize had again no effect on the fontsize of the titles, but only on suptitles.
Now one can use seaborn and still control the fontsize of the titles.
