### 2025-07-11 5367443
docstring return denoise_diffusion (#600)
