### 2025-07-18 #8292 @tachella
@crvernon thanks for checking-in, we have been working in reducing the word count, we should be updating the paper this evening (I will send a reminder once its done)
