### 2025-07-21 #8292 @tachella
@crvernon We have updated the draft to have less than 1500 words, and also updated all the DOIs (except for two papers which didn't have DOIs unfortunately). The updated paper is in the [joss branch](https://github.com/deepinv/deepinv/tree/joss).
