### 2025-09-20 #8923 @HectorMozo3110
### Review status
----
I have now completed the checklist, and overall the software looks in very good shape and consistent with the JOSS guidelines. The following suggestions remain pending:

---

- [ ] Unify or reduce the “Coding Practices”, “Documentation”, and “Perspectives” sections (manuscript).
- [ ] [PR #784](https://github.com/deepinv/deepinv/pull/784) adjusting the paper metadata.
---
After these are addressed, I’ll be ready to close my review.
