### 2025-10-06 #8923 @ScepticalRabbit
I have just completed my last few checks by running the tests and a bunch of the examples in the docs. I have also read the latest version of the manuscript and it looks great to me. The links in the manuscript out to the docs are particular helpful. Looking forward to using this package some more myself in the future, thanks @tachella!

This completes my review and I would recommend this paper is accepted. @vissarion my review is now complete as well.
